% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit_it.R
\name{at_risk}
\alias{at_risk}
\title{at risk counter}
\usage{
at_risk(x, tp = NULL, tidy = FALSE)
}
\arguments{
\item{x}{an object created by \code{survfit_it} or equivalent}

\item{tp}{timepoints wanted}

\item{tidy}{output in tidy format (default \code{FALSE})}
}
\value{
a data frame
}
\description{
count individuals at risk and (cumulative) events at given time points
}
