% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/between.R
\name{between}
\alias{between}
\title{Condition on a vector similar to SQL's between clause}
\usage{
between(x, low, high, inclusive = TRUE)
}
\arguments{
\item{x}{A vector.}

\item{low}{Lower limit}

\item{high}{Upper Limit}

\item{inclusive}{If set to \code{TRUE}, then the condition is inclusive of what are set for \code{low} and \code{high}. Other, the "between" is literal.}
}
\value{
A vector.
}
\description{
Condition on a vector similar to SQL's between clause
}
\examples{
k <- 1:100
between(k, 5, 24)

subset(mtcars, between(mpg, 18, 21, inclusive = FALSE))
}
\seealso{
\url{https://github.com/robertschnitman/dm}
}
