% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.R
\name{browse_docs}
\alias{browse_docs}
\title{Browse package documentation}
\usage{
browse_docs()
}
\value{
Opens the {dm}-website as a side-effect
}
\description{
Opens the documentation website in a browser.
}
\examples{
browse_docs()

}
\keyword{internal}
