% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHarvest.R
\name{calcHarvest}
\alias{calcHarvest}
\title{Calculate harvest rate}
\usage{
calcHarvest(bdat)
}
\arguments{
\item{bdat}{data for the bayesian specification of a DM run}
}
\value{
Harvest rate (scalar).
}
\description{
Calculate the harvest rate by year and age of return
}
