% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcInitVals.R
\name{calcInitVals}
\alias{calcInitVals}
\title{Calculate initial values for MCMC.}
\usage{
calcInitVals(mlEst, dat, input, bdat, sims = NULL)
}
\arguments{
\item{mlEst}{The maximum likelihood estimate for the four SR parameters: log(prod), log(capacity), log(msCoef), log(flowCoef).  Output from findOptimum().}

\item{dat}{data from the A & P file}

\item{input}{a list with the other values needed for a DM run. 
The following are examples naturalMort, analysisType = "DM", SRfunction, includeMarineSurvival, includeFlow}

\item{bdat}{The in a format appropriate for JAGS implementation of the model.}

\item{sims}{MCMC simulations from a previous fit. If sims=NULL all initial values are based on estimates derived from the data.}
}
\value{
A list with initial values for prod (production), logCap (log capacity), 
msCoef (the coefficient for marine survival), 
flowCoef (the coeficient for stream survival), rMu and maturationRate_logit.
}
\description{
\code{calcInitVals} computes initial values for MCMC algorithm used to compute the posterior distribution for the model parameters.
}
