% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateHarvested.R
\name{calculateHarvested}
\alias{calculateHarvested}
\title{Calculate harvested recruits and escapement}
\usage{
calculateHarvested(r1)
}
\arguments{
\item{r1}{a single run list output by the runModel function}
}
\value{
A list with harvested recruits and escapement numbers.
}
\description{
Calculate harvested recruits and escapement
}
