% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePEandSmsy.R
\name{calculatePEandSmsy}
\alias{calculatePEandSmsy}
\title{Calculate process error and SMSY.}
\usage{
calculatePEandSmsy(input, dat, postdraws)
}
\arguments{
\item{input}{a list with the other values needed for a DM run. 
The following are examples naturalMort, analysisType = "DM", SRfunction, includeMarineSurvival, includeFlow}

\item{dat}{data from the A & P file}

\item{postdraws}{posterior list from getPostDraws() call}
}
\value{
A dataframe tDat with a column for posterior draws from each parameter 
and the smsy for each draw.
}
\description{
This function takes input and dat produced from the A & P files
along with x (the posteriors) from getPosteriors().  It computes the process error 
estimates (MSE and autoCorr or mean and sterr) along with the SMSY for each posterior 
draw.  It then assembles all these into a dataframe tDat and returns that.
}
