% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdates.R
\name{cdate}
\alias{cdate}
\title{fix censored dates}
\usage{
cdate(x, sep = NULL, low.bound = NULL, ok.years = NULL)
}
\arguments{
\item{x}{dates}

\item{sep}{seprator in x, if any}

\item{low.bound}{the lower bound}

\item{ok.years}{span of years that are ok, values outside will generate an
error, \code{NULL} (default) all years are accepted}
}
\description{
fix censored dates of the type '20110000' or '20110200' by selecting the
    midpoint of the censored time interval. If a lower bound is given, the
    midpoint between this point and the end of the censored interval is
    chosen.
}
