% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{cdm_add_fk}
\alias{cdm_add_fk}
\title{Add a reference from one table of a \code{dm} to another}
\usage{
cdm_add_fk(dm, table, column, ref_table, check = TRUE)
}
\description{
Add a reference from one table of a \code{dm} to another
}
