% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-keys.R
\name{cdm_add_pk}
\alias{cdm_add_pk}
\title{Mark a column of a table in a \code{dm}-object as its primary key}
\usage{
cdm_add_pk(dm, table, column, check = TRUE, force = FALSE)
}
\description{
\code{cdm_add_pk()} marks the given column as the given table's primary key
in the \code{data_model}-part of the \code{dm}-object. If \code{check == TRUE}, it also first checks if
the given column is a unique key of the table. If \code{force == TRUE}, it replaces an already
set key.
}
\examples{
\dontrun{
library(nycflights13)
library(dplyr)

nycflights_dm <- dm(src_df(pkg = "nycflights13"))

# the following works
cdm_add_pk(nycflights_dm, planes, tailnum)
cdm_add_pk(nycflights_dm, airports, faa)
cdm_add_pk(nycflights_dm, planes, manufacturer, check = FALSE)

# the following does not work
cdm_add_pk(nycflights_dm, planes, manufacturer)
}
}
