% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_management.R
\name{cdm_add_tbls}
\alias{cdm_add_tbls}
\title{Add one or more tables to a \code{\link{dm}}}
\usage{
cdm_add_tbls(dm, ...)
}
\arguments{
\item{dm}{A \code{\link{dm}} object}

\item{...}{One or more tibbles to add to the \code{dm}. If not named, the tables will retain their name within the \code{dm}
(does not work in a pipe or in \code{map()}-style functions). Explicit naming is supported with the syntax \code{new_table_name = tbl}.}
}
\value{
The inital \code{dm} with the additional table(s).
}
\description{
Adds one or more tibbles to a \code{\link{dm}}, using tidyverse syntax.
}
