% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{cdm_apply_filters}
\alias{cdm_apply_filters}
\title{Apply all filters}
\usage{
cdm_apply_filters(dm)
}
\arguments{
\item{dm}{A \code{dm} object.}
}
\description{
All set filters are applied and their combined cascading effect on each table of the \code{\link{dm}} is taken into account.
An alias for this function is the method for \code{dm} class objects for \code{compute()}.
}
\examples{
cdm_nycflights13() \%>\%
  cdm_filter(flights, month == 3) \%>\%
  cdm_apply_filters()

library(dplyr)
cdm_nycflights13() \%>\%
  cdm_filter(planes, engine \%in\% c("Reciprocating", "4 Cycle")) \%>\%
  compute()

}
