% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key-helpers.R
\name{cdm_check_constraints}
\alias{cdm_check_constraints}
\title{Validate your \code{\link{dm}}: are all key constraints met?}
\usage{
cdm_check_constraints(dm)
}
\arguments{
\item{dm}{A \code{dm} object.}
}
\value{
A list of 2:
\enumerate{
\item \code{pk}: a named list of logical values. The names are a combination of the table name, a dollar sign
and the primary key column name.
\item \code{fk}: a named list of logical values. The names are a combination of the table name, a dollar sign
and the foreign key column name.
}
}
\description{
This function returns a concise list, reporting for each key constraint if it is met (\code{TRUE}) or violated (\code{FALSE}).
}
\details{
For the primary key constraints it is tested, if the values in the respective columns are all unique.
For the foreign key constraints the tests check, if for each foreign key constraint, the values of the foreign key column
are a subset of those of the referenced column.
}
\examples{
cdm_check_constraints(cdm_nycflights13())
}
