% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-keys.R
\name{cdm_check_for_pk_candidates}
\alias{cdm_check_for_pk_candidates}
\title{Which columns are candidates for a primary key column of a \code{dm}-object's table?}
\usage{
cdm_check_for_pk_candidates(dm, table)
}
\description{
\code{cdm_check_for_pk_candidates()} checks for each column of a
table of a \code{dm}-object if this column contains only unique values and is therefore
a unique key of this table.
}
\examples{
\dontrun{
library(nycflights13)
library(dplyr)

nycflights_dm <- dm(src_df(pkg = "nycflights13"))

nycflights_dm \%>\% cdm_check_for_pk_candidates("flights")
nycflights_dm \%>\% cdm_check_for_pk_candidates("airports")
}

}
