% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-dm.R
\name{cdm_draw}
\alias{cdm_draw}
\alias{cdm_set_colors}
\alias{cdm_get_colors}
\alias{cdm_get_available_colors}
\title{Draw schema of a 'dm'-object's data model}
\usage{
cdm_draw(dm, table_names = NULL, rankdir = "LR", col_attr = "column",
  view_type = "keys_only", columnArrows = TRUE, graph_attrs = "",
  node_attrs = "", edge_attrs = "", focus = NULL,
  graph_name = "Data Model")

cdm_set_colors(dm, ...)

cdm_get_colors(dm)

cdm_get_available_colors()
}
\arguments{
\item{...}{Colors to set in the form \code{table = "<color>"} . Fall-through syntax similarly to
\code{\link[=switch]{switch()}} is supported: \code{table1 = , table2 = "<color>"} sets the color for both \code{table1}
and \code{table2} .}
}
\value{
For \code{cdm_set_colors()}: the updated data model.

For \code{cdm_get_colors()}, a two-column tibble with one row per table.

For \code{cdm_get_available_colors()}, a tibble with the color in the first
column and auxiliary information in other columns.
}
\description{
\code{cdm_draw()} draws a schema of the data model using \code{datamodelr} (which in turn uses \code{DiagrammeR})

\code{cdm_set_colors()} allows to define the colors in which to display the tables of the data model.

\code{cdm_get_colors()} returns the colors defined for a data model.

\code{cdm_get_available_colors()} returns an overview of the available colors and their names
as a tibble.
}
