% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-dm.R
\name{cdm_draw_data_model}
\alias{cdm_draw_data_model}
\alias{cdm_draw_data_model_set_colors}
\alias{cdm_draw_data_model_print_colors}
\title{Draw schema of a 'dm'-object's data model}
\usage{
cdm_draw_data_model(dm, table_names = NULL, rankdir = "LR",
  col_attr = "column", view_type = "keys_only", columnArrows = TRUE,
  graph_attrs = "", node_attrs = "", edge_attrs = "", focus = NULL,
  graph_name = "Data Model")

cdm_draw_data_model_set_colors(dm, list_of_table_colors)

cdm_draw_data_model_print_colors()
}
\description{
\code{cdm_draw_data_model()} draws a schema of the data model using \code{datamodelr} (which in turn uses \code{DiagrammeR})

\code{cdm_draw_data_model_set_colors()} allows to define the colors in which to display the tables of the data model

\code{cdm_draw_data_model_print_colors()} prints an overview of the available colors and their names
}
