% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{cdm_enum_fk_candidates}
\alias{cdm_enum_fk_candidates}
\alias{enum_fk_candidates}
\title{Find foreign key candidates in a table}
\usage{
cdm_enum_fk_candidates(dm, table, ref_table)

enum_fk_candidates(zoomed_dm, ref_table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{The table whose columns should be tested for suitability of use as foreign keys.}

\item{ref_table}{A table with a primary key.}

\item{zoomed_dm}{A \code{dm} with a zoomed table.}
}
\value{
A table that lists which columns of \code{table} would be suitable candidates for
foreign key columns to reference \code{ref_table}, and which columns would not be suitable.
}
\description{
Determine which columns would be good candidates to be used as foreign keys of a table,
and reference the primary key column of another \code{\link{dm}} object's table.
}
\details{
\code{cdm_enum_fk_candidates()} first checks if \code{ref_table} has a primary key set.
Then, for each column of \code{table}, a join operation will be tried, with argument \code{by} matching
the respective column with the primary key of \code{ref_table}.
This tests implicitly for
type compatibility (on most sources).
Based on the result of the join, the
entry in the result column \code{why} is:
\itemize{
\item an empty entry, if the column is a candidate
\item the total percentage and individual numbers of missing matches between the entries of the
respective column in table \code{table} and the primary column entries in table \code{ref_table}.
\item the error message triggered by the error (often stating the mismatched column types)
}

\code{enum_fk_candidates()} works like \code{cdm_enum_fk_candidates()} with the zoomed table as \code{table}.
}
\examples{
cdm_enum_fk_candidates(cdm_nycflights13(), flights, airports)

}
\seealso{
Other foreign key functions: 
\code{\link{cdm_add_fk}()},
\code{\link{cdm_get_all_fks}()},
\code{\link{cdm_get_fk}()},
\code{\link{cdm_has_fk}()},
\code{\link{cdm_rm_fk}()}
}
\concept{foreign key functions}
