% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-dm.R
\name{cdm_filter}
\alias{cdm_filter}
\title{Cascading filter of a \code{dm}-object}
\usage{
cdm_filter(dm, table, ...)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}

\item{...}{Logical predicates defined in terms of the variables in .data.
Multiple conditions are combined with &. Only rows where the condition evaluates
to TRUE are kept.

The arguments in ... are automatically quoted and evaluated in the context of
the data frame. They support unquoting and splicing. See vignette("programming")
for an introduction to these concepts.}
}
\description{
'cdm_filter()' allows you to set a filter condition for one table
of a \code{dm}-objhect. If the \code{dm}'s tables are connected via key constrains, the
filtering will affect all other connected tables, leaving only the rows with
the corresponding key values.
}
