% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_tables_from_dm.R
\name{cdm_find_conn_tbls}
\alias{cdm_find_conn_tbls}
\title{Find the tables connecting two or more tables in a \code{dm}}
\usage{
cdm_find_conn_tbls(dm, ...)
}
\arguments{
\item{dm}{A \code{dm}-object}

\item{...}{Two or more table names of the \code{dm}-object's tables.}
}
\value{
Character vector with the names of the connecting tables.
}
\description{
Find all tables that need to be passed when traversing the \code{dm}-object
between the indicated tables along the foreign
key relations. Result includes the given tables.
}
