% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{cdm_flatten}
\alias{cdm_flatten}
\title{Flatten \code{dm} into a wide table}
\usage{
cdm_flatten(dm, join = left_join)
}
\arguments{
\item{dm}{A \code{\link{dm}} object}

\item{join}{The type of join to be performed, see \code{\link[dplyr]{join}}}
}
\value{
A wide table resulting of consecutively joining all tables together.
}
\description{
This function joins all the tables of your \code{dm} object together and returns one
table with unique columns. Use
\code{\link[=cdm_select_tbl]{cdm_select_tbl()}} if necessary to reduce the number of tables before calling this function.
}
\examples{
cdm_nycflights13() \%>\%
cdm_select_tbl(-weather) \%>\%
cdm_flatten()

}
\seealso{
Other Flattening functions: \code{\link{cdm_join_tbl}}
}
\concept{Flattening functions}
