% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{cdm_flatten_to_tbl}
\alias{cdm_flatten_to_tbl}
\title{Flatten part of a \code{dm} into a wide table}
\usage{
cdm_flatten_to_tbl(dm, start, ..., join = left_join)
}
\arguments{
\item{dm}{A \code{\link{dm}} object}

\item{start}{Table to start from. From this table all outgoing foreign key relations are
considered to establish a processing order for the joins. An interesting choice could be
for example a fact table in a star schema.}

\item{...}{Unquoted table names to include in addition to \code{start}. If empty, all tables that can
be reached are included.}

\item{join}{The type of join to be performed, see \code{\link[dplyr:join]{dplyr::join()}}}
}
\value{
A wide table resulting of consecutively joining all tables together.
}
\description{
This function joins all the tables of your \code{\link{dm}} object together, that can be reached
from table \code{start} in the direction that the foreign keys are pointing, using the
foreign key relations to determine the parameter \code{by} for the necessary joins.
It returns one table with unique columns. Use the \code{...} if necessary to
reduce the number of tables that should be included in the result.
}
\details{
Applied the necessary filters (in any case the cascading effect of the filters on \code{start}
is taken into account) and \code{\link[=cdm_disambiguate_cols]{cdm_disambiguate_cols()}} first, to get a "clean" \code{\link{dm}}.
Then the order of the joins is determined and the joins are performed.

Mind, that for \code{join = right_join} the result two or more consecutive joins depends on the order
of the joins and eventually on the order of the table in the \code{dm}.
}
\examples{
cdm_nycflights13() \%>\%
  cdm_select_tbl(-weather) \%>\%
  cdm_flatten_to_tbl(flights)
}
\seealso{
Other flattening functions: \code{\link{cdm_join_to_tbl}}
}
\concept{flattening functions}
