% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{cdm_get_all_fks}
\alias{cdm_get_all_fks}
\title{Retrieve all foreign key constraints in a \code{dm}}
\usage{
cdm_get_all_fks(dm)
}
\arguments{
\item{dm}{A \code{dm} object.}
}
\value{
A tibble with columns:

"child_table": child table,
"child_fk_col": foreign key column in child table,
"col_class": class of this column,
"parent_table": parent table,
}
\description{
Get an overview about all foreign key relations in a \code{dm}
}
