% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-keys.R
\name{cdm_get_all_pks}
\alias{cdm_get_all_pks}
\title{Get all primary keys of a \code{\link{dm}} object}
\usage{
cdm_get_all_pks(dm)
}
\arguments{
\item{dm}{A \code{dm} object.}
}
\description{
\code{cdm_get_all_pks()} checks the \code{dm} object for set primary keys and
returns the tables, the respective primary key columns and their classes.
}
\seealso{
Other Primary key functions: \code{\link{cdm_add_pk}},
  \code{\link{cdm_get_pk}}, \code{\link{cdm_has_pk}},
  \code{\link{cdm_rm_pk}}, \code{\link{enum_pk_candidates}}
}
\concept{Primary key functions}
