% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{cdm_get_fk}
\alias{cdm_get_fk}
\title{Get foreign key columns pointing from one table of a \code{dm} to another}
\usage{
cdm_get_fk(dm, table, ref_table)
}
\description{
Get foreign key columns pointing from one table of a \code{dm} to another
}
