% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{cdm_has_fk}
\alias{cdm_has_fk}
\title{Does a reference from one table of a \code{dm} to another exist?}
\usage{
cdm_has_fk(dm, table, ref_table)
}
\description{
Does a reference from one table of a \code{dm} to another exist?
}
