% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-keys.R
\name{cdm_has_pk}
\alias{cdm_has_pk}
\title{Does a table of a \code{dm} object have a column set as primary key?}
\usage{
cdm_has_pk(dm, table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}
}
\description{
\code{cdm_has_pk()} checks in the \code{data_model} part
of the \code{dm} object if a given table has a column marked as primary key.
}
\examples{
library(dplyr)
nycflights_dm <- cdm_nycflights13()

nycflights_dm \%>\%
  cdm_has_pk(planes)
}
