% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{cdm_learn_from_db}
\alias{cdm_learn_from_db}
\title{Learn a \code{\link{dm}} object from a database (DB)}
\usage{
cdm_learn_from_db(dest)
}
\arguments{
\item{dest}{A \code{src}-object on a DB or a connection to a DB.}
}
\value{
A \code{\link{dm}} object with the tables from the DB and the respective key relations.
}
\description{
If there are any permament tables on a DB, a new \code{\link{dm}} object can be created that contains those tables,
along with their primary and foreign key constraints.

Currently this only works with MSSQL and Postgres databases.

The default database schema will be used; it is currently not possible to parametrize the funcion with a specific database schema.
}
\seealso{
Other DB interaction functions: 
\code{\link{cdm_copy_to}()}
}
\concept{DB interaction functions}
