% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{cdm_remove_fk}
\alias{cdm_remove_fk}
\title{Remove reference(s) from one table of a \code{dm} to another}
\usage{
cdm_remove_fk(dm, table, column, ref_table)
}
\description{
Remove reference(s) from one table of a \code{dm} to another
}
\details{
Can either remove one reference between the two tables or all at once if parameter \code{column = NULL}.
All parameters can be provided unquoted or quoted.'
}
