% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{cdm_rename_table}
\alias{cdm_rename_table}
\alias{cdm_rename_tables}
\title{Rename tables of a \code{dm}}
\usage{
cdm_rename_table(dm, old_name, new_name)

cdm_rename_tables(dm, old_table_names, new_table_names)
}
\arguments{
\item{dm}{A \code{dm} object}

\item{old_name}{The original name of the table}

\item{new_name}{The new name of the table}

\item{old_table_names}{Character vector or list of the original names of the tables which are to change}

\item{new_table_names}{Character vector or list of the new names of the tables}
}
\description{
\code{cdm_rename_table()} changes the name of one of a \code{dm}'s tables.

\code{cdm_rename_tables()} changes the names one or more tables of a \code{dm}.
}
