% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{cdm_rename_tbl}
\alias{cdm_rename_tbl}
\title{Change names of tables in a \code{dm}}
\usage{
cdm_rename_tbl(dm, ...)
}
\arguments{
\item{dm}{A \code{dm} object}

\item{...}{Named character vector (new_name = old_name)}
}
\description{
\code{cdm_rename_tbl()} changes the names of one or more tables of a \code{dm}.
}
