% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-keys.R
\name{cdm_rm_pk}
\alias{cdm_rm_pk}
\title{Remove primary key from a table in a \code{dm}-object}
\usage{
cdm_rm_pk(dm, table)
}
\description{
\code{cdm_rm_pk()} removes a potentially set primary key from a table in the
underlying \code{data_model}-object and otherwise leaves the \code{dm}-object untouched.
}
\examples{
\dontrun{
library(nycflights13)
library(dplyr)

nycflights_dm <- dm(src_df(pkg = "nycflights13"))

# the following works
cdm_obj_with_keys <- cdm_add_pk(nycflights_dm, "planes", "tailnum") \%>\%
  cdm_add_pk("airports", faa)

cdm_obj_with_keys \%>\%
  cdm_rm_pk("airports") \%>\%
  cdm_has_pk("planes")

cdm_obj_with_keys \%>\%
  cdm_rm_pk("planes") \%>\%
  cdm_has_pk("planes")
}

}
