% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_management.R
\name{cdm_rm_tbls}
\alias{cdm_rm_tbls}
\title{Remove one or more tables to a \code{\link{dm}}}
\usage{
cdm_rm_tbls(dm, ...)
}
\arguments{
\item{dm}{A \code{\link{dm}} object}

\item{...}{One or more unquoted tibble names to remove from the \code{dm}.}
}
\value{
The inital \code{dm} without the removed table(s).
}
\description{
Removes one or more tibbles from a \code{\link{dm}} by wrapping \code{cdm_select_tbls()}. Mind, that all foreign keys are
dropped that pointed to or from the removed table(s).
}
