% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-tbl.R, R/table_management.R
\name{cdm_select_tbl}
\alias{cdm_select_tbl}
\alias{cdm_rename_tbl}
\alias{cdm_add_tbl}
\title{Select and rename tables}
\usage{
cdm_select_tbl(dm, ...)

cdm_rename_tbl(dm, ...)

cdm_add_tbl(dm, table, table_name = NULL)
}
\arguments{
\item{dm}{A \code{\link{dm}} object}

\item{...}{One or more table names of the \code{\link{dm}} object's tables.
See \code{\link[tidyselect:vars_select]{tidyselect::vars_select()}} and \code{\link[tidyselect:vars_rename]{tidyselect::vars_rename()}}
for details on the semantics.}

\item{table}{A tibble}

\item{table_name}{The name for the new table. If left \code{NULL}, the new table will retain its original name
(does not work in a pipe or in \code{map()}-style functions)}

\item{dm}{A \code{\link{dm}} object}
}
\value{
The input \code{dm} with tables renamed or removed.

The inital \code{dm} with the additional table.
}
\description{
\code{cdm_select_tbl()} keeps the selected tables and their relationships,
optionally renaming them.

\code{cdm_rename_tbl()} renames tables.

Adds a tibble to a \code{\link{dm}}.
}
