% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-dm.R
\name{cdm_semi_join}
\alias{cdm_semi_join}
\title{Semi-join a \code{\link{dm}} object with one of its reduced tables}
\usage{
cdm_semi_join(dm, table, reduced_table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}

\item{reduced_table}{The table indicated in argument \code{table}, but in a filtered state (cf. \code{dplyr::filter()}).}
}
\description{
Similar to the subsequent calls of \code{\link[=cdm_filter]{cdm_filter()}} and \code{\link[=cdm_apply_filters]{cdm_apply_filters()}}, \code{cdm_semi_join()}
performs a cascading "row reduction" of a \code{\link{dm}} object
by an inital semi-join of one of its tables with the same but filtered table. Subsequently, the
key constraints are used to compute the remainders of the other tables of the \code{\link{dm}} object and
a new \code{\link{dm}} object is returned.
}
\references{
\code{\link[=cdm_filter]{cdm_filter()}}
}
