% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-interface.R
\name{cdm_set_key_constraints}
\alias{cdm_set_key_constraints}
\title{Set key constraints on a DB for a \code{dm}-obj with keys.}
\usage{
cdm_set_key_constraints(dm)
}
\arguments{
\item{dm}{A \code{dm} object.}
}
\description{
\code{cdm_set_key_constraints()} takes a \code{dm}-object that lives on a DB (so far
it works exclusively for MSSQL and Postgres) and mirrors the \code{dm} key constraints
on the database.
}
\examples{
src_sqlite <- dplyr::src_sqlite(":memory:", create = TRUE)
iris_dm <- cdm_copy_to(
  src_sqlite,
  as_dm(list(iris = iris)),
  set_key_constraints = FALSE
)

# there are no key constraints in `as_dm(list(iris = iris))`,
# but if there were, and if we had already implemented setting key
# constraints for SQLite, this would do something:
cdm_set_key_constraints(iris_dm)
}
