% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.R
\name{cdm_with}
\alias{cdm_with}
\title{Among others: address table by their unquoted name}
\usage{
cdm_with(dm, code)
}
\description{
FIXME: missing description
}
\examples{
cdm_with(cdm_nycflights13(), airports)
}
