% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{cdm_zoom_to_tbl}
\alias{cdm_zoom_to_tbl}
\alias{cdm_insert_zoomed_tbl}
\alias{cdm_update_zoomed_tbl}
\alias{cdm_zoom_out}
\title{Single out a table of a \code{dm}}
\usage{
cdm_zoom_to_tbl(dm, table)

cdm_insert_zoomed_tbl(dm, new_tbl_name)

cdm_update_zoomed_tbl(dm)

cdm_zoom_out(dm)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}

\item{new_tbl_name}{Name of the new table}
}
\description{
Zooming to a table of a \code{\link{dm}} allows for the use of many \code{dplyr}-verbs directly on this table, while retaining the
context of the \code{dm} object.
}
\details{
\code{cdm_zoom_to_tbl()}: zooms to the given table

\code{cdm_update_zoomed_tbl()}: overwrites the originally zoomed table with the manipulated table.
The filter conditions for the zoomed table are added to the original filter conditions.

\code{cdm_insert_zoomed_tbl()}: adds a new table to the \code{dm}.

\code{cdm_zoom_out()}: discards the zoomed table, returning the \code{dm} from before zooming

Whenever possible, the key relations of the original table are transferred to the resulting table
when using \code{cdm_insert_zoomed_tbl()} or \code{cdm_update_zoomed_tbl()}.

Functions from \code{dplyr}, that are supported for a \code{zoomed_dm}: \code{group_by()}, \code{summarise()}, \code{mutate()},
\code{transmute()}, \code{select()}, \code{rename()} and \code{ungroup()}. You can use these functions just like you would
with a normal table.

\code{filter()} is also supported, but treated in a special way: the filter expression for the zoomed table is
stored in the \code{dm} and is treated depending on which function you use to return to a normal \code{dm}:
\enumerate{
\item \code{cdm_zoom_out()}: all filter conditions for the zoomed table are discarded
\item \code{cdm_update_zoomed_tbl()}: the filter conditions of the original table and those of the zoomed table are combined
\item \code{cdm_insert_zoomed_tbl()}: the filter conditions of the original table stay there and those of the zoomed table are
transferred to the new table of the \code{dm}
}
}
