% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{check_foreign_key}
\alias{check_foreign_key}
\title{Test foreign key property for two tables and two columns (NOT column combinations) in one direction}
\usage{
check_foreign_key(t1, c1, t2, c2)
}
\arguments{
\item{t1}{First data frame whose column `c1` should be tested for foreign key properties.}

\item{c1}{Column of first data frame which should be tested for foreign key property w.r.t. the second table,
i.e. if all values of `c1` are also values of `c2`.}

\item{t2}{Second data frame.}

\item{c2}{Column of second data frame which has to contain all values of `c1` to avoid an error.}
}
\description{
`check_foreign_key()` tests, if the chosen column `c1` of data frame `t1` is a foreign key for
data frame `t2` (when checked against column `c2`).
}
\examples{
\dontrun{
data_1 <- tibble(a = c(1, 2, 1), b = c(1, 4, 1), c = c(5, 6, 7))
data_2 <- tibble(a = c(1, 2, 3), b = c(4, 5, 6), c = c(7, 8, 9))
# this is passing:
check_foreign_key(data_1, a, data_2, a)

# this is failing:
check_foreign_key(data_2, a, data_1, a)
}
}
