% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key-helpers.R
\name{check_if_subset}
\alias{check_if_subset}
\title{Test if values of one column are a subset of values of another column}
\usage{
check_if_subset(t1, c1, t2, c2, verbose = TRUE)
}
\arguments{
\item{t1}{Data frame containing the column \code{c1}.}

\item{c1}{Column of \code{t1} that should only contain values that are also in \code{c2} of data frame \code{t2}.}

\item{t2}{Data frame containing the column \code{c2}.}

\item{c2}{Column of second data frame which has to contain all values of \code{c1} to avoid an error.}

\item{verbose}{Boolean, if \code{TRUE} (default), the lines with missing column entries will be printed before an error is thrown.}
}
\description{
\code{check_if_subset()} tests, if the values of the chosen column \code{c1} of data frame \code{t1} are a subset of the values
of column \code{c2} of data frame \code{t2}.
}
\examples{
data_1 <- tibble::tibble(a = c(1, 2, 1), b = c(1, 4, 1), c = c(5, 6, 7))
data_2 <- tibble::tibble(a = c(1, 2, 3), b = c(4, 5, 6), c = c(7, 8, 9))
# this is passing:
check_if_subset(data_1, a, data_2, a)

# this is failing:
try(check_if_subset(data_2, a, data_1, a))
}
