% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{check_overlap}
\alias{check_overlap}
\title{Test foreign key properties for two tables and two columns (NOT column combinations) in both directions}
\usage{
check_overlap(t1, c1, t2, c2)
}
\arguments{
\item{t1}{First data frame whose column `c1` should be tested for key properties.}

\item{c1}{Column of first data frame which should be tested for foreign key property w.r.t. the second table,
i.e. if all values of `c1` are also values of `c2`.}

\item{t2}{Second data frame whose column should be tested for key properties.}

\item{c2}{Column of second data frame which should be tested for foreign key property w.r.t. the first table,
i.e. if all values of `c2` are also values of `c1`.}
}
\description{
`check_overlap()` is a wrapper of `check_foreign_key()`. It tests the foreign key property in both directions.
}
\examples{
\dontrun{
data_1 <- tibble(a = c(1, 2, 1), b = c(1, 4, 1), c = c(5, 6, 7))
data_2 <- tibble(a = c(1, 2, 3), b = c(4, 5, 6), c = c(7, 8, 9))
# this is failing:
check_overlap(data_1, a, data_2, a)

data_3 <- tibble(a = c(2, 1, 2), b = c(4, 5, 6), c = c(7, 8, 9))
# this is passing:
check_overlap(data_1, a, data_3, a)
}
}
