% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kin.R
\name{cluster.by}
\alias{cluster.by}
\alias{cluster.by.incl.next}
\title{clustering}
\usage{
cluster.by(x, by)

cluster.by.incl.next(incl.next)
}
\arguments{
\item{x}{variable from which to create cluster}

\item{by}{name of method}

\item{incl.next}{a logical variable that for each line include wether the
next line should belong to the same cluster as the current line}
}
\description{
create cluster variable indicated by other variable
}
\section{Functions}{
\itemize{
\item \code{cluster.by.incl.next}: cluster by 'include.next' indicator
}}

