% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncenterProdCap.R
\name{conditionalcenterProdCap}
\alias{conditionalcenterProdCap}
\title{Center productivty and log capacity parameters conditioned on values in input}
\usage{
conditionalcenterProdCap(input, dat, params)
}
\arguments{
\item{dat}{data from the A & P file}

\item{params}{a vector of parameters with names flowCoef, msCoef, prod, cap.  Assumed to be uncentered, e.g. the ML estimates from the SR model with covariates.}
}
\value{
a vector of centered parameters with names flowCoef, msCoef, prod, cap.
}
\description{
Center capacity and productivity so that parameters can be compared to estimates from an analysis where the covariates were (possibly) centered.  VRAP uses the uncentered parameters. The input$centerFlow and input$centerMS variables specify whether and which covariates were centered for the Bayesian analysis.
}
