% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_rms2df.R
\name{contrast_rms2df}
\alias{contrast_rms2df}
\alias{contrast_rms2tidy}
\title{contrast to data frame}
\usage{
contrast_rms2df(c)

contrast_rms2tidy(c)
}
\arguments{
\item{c}{a contrast rms object}
}
\value{
a data.frame
}
\description{
turn a contrast.rms object into a data frame
}
\section{Functions}{
\itemize{
\item \code{contrast_rms2tidy}: mimick \code{broom::tidy} behaviour  as applied
to a contrast.rms object
}}

