% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxreg_it.R
\name{cph2df}
\alias{cph2df}
\title{cph to data frame}
\usage{
cph2df(cph)
}
\arguments{
\item{cph}{an object created by \code{rms::cph}}
}
\value{
a data.frame
}
\description{
convert a \code{rms::cph} object to a data frame
}
