% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxreg_it.R
\name{cph_it}
\alias{cph_it}
\alias{cph_it_display_code}
\title{iterated use of \code{rms::cph}}
\usage{
cph_it(
  data,
  surv,
  fRHS,
  cont = NULL,
  w = NULL,
  prefix = c(time = "t.", event = "ev.")
)

cph_it_display_code()
}
\arguments{
\item{data}{a data frame}

\item{surv}{outcomes, as surv objects or specification of the outcome names
where the time- and event components differ only in prefix}

\item{fRHS}{right hand side of formula (do not include '~')}

\item{cont}{names of variables (included in fHRS) for which the actual
effects are wanted (typically continuous ones, the others are recorded anyway)}

\item{w}{weights}

\item{prefix}{prefix consistently used for time- and event components of
(what is to become) Surv objects (only needed when surv does not
indicated Surv variables already)}
}
\value{
a list with components
\itemize{
 \item{"model"}{list of all models created, named from surv argument}
 \item{"HR"}{data frame containing all model effects}
 \item{"resid"}{list of \code{cox.zph} objects, named from surv argument}
 \item{"effect"}{data frame of (continuous) effects}
}
}
\description{
run the same (RHS) formula for multiple outcomes and record models, effect,
    residuals and continuous effects
}
\section{Functions}{
\itemize{
\item \code{cph_it_display_code}: example  of code to loop out results  into a rnw document
(requires subdirectory 'saved-figure' but this is easily modified)
}}

