% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBUGSdata.R
\name{createBUGSdata}
\alias{createBUGSdata}
\title{Create bugs data for model with or without age data.}
\usage{
createBUGSdata(dat, input, priors = NULL, allData = FALSE)
}
\arguments{
\item{dat}{data from the A & P file}

\item{input}{a list with the other values needed for a DM run. 
The following are examples naturalMort, analysisType = "DM" or "SS", SRfunction, includeMarineSurvival, includeFlow}

\item{priors}{the parameters for the prior distributions used in the JAGS models.}

\item{allData}{if allData=TRUE the data for both model types is returned. If allData=FALSE, just the data necessary for the specified model is returned.}
}
\value{
dataframe returns a list with the data and priors formatted as input for the JAGS model.
}
\description{
This function creates bugs data (bdat).
}
