% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPriors.R
\name{createPriors}
\alias{createPriors}
\title{Create a priors list.}
\usage{
createPriors(pMode = 20, pSig = 7, pMin = 0, pMax = 40, cMu = 9,
  cSig = 50, cMin = 3, cMax = 12, msMin = 1e-04, flowMin = -1000,
  flowMax = -1e-04)
}
\arguments{
\item{pMode}{Mode of the production prior}

\item{pSig}{Sigma of the production prior}

\item{pMin}{Lowerbound of the production prior}

\item{pMax}{Upperbound of the production prior}

\item{cSig}{Sigma of the capacity prior}

\item{cMin}{Lowerbound of the capacity prior}

\item{cMax}{Upperbound of the capacity prior}

\item{msMin}{Lowerbound of the marine survival coefficient prior}

\item{flowMax}{Upperbound of the flow coefficient prior; the coefficient is always negative}

\item{cMode}{Mode of the capacity prior}
}
\value{
a list with parameters for each prior.
}
