% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRAVfile.R
\name{createRAVfile}
\alias{createRAVfile}
\title{Create rav file.}
\usage{
createRAVfile(bdat, input, tDat, dat, filename = "temp_rav.rav",
  estType = "median", sim = 1, rav.options = list())
}
\arguments{
\item{bdat}{data for the bayesian specification of a DM run}

\item{input}{a list with the other values needed for a DM run. 
The following are examples naturalMort, analysisType = "DM", SRfunction, includeMarineSurvival, includeFlow}

\item{tDat}{a data frame of the transformed posteriors and calculated process error and 
Smsy for each draw.  From calculatePEandSmsy().  This is specifically for VRAP input.  The
productivity (a) and capacity (b) posteriors are not adjusted (centered) if the covariates
do not have 0 mean.  VRAP needs the SR parameters in this raw form using non-centered (de-meaned) covariates.}

\item{dat}{data from the A & P file}

\item{filename}{name to give the rav file}

\item{estType}{("median")}

\item{sim}{indicates which MCMC sim from the posteriors to use if estType is not median}

\item{rav.options}{list of the rav file options to use that do not come from the posteriors.
\describe{
  \item{randomSeed}{(0)}
  \item{numRuns}{(1000) The number of simulations to run for computing 
  probabilities in VRAP.}
  \item{numYears}{(25) The number of years to project forward in the 
  simulations.}
  \item{minAge}{(2)}
  \item{maxAge}{(5)}
  \item{convergeCrit}{(0.001)}
  \item{centerCovFlag}{("NO") "NO"/"YES"}
  \item{marineSurvType}{("Autoc") "Autoc"/"Cycle"/"Trend"}
  \item{flowType}{("Autoc") "Autoc"/"Cycle"/"Trend"}
  \item{modelDepensation}{("NO") "NO"/"YES"}
  \item{depensation}{(300)}
  \item{QETcritical}{(63)}
  \item{depensPar3}{(1)}
  \item{recruitsFromAdultSpawners}{("YES") "NO"/"YES"}
  \item{SRvariation}{("YES") "NO"/"YES"}
  \item{smoltToAdultVar}{("NO") "NO"/"YES"}
  \item{baseExploitationRate}{(0.67)}
  \item{includeManagementError}{("YES") "NO"/"YES"}
  \item{manageErrorA}{(65.3946) derived from estimates from Puget Sound for all except STL and WRS}
  \item{manageErrorB}{(0.0158) derived from estimates from Puget Sound for all except STL and WRS}
  \item{lowerEscThreshold}{(200)}
  \item{numYearsToAvg}{(5)}
  \item{runType}{("ER") "ER"/"Pop"}
  \item{bufferMin}{(0)}
}}
}
\value{
nothing is returned but the rav file is written.
}
\description{
This function takes the output from the DM model and creates a 
rav file that can be used forrunning the VRAP simulations. 
The parameter simInd indicates which MCMC sim to use. 
This takes the information from the input file in the DM 
spreadsheet (read in by readDMData) and the results of the 
MCMC simulations. Current version has no uncertainty in 
management error and base exploitation rate must be included 
as a parameter. This could be calculated with the SR params, 
maturation rates, natMort, and harvest.
}
