% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_find.R
\name{db_info}
\alias{db_info}
\title{extract variable information}
\usage{
db_info(df, attrib = "label", name = NULL)
}
\arguments{
\item{df}{a data frame (or similar object)}

\item{attrib}{name of attribute (default 'label')}

\item{name}{optionally, the name of the data frame for output}
}
\value{
a data frame
}
\description{
get names, associated labels (or any specific attribute), and
    class of variables in a data set
}
