% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_table.R
\name{db_table}
\alias{db_table}
\title{table wrapper}
\usage{
db_table(data, ...)
}
\arguments{
\item{data}{a data frame (or similar)}

\item{...}{variables (unquoted)}
}
\value{
a table of the variables
}
\description{
table for variables in a data frame
}
