% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedup.R
\name{dedup}
\alias{dedup}
\title{Deduplicate a data frame.}
\usage{
dedup(x, ...)
dup_rows(x)
dup_nrow(x)
dup_mark(x)
}
\arguments{
\item{x}{A data frame.}

\item{...}{Additional parameters passed to \code{unique()}.}
}
\description{
Deduplicates a data frame.
}
\details{
\code{dedup()} removes duplicates from a data frame. Same as casting \code{unique()} on a data frame.

\code{dup_rows()} finds the duplicated rows in a data frame.

\code{dup_nrow} counts the number of duplicates.

\code{dup_mark} codes duplicate rows with a 1; otherwise, 0.
}
\examples{
TBD

}
\seealso{
\url{https://github.com/robertschnitman/dm}
}
