% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-write.R
\name{dm_write_csv}
\alias{dm_write_csv}
\alias{dm_read_csv}
\alias{dm_write_zip}
\alias{dm_read_zip}
\alias{dm_write_xlsx}
\alias{dm_read_xlsx}
\title{Read/write a local dm from/to files}
\usage{
dm_write_csv(dm, csv_directory)

dm_read_csv(csv_directory)

dm_write_zip(dm, zip_file_path, overwrite = FALSE)

dm_read_zip(zip_file_path)

dm_write_xlsx(dm, xlsx_file_path, overwrite = FALSE)

dm_read_xlsx(xlsx_file_path)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{csv_directory}{For \code{dm_write_csv()}: The path to a non-existent or empty
directory to write the \code{csv}-files defining the \code{dm} to.

For \code{dm_read_csv()}: The path to the directory containing the \code{dm} as \code{csv-files}.}

\item{zip_file_path}{For \code{dm_write_zip}: The file path to the \code{zip}-file to write.

For \code{dm_read_zip}: The file path to the \code{zip}-file to read the \code{dm} from.}

\item{overwrite}{Logical, default: \code{FALSE}. In case the file already exists, should it be overwritten?}

\item{xlsx_file_path}{For \code{dm_write_xlsx()}: The file path to the \code{xlsx}-file to write.

For \code{dm_read_xlsx()}: The file path to the \code{xlsx}-file to read the \code{dm} from.}
}
\value{
\code{read}-family: A \code{dm} object.

\code{write}-family: The \code{dm}, invisibly.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Store a local \code{dm} in a file or directory and retrieve it later.
}
\details{
\code{dm_write_csv()}: write a \code{dm} to a collection of \code{csv}-files to a non-existent or empty directory.

\code{dm_read_csv()}: read a \code{dm} from a directory created using \code{dm_write_csv()}.

\code{dm_write_zip()}: write a \code{dm} to a \code{zip}-file containing a collection of \code{csv}-files.

\code{dm_read_zip()}: read a \code{dm} from a \code{zip}-file created using \code{dm_write_zip()}.

\code{dm_write_xlsx()}: write a \code{dm} to an \code{xlsx}-file containing several tables defining the \code{dm}.

\code{dm_read_xlsx()}: read a \code{dm} from an \code{xlsx}-file created using \code{dm_write_xlsx()}.
}
\examples{
\dontrun{
dm_write_csv(dm_nycflights13(), "nyc_dm_as_csv")
dm_read_csv("nyc_dm_as_csv")

dm_write_zip(dm_nycflights13(), "dm.zip")
dm_read_zip("dm.zip")

dm_write_xlsx(dm_nycflights13(), "dm.xlsx")
dm_read_xlsx("dm.xlsx")
}
}
