% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_primary_keys.R
\name{dm_add_primary_key}
\alias{dm_add_primary_key}
\title{Mark a column of a table in a \code{dm}-object as its primary key}
\usage{
dm_add_primary_key(dm, table, column, check_if_unique_key = TRUE)
}
\description{
\code{dm_add_primary_key()} first checks , if the given column
is a unique key of the given table in the \code{dm}-object. If yes, it then marks
that column as the table's primary key in the \code{data_model}-part of the \code{dm}-object.
}
\examples{
\dontrun{
iris_mtcars_dm_obj <- as_dm(
    list(iris = iris \%>\%
             mutate(id_column = row_number()) \%>\%
             select(id_column, everything()),
         mtcars = mtcars, iris_2 = iris)
     )
# the following works
dm_add_primary_key(iris_mtcars_dm_obj, "iris", "id_column")
dm_add_primary_key(iris_mtcars_dm_obj, "iris", id_column)
dm_add_primary_key(iris_mtcars_dm_obj, "iris", "Sepal.Width", check_if_unique_key = FALSE)

# the following does not work
dm_add_primary_key(iris_mtcars_dm_obj, "iris", "Sepal.Width")
}
}
