% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-tbl.R
\name{dm_add_tbl}
\alias{dm_add_tbl}
\title{Add tables to a \code{\link{dm}}}
\usage{
dm_add_tbl(.dm, ..., .repair = "unique", .quiet = FALSE)
}
\arguments{
\item{...}{One or more tables to add to the \code{dm}.
If no explicit name is given, the name of the expression is used.}

\item{.repair}{Treatment of problematic column names:
\itemize{
\item \code{"minimal"}: No name repair or checks, beyond basic existence,
\item \code{"unique"}: Make sure names are unique and not empty,
\item \code{"check_unique"}: (default value), no name repair, but check they are
\code{unique},
\item \code{"universal"}: Make the names \code{unique} and syntactic
\item a function: apply custom name repair (e.g., \code{.name_repair = make.names}
for names in the style of base R).
\item A purrr-style anonymous function, see \code{\link[rlang:as_function]{rlang::as_function()}}
}

This argument is passed on as \code{repair} to \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}.
See there for more details on these terms and the strategies used
to enforce them.}

\item{.quiet}{Whether to suppress messages about name repair.}

\item{dm}{A \code{\link{dm}} object.}
}
\value{
The initial \code{dm} with the additional table(s).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Adds one or more new tables to a \code{\link{dm}}.
Existing tables are not overwritten.
}
\examples{
dm() \%>\%
  dm_add_tbl(mtcars, flowers = iris)

# renaming table names if necessary (depending on the `repair` argument)
dm() \%>\%
  dm_add_tbl(new_tbl = mtcars, new_tbl = iris)
}
\seealso{
\code{\link[=dm_mutate_tbl]{dm_mutate_tbl()}}, \code{\link[=dm_rm_tbl]{dm_rm_tbl()}}
}
