% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{dm_bind}
\alias{dm_bind}
\title{Merge several \code{dm}}
\usage{
dm_bind(..., repair = "check_unique", quiet = FALSE)
}
\arguments{
\item{...}{\code{dm} objects to bind together.}

\item{repair}{Treatment of problematic column names:
\itemize{
\item \code{"minimal"}: No name repair or checks, beyond basic existence,
\item \code{"unique"}: Make sure names are unique and not empty,
\item \code{"check_unique"}: (default value), no name repair, but check they are
\code{unique},
\item \code{"universal"}: Make the names \code{unique} and syntactic
\item a function: apply custom name repair (e.g., \code{.name_repair = make.names}
for names in the style of base R).
\item A purrr-style anonymous function, see \code{\link[rlang:as_function]{rlang::as_function()}}
}

This argument is passed on as \code{repair} to \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}.
See there for more details on these terms and the strategies used
to enforce them.}

\item{quiet}{Whether to suppress messages about name repair.}
}
\value{
\code{dm} containing the tables and key relations of all \code{dm} objects.
}
\description{
Create a single \code{dm} from two or more \code{dm} objects.
}
\details{
The \code{dm} objects have to share the same \code{src}. By default table names need to be unique.
}
\examples{
\dontshow{if (rlang::is_installed("nycflights13")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dm_1 <- dm_nycflights13()
dm_2 <- dm(mtcars, iris)
dm_bind(dm_1, dm_2)
\dontshow{\}) # examplesIf}
}
