% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_keys.R
\name{dm_check_if_table_has_primary_key}
\alias{dm_check_if_table_has_primary_key}
\title{Does a table of a \code{dm}-object have a column set as primary key?}
\usage{
dm_check_if_table_has_primary_key(dm, table)
}
\description{
\code{dm_check_if_table_has_primary_key()} checks in the \code{data_model} part
of the \code{dm}-object if a given table has a column marked as primary key.
}
\examples{
\dontrun{
library(nycflights13)
library(dplyr)

nycflights_dm <- dm(src_df(pkg = "nycflights13"))
dm_obj_with_keys <- dm_add_primary_key(nycflights_dm, "planes", "tailnum")

dm_obj_with_keys \%>\%
  dm_check_if_table_has_primary_key("planes")
}

}
