% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disentangle.R
\name{dm_clone_pt}
\alias{dm_clone_pt}
\title{Remove cycles from a \code{\link{dm}}}
\usage{
dm_clone_pt(dm, parent_table, naming_template = NULL, quiet = FALSE)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{parent_table}{A table of the \code{dm} with incoming foreign keys}

\item{naming_template}{Naming template for the tables to be created as a length 1 character variable.
Default \code{NULL} leads to names of the form: \verb{<parent_table>_<row_number>}.
Keywords are given in {glue}-style:
\itemize{
\item \code{{.pt}}  : original name of parent table
\item \code{{.pkc}} : primary key column name(s) of original parent table. Multiple columns are separated by an underscore.
\item \code{{.ct}}  : name of child table
\item \code{{.fkc}} : foreign key column name(s). Multiple columns are separated by an underscore
\item \code{{.n}} : new table number: integer counting the tables based on each original parent table
}}

\item{quiet}{Boolean.
By default, this function lists the renamed columns in a message, pass \code{TRUE} to suppress this message.}
}
\value{
A \code{dm} object with untangled foreign key relations regarding \code{parent_table}.
}
\description{
Multiply a specific parent table based on its foreign key relations
}
\details{
Firstly, a plan for the multiplication of the given \code{parent_table} is created.
This depends on the number of incoming foreign keys and the number of possible paths
between the \code{parent_table} and its child tables.

Subsequently the table is being cloned as many times as is necessary to avoid
cycles in the graph representation of the \code{dm} originating from this table,
distributing the originally incoming foreign keys among the clones.

For more details about the multiplication plan see \code{\link{dm_disentangle}}.
}
\examples{
dm_clone_pt(dm_nycflights13(), flights)
dm_clone_pt(dm_nycflights13(), airports)
}
\seealso{
Other remove cyclic FK relations: 
\code{\link{dm_disentangle}()}
}
\concept{remove cyclic FK relations}
