% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{dm_derive}
\alias{dm_derive}
\title{get or reset documentation for derivation}
\usage{
dm_derive(kill = FALSE, prompt = TRUE)
}
\arguments{
\item{kill}{if \code{TRUE} deletes all documentation}

\item{prompt}{if \code{TRUE} prompts before deleting}
}
\description{
get or reset the list that has documented a derivation process
}
