% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disentangle.R
\name{dm_disentangle}
\alias{dm_disentangle}
\title{Remove cycles from a \code{\link{dm}}}
\usage{
dm_disentangle(dm)
}
\arguments{
\item{dm}{A \code{dm} object.}
}
\value{
A cycle-free \code{\link{dm}} object.
}
\description{
Produce a cycle-free \code{\link{dm}} by multiplying parent tables
}
\details{
Starting from the graph representation of the \code{\link{dm}}, it is tested which
of the graph's components (connected undirectional subgraphs) has a cycle.
Now taking the direction of the relations into account, for components with a cycle
it is now it is determined which vertices (here: parent tables) have both:
\itemize{
\item 2 or more incoming foreign keys
\item multiple possible (undirected) paths between the participating tables in at last one of those foreign keys
}

Those tables are copied and reinserted into the \code{\link{dm}} with a new name while distributing the
foreign keys among the newly created tables.

This function is a no-op if:
\itemize{
\item no cycles are detected.
\item at least one "endless" cycles is detected (possible to walk endlessly in direction of arrows).
}
}
\examples{
dm_disentangle(dm_nycflights13())
dm_disentangle(dm_nycflights13(cycle = TRUE))
}
