% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{dm_doc}
\alias{dm_doc}
\title{get or reset documentation for dm}
\usage{
dm_doc(kill = FALSE, prompt = TRUE)
}
\arguments{
\item{kill}{if \code{TRUE} deletes all documentation}

\item{prompt}{if \code{TRUE} prompts before deleting}
}
\description{
get or reset the list that has documentet the data management process done by dm
}
