% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{dm_doc_set}
\alias{dm_doc_set}
\title{add to dm documentation}
\usage{
dm_doc_set(name, value)
}
\arguments{
\item{name}{name of new list element}

\item{value}{value of new list element}
}
\description{
add list element to dm documentation (internal use)
}
