% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dm_filter2dist}
\alias{dm_filter2dist}
\title{filter distances}
\usage{
dm_filter2dist(f = NULL, plot = TRUE, m = 1)
}
\arguments{
\item{f}{a dm_filter}

\item{plot}{logical; plot cluster dendogram?}

\item{m}{1 or 2; method (needs documentation)}
}
\description{
some kind of metric on the filters
}
\details{
this needs thinking about, but the idea was that sequantially
    looking at inclusion/exclusion of filters can be very misleading
}
