% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dm_filter2latexlist}
\alias{dm_filter2latexlist}
\title{filtering as latex list}
\usage{
dm_filter2latexlist(f = NULL)
}
\arguments{
\item{f}{dm_filter object or similar}
}
\value{
NULL and side effects
}
\description{
...
}
