% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{dm_filter_set}
\alias{dm_filter_set}
\title{add to filter documentation}
\usage{
dm_filter_set(name, value)
}
\arguments{
\item{name}{name of new list element}

\item{value}{value of new list element}
}
\description{
add list element to filter documentation (internal use)
}
