% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial.R
\name{dm_financial}
\alias{dm_financial}
\alias{dm_financial_sqlite}
\title{Creates a dm object for the Financial data}
\usage{
dm_financial()

dm_financial_sqlite()
}
\value{
A \code{dm} object.
}
\description{
\code{dm_financial()} creates an example \code{\link{dm}} object from the tables at https://relational.fel.cvut.cz/dataset/Financial.
The connection is established once per session,
subsequent calls return the same connection.

\code{dm_financial_sqlite()} copies the data to a temporary SQLite database.
The data is downloaded once per session, subsequent calls return the same database.
The \code{trans} table is excluded due to its size.
}
\examples{
\dontshow{if (dm:::dm_has_financial() && rlang::is_installed("DiagrammeR")) withAutoprint(\{ # examplesIf}
dm_financial() \%>\%
  dm_draw()
\dontshow{\}) # examplesIf}
}
