% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial.R
\name{dm_financial_sqlite}
\alias{dm_financial_sqlite}
\title{Creates a dm object for the Financial data}
\usage{
dm_financial_sqlite()
}
\value{
A \code{dm} object.
}
\description{
Creates an example \code{\link{dm}} object from the tables at
\url{https://relational.fit.cvut.cz/dataset/Financial}.
The data is downloaded once per session,
subsequent calls return the same database.
The \code{trans} table is excluded due to its size.
}
\examples{
if (rlang::is_installed("RMariaDB") && rlang::is_installed("RSQLite")) {
  dm_financial_sqlite() \%>\%
    dm_draw()
}
}
