% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{dm_get_fk}
\alias{dm_get_fk}
\title{Retrieve the name of the column marked as a foreign key, pointing from one table of a \code{\link{dm}} to another table.}
\usage{
dm_get_fk(dm, table, ref_table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}.}

\item{ref_table}{The table that is referenced from \code{table}.}
}
\value{
A character vector with the column name(s) of \code{table}, pointing to the primary key of \code{ref_table}.
}
\description{
\code{dm_get_fk()} returns the name of the
column marked as foreign key of table \code{table} with respect to table \code{ref_table} within a \code{\link{dm}} object.
If no foreign key is set between the tables, an empty character vector is returned.
}
\examples{
dm_get_fk(dm_nycflights13(), flights, airports)
}
\seealso{
Other foreign key functions: 
\code{\link{dm_add_fk}()},
\code{\link{dm_enum_fk_candidates}()},
\code{\link{dm_get_all_fks}()},
\code{\link{dm_has_fk}()},
\code{\link{dm_rm_fk}()}
}
\concept{foreign key functions}
