% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-keys.R
\name{dm_get_pk}
\alias{dm_get_pk}
\title{Primary key column names}
\usage{
dm_get_pk(dm, table, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated because of its limited use
and its unintuitive return value.
 Use \code{\link[=dm_get_all_pks]{dm_get_all_pks()}} instead.
}
\keyword{internal}
