% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{dm_get_referencing_tables}
\alias{dm_get_referencing_tables}
\title{Get the names of a \code{\link{dm}}'s tables referencing a given table.}
\usage{
dm_get_referencing_tables(dm, table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}
}
\value{
Character vector of the names of the tables pointing to the primary
key of parameter \code{table}.
}
\description{
Get the names of a \code{\link{dm}}'s tables referencing a given table.
}
\seealso{
Other functions utilizing foreign key relations: \code{\link{dm_is_referenced}}
}
\concept{functions utilizing foreign key relations}
