% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{dm_import_csv}
\alias{dm_import_csv}
\alias{dm_export_csv}
\title{Import CSV files as dm object}
\usage{
dm_import_csv(path = ".", pk = NULL, fk = NULL)

dm_export_csv(x, path)
}
\arguments{
\item{path}{Path to directory from where csv files has to be read.
For \code{dm_export_csv} it is a path where csv files will be written to.}

\item{pk}{Specification of primary keys, not yet implemented.}

\item{fk}{Specification of foreign keys, not yet implemented.}

\item{x}{A \code{dm} object to be exported to csv files.}
}
\value{
For \code{dm_import_csv} a \code{dm} object.

For \code{dm_export_csv} \code{NULL} invisibly.
}
\description{
By specifying \code{path} to a directory of csv files, files are read and
\code{dm} object is created.

\code{dm_export_csv} is a helper function to easily save existing \code{dm} object
into csv files.
}
\examples{
## extract example dm into csv files
x <- dm_nycflights13()
dm_export_csv(x, path="csv")
list.files("csv")
## read csv files into db
dm <- dm_import_csv("csv")
dm
}
\seealso{
\itemize{
\item \code{\link[=dm_nycflights13]{dm_nycflights13()}}  for creating an example \code{dm} object
}
}
