% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{dm_is_referenced}
\alias{dm_is_referenced}
\title{Is a \code{\link{dm}}'s table referenced by another table?}
\usage{
dm_is_referenced(dm, table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}
}
\value{
\code{TRUE}, if at least one foreign key exists, pointing to the primary
key of parameter \code{table}, \code{FALSE} otherwise.
}
\description{
Is a \code{\link{dm}}'s table referenced by another table?
}
\seealso{
Other functions utilizing foreign key relations: \code{\link{dm_get_referencing_tables}}
}
\concept{functions utilizing foreign key relations}
