% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_kill_attr.R
\name{dm_kill_attr}
\alias{dm_kill_attr}
\title{remove attributes}
\usage{
dm_kill_attr(x, attr)
}
\arguments{
\item{x}{data frame or similar}

\item{attr}{name of one or more attribute to nullify}
}
\description{
remove unwanted attributes from data frame or similar object
}
