% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-helpers.R
\name{dm_nrow}
\alias{dm_nrow}
\title{Number of rows}
\usage{
dm_nrow(dm)
}
\arguments{
\item{dm}{A \code{\link{dm}} object.}
}
\description{
Returns a named vector with the number of rows for each table.
}
