% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_recode2latex.R
\name{dm_recode2latex}
\alias{dm_recode2latex}
\title{Turn all recoded tables in doc into latex}
\usage{
dm_recode2latex(
  doc = NULL,
  file = "",
  lab.prefix = "tab:recode_",
  where = "htb",
  clearpage = NULL,
  ...
)
}
\arguments{
\item{doc}{the dm documentation (if \code{NULL} we look in \code{dm_doc()})}

\item{file}{'file' argument for \code{Hmisc::latex} (default = "")}

\item{lab.prefix}{prefix for labels for LaTeX tables (default =
"tab:recode_") which concatenates with the variable name}

\item{where}{LaTeX indicator of position for float (Default: "htb")}

\item{clearpage}{how many floats before a LaTeX clearpage? Recycled if
necessary}

\item{...}{arguments passed to \code{Hmisc::latex}.}
}
\description{
...
}
