% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_keys.R
\name{dm_remove_primary_key}
\alias{dm_remove_primary_key}
\title{Remove primary key from a table in a \code{dm}-object}
\usage{
dm_remove_primary_key(dm, table)
}
\description{
\code{dm_remove_primary_key()} removes a potentially set primary key from a table in the
underlying \code{data_model}-object and otherwise leaves the \code{dm}-object untouched.
}
\examples{
\dontrun{
library(nycflights13)
library(dplyr)

nycflights_dm <- dm(src_df(pkg = "nycflights13"))

# the following works
dm_obj_with_keys <- dm_add_primary_key(nycflights_dm, "planes", "tailnum") \%>\%
  dm_add_primary_key("airports", faa)

dm_obj_with_keys \%>\%
  dm_remove_primary_key("airports") \%>\%
  dm_check_if_table_has_primary_key("planes")

dm_obj_with_keys \%>\%
  dm_remove_primary_key("planes") \%>\%
  dm_check_if_table_has_primary_key("planes")
}

}
