% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-tbl.R
\name{dm_rm_tbl}
\alias{dm_rm_tbl}
\title{Remove tables from a \code{\link{dm}}}
\usage{
dm_rm_tbl(dm, ...)
}
\arguments{
\item{dm}{A \code{\link{dm}} object}

\item{...}{One or more unquoted tibble names to remove from the \code{dm}.}
}
\value{
The inital \code{dm} without the removed table(s).
}
\description{
Removes one or more tibbles from a \code{\link{dm}}.
}
\seealso{
\code{\link[=dm_add_tbl]{dm_add_tbl()}}, \code{\link[=dm_select_tbl]{dm_select_tbl()}}
}
