% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-surgery.R
\name{dm_separate_tbl}
\alias{dm_separate_tbl}
\title{Decompose a table into two linked tables}
\usage{
dm_separate_tbl(
  dm,
  table,
  new_key_column,
  ...,
  new_table_name = NULL,
  repair = "check_unique",
  quiet = FALSE,
  on_delete = c("no_action", "cascade")
)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}.}

\item{new_key_column}{Name of the identifier column (primary key column) for the
newly created lookup (parent) table. A column of this name is also added in \code{table}.}

\item{...}{The columns to be extracted from \code{table}. \code{tidyselect} is supported, see
\code{\link[dplyr:select]{dplyr::select()}} for details on the semantics.}

\item{new_table_name}{Unquoted name for the new table. If \code{NULL}, the name will be
\code{table} with the suffix \verb{_lookup}.}

\item{repair}{Either a string or a function. If a string, it must
be one of \code{"check_unique"}, \code{"minimal"}, \code{"unique"}, or \code{"universal"}.
If a function, it is invoked with a vector of minimal names and must
return minimal names, otherwise an error is thrown.
\itemize{
\item Minimal names are never \code{NULL} or \code{NA}. When an element doesn't
have a name, its minimal name is an empty string.
\item Unique names are unique. A suffix is appended to duplicate
names to make them unique.
\item Universal names are unique and syntactic, meaning that you can
safely use the names as variables without causing a syntax
error.
}

The \code{"check_unique"} option doesn't perform any name repair.
Instead, an error is raised if the names don't suit the
\code{"unique"} criteria.}

\item{quiet}{By default, the user is informed of any renaming
caused by repairing the names. This only concerns unique and
universal repairing. Set \code{quiet} to \code{TRUE} to silence the
messages.}

\item{on_delete}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Defines behavior if a row in the parent table is deleted.
- \code{"no_action"}, the default, means that no action is taken
and the operation is aborted if child rows exist
- \code{"cascade"} means that the child row is also deleted
This setting is picked up by \code{\link[=copy_dm_to]{copy_dm_to()}} with \code{set_key_constraints = TRUE},
and might be considered by \code{\link[=dm_rows_delete]{dm_rows_delete()}} in a future version.}
}
\value{
A \code{dm} with one of its tables split into two tables which are linked by
a foreign key relation.
}
\description{
Extract a lookup table from a table in a \code{dm}, linking the original table and the new table by a key.

\code{dm_separate_table()} accepts a \code{dm}, an unquoted table name, a name for the new column
which will be created to link the two tables and the names of the columns that will be
used in order to form the new table.

In addition you can specify the name of the newly created table and how possible
conflicts with existing table names should be handled.
}
\examples{
dm_nycflights13() \%>\%
  dm_separate_tbl(flights, ymd, year, month, day)
}
\seealso{
Other normalization: 
\code{\link{decompose_table}()},
\code{\link{dm_unite_tbls}()},
\code{\link{reunite_parent_child}()}
}
\concept{normalization}
