% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dm.R
\name{dm_test_load}
\alias{dm_test_load}
\title{Loads \code{dm}-objects into one or more registered sources}
\usage{
dm_test_load(x, name = NULL, srcs = dbplyr:::test_srcs$get(),
  ignore = character())
}
\description{
Works like \code{dbplyr::test_load()}, just for \code{dm}_objects.
}
\examples{
dbplyr::test_register_src("df", dplyr::src_df(env = new.env()))
dbplyr::test_register_src("sqlite", dplyr::src_sqlite(":memory:", create = TRUE))

dm_test_obj <- dm(dplyr::src_df(pkg = "nycflights13"))
dm_test_obj_srcs <- dm_test_load(dm_test_obj)

}
