% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_to_tibble.R
\name{dm_to_tibble}
\alias{dm_to_tibble}
\title{Wrap dm into a single tibble}
\usage{
dm_to_tibble(dm, root, silent = FALSE)
}
\arguments{
\item{dm}{A cycle free dm object}

\item{root}{Table to wrap the dm into (unquoted)}

\item{silent}{Whether to print the code that reverse the transformation. See details.

When silent is \code{FALSE} (default) we print the steps required to achieve
the reverse transformation without using a prototype. This is sequence of
calls to \code{dm()}, \code{dm_unpack_tbl()} and \code{dm_unnest_tbl()}.

The reverse transformation i generally not a perfect round trip since
\code{dm_to_tibble()} keeps only information related to \code{root}}
}
\value{
A tibble
}
\description{
\code{dm_to_tibble()} creates a tibble built from the \code{root} table and containing
all the data related to it through the relationships stored in the dm.
}
\examples{
dm_to_tibble(dm_nycflights13(), airlines)
}
\seealso{
\link{tibble_to_dm}, \link{dm_wrap}, \link{dm_unwrap},
\link{dm_wrap_all}, \link{dm_unwrap_all},
}
