% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-surgery.R
\name{dm_unite_tbls}
\alias{dm_unite_tbls}
\title{Merge two tables that are linked by a foreign key relation}
\usage{
dm_unite_tbls(dm, table_1, table_2, rm_key_col = TRUE)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table_1}{A table of the \code{dm} that is directly linked to \code{table_2}}

\item{table_2}{A table of the \code{dm} that is directly linked to \code{table_1}}

\item{rm_key_col}{Boolean, if \code{TRUE} (default), the FK column is removed after
joining the two tables. Otherwise the key column of the child table is kept.}
}
\value{
The original \code{dm} with two of its tables merged into one.
}
\description{
Join two tables together by their foreign key and update the
\code{dm} accordingly. This is similar to \code{dm_join_to_tbl()}, only that the resulting
table will be part of the \code{dm}. The name of the child table (the table with the
foreign key relation pointing away from it) will be used for the resulting table.

Foreign key relations are kept whenever possible.
}
\examples{
dm_nycflights13() \%>\%
  dm_unite_tbls(flights, planes)
}
\seealso{
Other normalization: 
\code{\link{decompose_table}()},
\code{\link{dm_separate_tbl}()},
\code{\link{reunite_parent_child}()}
}
\concept{normalization}
