% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_to_tibble.R
\name{dm_unwrap_all}
\alias{dm_unwrap_all}
\title{Unwrap a single table dm}
\usage{
dm_unwrap_all(dm, prototype)
}
\arguments{
\item{dm}{A dm}

\item{prototype}{A dm, might be an empty prototype.}
}
\value{
A dm
}
\description{
Unwrap a single table dm
}
\examples{
wrapped_dm <- dm_wrap_all(dm_nycflights13(), airlines, silent = TRUE)
dm_unwrap_all(wrapped_dm, dm_nycflights13())
}
\seealso{
\link{dm_wrap_all}, \link{dm_to_tibble}, \link{tibble_to_dm},
\link{dm_wrap}, \link{dm_unwrap},
}
