% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_to_tibble.R
\name{dm_wrap}
\alias{dm_wrap}
\alias{dm_pack_tbl}
\alias{dm_nest_tbl}
\title{Wrap a table inside its dm}
\usage{
dm_wrap(dm, table, into = NULL, silent = FALSE)

dm_pack_tbl(dm, table, into = NULL, silent = FALSE)

dm_nest_tbl(dm, table, into = NULL, silent = FALSE)
}
\arguments{
\item{dm}{A dm}

\item{table}{A table}

\item{into}{The table to wrap \code{table} into, optional as it can be guessed
from the foreign keys unambiguously but useful to be explicit.}

\item{silent}{if not silent (the default), the code to unwrap will be printed
\itemize{
\item \code{dm_nest_tbl()} will nest a given \code{table} into its parent, \code{table} itself
should not have children tables (i.e. it needs to be a \emph{terminal child table})
\item \code{dm_pack_tbl()} will pack a given \code{table} into its child, \code{table} itself
should not have parent tables (i.e. it needs to be a \emph{terminal parent table})
\item \code{dm_wrap()} is behave either the same as \code{dm_nest_tbl()} or \code{dm_pack_tbl()}
depending on which one is applicable.
}}
}
\description{
These functions pack or nest a table inside its neighbour table (parent or child)
}
\seealso{
\link{dm_wrap_all}, \link{dm_unwrap_all},
\link{dm_to_tibble}, \link{tibble_to_dm}
}
