% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_to_tibble.R
\name{dm_wrap_all}
\alias{dm_wrap_all}
\title{Wrap dm into a single tibble dm}
\usage{
dm_wrap_all(dm, root, silent = FALSE, strict = TRUE)
}
\arguments{
\item{dm}{A cycle free dm object}

\item{root}{Table to wrap the dm into (unquoted)}

\item{silent}{Whether to print the code that reverse the transformation. See details.

When silent is \code{FALSE} (default) we print the steps required to achieve
the reverse transformation without using a prototype. This is sequence of
calls to \code{dm()}, \code{dm_unpack_tbl()} and \code{dm_unnest_tbl()}.

The reverse transformation i generally not a perfect round trip since
\code{dm_to_tibble()} keeps only information related to \code{root}}

\item{strict}{Whether to fail for cyclic dms that cannot be wrapped into a
single table, if \code{FALSE} a partially wrapped dm will be returned

When silent is \code{FALSE} (default) we print the steps required to achieve
the reverse transformation without using a prototype. This is a sequence of
calls to \code{dm_unpack_tbl()} and \code{dm_unnest_tbl()}.

The reverse transformation i generally not a perfect round trip since
\code{dm_to_tibble()} keeps only information related to \code{root}}
}
\value{
A single table dm
}
\description{
\code{dm_wrap_all()} creates a single tibble tibble dm containing the \code{root} table
enhanced with all the data related to it through the relationships stored in the dm.
}
\examples{
dm_wrap_all(dm_nycflights13(), airlines)
}
\seealso{
\link{dm_unwrap_all},  \link{dm_to_tibble}, \link{tibble_to_dm},
\link{dm_wrap}, \link{dm_unwrap},
}
