% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom2.R
\name{dm_zoom2_to}
\alias{dm_zoom2_to}
\alias{dm_update_zoom2ed}
\alias{dm_insert_zoom2ed}
\title{Mark table for manipulation (v2)}
\usage{
dm_zoom2_to(dm, table)

dm_update_zoom2ed(zoomed_tbl)

dm_insert_zoom2ed(
  zoomed_tbl,
  new_tbl_name = NULL,
  repair = "unique",
  quiet = FALSE
)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}.}

\item{zoomed_tbl}{A \code{dm_keyed_tbl} object returned by \code{dm_zoom2_to()}
or modified via dplyr operations.}

\item{new_tbl_name}{Name of the new table.}

\item{repair}{Either a string or a function. If a string, it must be one of
\code{"check_unique"}, \code{"minimal"}, \code{"unique"}, \code{"universal"}, \code{"unique_quiet"},
or \code{"universal_quiet"}. If a function, it is invoked with a vector of
minimal names and must return minimal names, otherwise an error is thrown.
\itemize{
\item Minimal names are never \code{NULL} or \code{NA}. When an element doesn't
have a name, its minimal name is an empty string.
\item Unique names are unique. A suffix is appended to duplicate
names to make them unique.
\item Universal names are unique and syntactic, meaning that you can
safely use the names as variables without causing a syntax
error.
}

The \code{"check_unique"} option doesn't perform any name repair.
Instead, an error is raised if the names don't suit the
\code{"unique"} criteria.

The options \code{"unique_quiet"} and \code{"universal_quiet"} are here to help the
user who calls this function indirectly, via another function which exposes
\code{repair} but not \code{quiet}. Specifying \code{repair = "unique_quiet"} is like
specifying \verb{repair = "unique", quiet = TRUE}. When the \code{"*_quiet"} options
are used, any setting of \code{quiet} is silently overridden.}

\item{quiet}{By default, the user is informed of any renaming
caused by repairing the names. This only concerns unique and
universal repairing. Set \code{quiet} to \code{TRUE} to silence the
messages.

Users can silence the name repair messages by setting the
\code{"rlib_name_repair_verbosity"} global option to \code{"quiet"}.}
}
\value{
For \code{dm_zoom2_to()}: A \code{dm_keyed_tbl} object with \code{"dm_zoom2"} attributes.

For \code{dm_update_zoom2ed()} and \code{dm_insert_zoom2ed()}: A \code{dm} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{dm_zoom2_to()} zooms to the given table, returning a keyed table
that has the dm object as an attribute.
Key column tracking (primary and foreign) is the responsibility
of the \code{dm_keyed_tbl} object.

\code{dm_update_zoom2ed()} overwrites the originally zoomed table
with the manipulated table.

\code{dm_insert_zoom2ed()} adds the manipulated table as a new table to the dm.
}
\examples{
\dontshow{if (rlang::is_installed(c("nycflights13", "DiagrammeR"))) withAutoprint(\{ # examplesIf}
flights_keyed <- dm_zoom2_to(dm_nycflights13(), flights)

flights_keyed

flights_keyed_transformed <-
  flights_keyed \%>\%
  mutate(am_pm_dep = ifelse(dep_time < 1200, "am", "pm"))

# replace table `flights` with the zoomed table
flights_keyed_transformed \%>\%
  dm_update_zoom2ed()

# insert the zoomed table as a new table
flights_keyed_transformed \%>\%
  dm_insert_zoom2ed("extended_flights")
\dontshow{\}) # examplesIf}
}
