% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{dm_zoom_to_tbl}
\alias{dm_zoom_to_tbl}
\alias{dm_insert_zoomed_tbl}
\alias{dm_update_zoomed_tbl}
\alias{dm_zoom_out}
\title{Single out a table of a \code{dm}}
\usage{
dm_zoom_to_tbl(dm, table)

dm_insert_zoomed_tbl(dm, new_tbl_name = NULL, repair = "unique", quiet = FALSE)

dm_update_zoomed_tbl(dm)

dm_zoom_out(dm)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}.}

\item{new_tbl_name}{Name of the new table.}

\item{repair}{Either a string or a function. If a string, it must
be one of \code{"check_unique"}, \code{"minimal"}, \code{"unique"}, or \code{"universal"}.
If a function, it is invoked with a vector of minimal names and must
return minimal names, otherwise an error is thrown.
\itemize{
\item Minimal names are never \code{NULL} or \code{NA}. When an element doesn't
have a name, its minimal name is an empty string.
\item Unique names are unique. A suffix is appended to duplicate
names to make them unique.
\item Universal names are unique and syntactic, meaning that you can
safely use the names as variables without causing a syntax
error.
}

The \code{"check_unique"} option doesn't perform any name repair.
Instead, an error is raised if the names don't suit the
\code{"unique"} criteria.}

\item{quiet}{By default, the user is informed of any renaming
caused by repairing the names. This only concerns unique and
universal repairing. Set \code{quiet} to \code{TRUE} to silence the
messages.}
}
\value{
For \code{cdm_zoom_to_tbl()}: A \code{zoomed_dm} object

For \code{cdm_insert_zoomed_tbl()}, \code{cdm_update_zoomed_tbl()} and \code{cdm_zoomed_out()}: A \code{dm} object
}
\description{
Zooming to a table of a \code{\link{dm}} allows for the use of many \code{dplyr}-verbs directly on this table, while retaining the
context of the \code{dm} object.
}
\details{
\code{dm_zoom_to_tbl()}: zooms to the given table.

\code{dm_update_zoomed_tbl()}: overwrites the originally zoomed table with the manipulated table.
The filter conditions for the zoomed table are added to the original filter conditions.

\code{dm_insert_zoomed_tbl()}: adds a new table to the \code{dm}.

\code{dm_zoom_out()}: discards the zoomed table and returns the \code{dm} as it was before zooming.

Whenever possible, the key relations of the original table are transferred to the resulting table
when using \code{dm_insert_zoomed_tbl()} or \code{dm_update_zoomed_tbl()}.

Functions from \code{dplyr} that are supported for a \code{zoomed_dm}: \code{group_by()}, \code{summarise()}, \code{mutate()},
\code{transmute()}, \code{select()}, \code{rename()} and \code{ungroup()}.
You can use these functions just like you would
with a normal table.

\code{filter()} is also supported, but treated in a special way: the filter expression for the zoomed table is
stored in the \code{dm} and is treated in a way that depends on which function you use to return to a normal \code{dm}:
\enumerate{
\item \code{dm_zoom_out()}: all filter conditions for the zoomed table are discarded
\item \code{dm_update_zoomed_tbl()}: the filter conditions of the original table and those of the zoomed table are combined
\item \code{dm_insert_zoomed_tbl()}: the filter conditions of the original table stay there and those of the zoomed table are
transferred to the new table of the \code{dm}
}
}
\examples{
library(dplyr)
flights_zoomed <- cdm_zoom_to_tbl(cdm_nycflights13(), flights)

flights_zoomed

flights_zoomed_transformed <-
  flights_zoomed \%>\%
  mutate(am_pm_dep = if_else(dep_time < 1200, "am", "pm")) \%>\%
  # `by`-argument of `left_join()` can be explicitly given
  # otherwise the key-relation is used
  left_join(airports) \%>\%
  select(year:dep_time, am_pm_dep, everything())

# replace table `flights` with the zoomed table
cdm_update_zoomed_tbl(flights_zoomed_transformed)

# insert the zoomed table as a new table
cdm_insert_zoomed_tbl(flights_zoomed_transformed, extended_flights)

# discard the zoomed table
cdm_zoom_out(flights_zoomed_transformed)

}
