% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmd.R
\name{dmd}
\alias{dmd}
\title{derivation documentation}
\usage{
dmd(name, expr, dmd = NULL, label = NULL, overwrite = FALSE)
}
\arguments{
\item{name}{name of variable assigned}

\item{expr}{the expression for creating variable}

\item{dmd}{the documentation}

\item{label}{a label to assign to the new variable}
}
\description{
save documentation on derived variables
}
