% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dmf}
\alias{dmf}
\title{document filtering}
\usage{
dmf(f, name, comment)
}
\arguments{
\item{f}{logical vector}

\item{name}{character}

\item{comment}{description of filtering}
}
\value{
nuffin
}
\description{
stores filtering documentation in dm_filter in dm_envir
}
