% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dmf_create}
\alias{dmf_create}
\title{create filter}
\usage{
dmf_create(x = NULL)
}
\arguments{
\item{x}{an object from \code{dmf}}
}
\value{
logical vector
}
\description{
use dmf-created list to make overall filter
}
