% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key-helpers.R
\name{examine_if_subset}
\alias{examine_if_subset}
\title{Test if the values of one column are a subset of the values of another column}
\usage{
examine_if_subset(t1, c1, t2, c2)
}
\arguments{
\item{t1}{The data frame that contains column \code{c1}.}

\item{c1}{The column of \code{t1} that should only contain the values that are also present in column \code{c2} of data frame \code{t2}.}

\item{t2}{The data frame that contains column \code{c2}.}

\item{c2}{The column of the second data frame that has to contain all values of \code{c1} to avoid an error.}
}
\value{
Returns \code{t1}, invisibly, if the check is passed.
Otherwise an error is thrown and the reason for it is explained.
}
\description{
\code{examine_if_subset()} tests if the values of the chosen column \code{c1} of data frame \code{t1} are a subset of the values
of column \code{c2} of data frame \code{t2}.
}
\examples{
data_1 <- tibble::tibble(a = c(1, 2, 1), b = c(1, 4, 1), c = c(5, 6, 7))
data_2 <- tibble::tibble(a = c(1, 2, 3), b = c(4, 5, 6), c = c(7, 8, 9))
# this is passing:
examine_if_subset(data_1, a, data_2, a)

# this is failing:
try(examine_if_subset(data_2, a, data_1, a))
}
