% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected.R
\name{expected}
\alias{expected}
\title{Stop execution if a specific class is expected}
\usage{
expected(x, expect)
}
\arguments{
\item{x}{A scalar string.}
}
\description{
Stop execution if a specific class is expected.
}
\seealso{
\url{https://github.com/robertschnitman/dm}
}
