% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial-db-con.R
\name{financial_db_con}
\alias{financial_db_con}
\title{Connection to SQL Financial Database}
\usage{
financial_db_con(source = "relational.fit")
}
\arguments{
\item{source}{(type: character) which service to use, by default is
"relational.fit" but it can be set to "db-edu" as well}
}
\value{
A \code{MariaDBConnection} object
}
\description{
Connects to relational.fit.cvut.cz unless the service is
unavailable, in which case db-edu.pacha.dev is used as a fallback
}
\examples{
\dontshow{if (dm:::dm_has_financial() && rlang::is_installed("RMariaDB")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
financial_db_con()
\dontshow{\}) # examplesIf}
}
