% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmatch_weight_by_dist.R
\name{fmatch_weight_by_dist}
\alias{fmatch_weight_by_dist}
\title{ATT weights rescaled by matching distance}
\usage{
fmatch_weight_by_dist(tr, data, cl = "cl", dist = "dist", id = "id",
  trv = 1, method = "test", method.params = NULL)
}
\arguments{
\item{tr}{name of treatment variable}

\item{data}{data frame}

\item{cl}{name of cluster variable}

\item{dist}{name of distance variable}

\item{id}{name of id variable}

\item{trv}{value for treated in treatment variable}

\item{method}{name of method}

\item{method.params}{parameters (as list) passed to method}
}
\description{
assign (possibly) non-uniforms weights within matching cluster based on the
    intercluster distances
}
