% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows-db.R
\name{get_returned_rows}
\alias{get_returned_rows}
\title{Extract the RETURNING rows}
\usage{
get_returned_rows(x)
}
\arguments{
\item{x}{A lazy tbl.}
}
\value{
A tibble.
}
\description{
Extract the RETURNING rows produced by \code{rows_insert()}, \code{rows_update()},
\code{rows_upsert()}, or \verb{rows_delete().}
}
