% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepict2surv.R
\name{grepict2surv}
\alias{grepict2surv}
\title{turn grepict alias into survival data}
\usage{
grepict2surv(g, prefix = c(time = "t.", event = "ev."))
}
\arguments{
\item{g}{returned from grepict (long format, stacked)}

\item{prefix}{prefix for time and event variable}
}
\value{
a data frame
}
\description{
make a grepict long format into wide survival data variables
}
