% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepict_washout.R
\name{grepict_washout}
\alias{grepict_washout}
\title{left censor / wash out}
\usage{
grepict_washout(g, w = NULL, u = "days")
}
\arguments{
\item{g}{object from \code{grepict}}

\item{w}{a boundary time}

\item{u}{units for diff between 'begin' and 'end' (default 'days')}
}
\value{
a data frame like that returned from \code{grepict}
}
\description{
ignore events prior to some boundary time
}
