% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grepr.R
\name{grepr}
\alias{grepr}
\title{'grep' and return hits}
\usage{
grepr(pattern, x, index = FALSE, ...)
}
\arguments{
\item{pattern}{pattern to look for}

\item{x}{character string to inspect}

\item{index}{logical; whether to also return indexes of hits (defaul FALSE)}

\item{...}{arguments to be passed to \code{\link{grep}}}
}
\description{
This is a wrapper for \code{\link{grep}} to return not only the indexes
of the hits (optional), but the hits themselves. (This is mainly a helper function
for \code{find_var} but can at times be useful.)
}
\examples{
grepr(pattern="a", x=names(datasets::mtcars))
}
\seealso{
\code{\link{grep}}, \code{\link{grepl}}
}
