% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows-db.R
\name{has_returned_rows}
\alias{has_returned_rows}
\title{Does \code{x} have RETURNING rows?}
\usage{
has_returned_rows(x)
}
\arguments{
\item{x}{A lazy tbl.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Checks if \code{x} has stored RETURNING rows produced by \code{rows_insert()}, \code{rows_update()},
\code{rows_upsert()}, or \verb{rows_delete().}
}
