% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{Helper file for creating text for the reports.}
\usage{
info(RData.file)
}
\arguments{
\item{RData.file}{a saved DM run as a RData file with the objects "input","dat","result", and "tDat"}
}
\value{
a list with text strings.
}
\description{
This uses data in the RData file from a saved DM run 
to create text describing the model fit.
}
