% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.evenodd.R
\name{is.evenodd}
\alias{is.evenodd}
\alias{is.even}
\alias{is.odd}
\title{Logical statements for even and odd values.}
\usage{
is.even(x)

is.odd(x)
}
\arguments{
\item{x}{Numeric (integer).}
}
\value{
Boolean.
}
\description{
Logical statements for even and odd values.
}
\examples{
is.even(1)
sapply(mtcars$gear, is.even)
is.odd(2)
sapply(mtcars$gear, is.odd)

}
\seealso{
\url{https://github.com/robertschnitman/afp}; `iseven()` and `isodd` in \url{https://docs.julialang.org/en/release-0.4/stdlib/numbers/}
}
