% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_nest.R
\name{json_nest}
\alias{json_nest}
\title{JSON nest}
\usage{
json_nest(.data, ..., .names_sep = NULL)
}
\arguments{
\item{.data}{A data frame, a data frame extension (e.g. a tibble), or  a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{...}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> Columns to pack, specified
using name-variable pairs of the form \code{new_col = c(col1, col2, col3)}.
 The right hand side can be any valid tidy select expression.}

\item{.names_sep}{If \code{NULL}, the default, the names will be left as is.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A wrapper around \code{\link[tidyr:nest]{tidyr::nest()}} which stores the nested data into JSON columns.
}
\examples{
df <- tibble::tibble(x = c(1, 1, 1, 2, 2, 3), y = 1:6, z = 6:1)
nested <- json_nest(df, data = c(y, z))
nested
}
\seealso{
\code{\link[tidyr:nest]{tidyr::nest()}}, \code{\link[=json_nest_join]{json_nest_join()}}
}
