% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.R
\name{keep}
\alias{keep}
\title{Keep rows a la Stata}
\usage{
keep(x, ...)
}
\arguments{
\item{x}{An object, typically a data frame.}

\item{...}{Parameters passed to \code{subset}.}
}
\description{
A shorthand for \code{subset}.
}
\seealso{
\url{https://github.com/robertschnitman/dm}, \url{https://www.stata.com/manuals13/gsw12.pdf}
}
