% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_libraries.R
\name{load_libraries}
\alias{load_libraries}
\title{Load libraries}
\usage{
load_libraries(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
Load libraries, and install them if not already installed.
}
\seealso{
\url{https://github.com/robertschnitman/dm}
}
