% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numNA.R
\name{numNA}
\alias{numNA}
\title{Count the number of missing values in a dataset}
\usage{
numNA(x, m)
rowNA(x)
colNA(x)
}
\arguments{
\item{x}{An object, usually a data frame.}

\item{m}{Margin. 1 for row-wise, 2 for column-wise.}
}
\value{
A vector.
}
\description{
Count the number of missing values in a dataset.
}
\details{
Similar to Stata, \code{numNA} counts the number of missing values: if \code{m} is set to 1, then it counts the number of missing values per row; if set to 2, then it counts the number of missing values per column; otherwise, it counts the total number of missing values.
}
\examples{
numNA(airquality)
rowNA(airquality)
colNA(airquality)

}
\seealso{
\url{https://github.com/robertschnitman/dm}
}
