% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmatch_weight_by_dist.R
\name{optmatch_distances}
\alias{optmatch_distances}
\title{matching distances}
\usage{
optmatch_distances(matchobj, distance, data)
}
\arguments{
\item{matchobj}{matching object, \code{fullmatch} or \code{pairmatch} (from
the optmatch package)}

\item{distance}{distance matrix created by \code{match_on}}

\item{data}{data frame}
}
\description{
wrapper for \code{matched.distances}
}
