% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\docType{data}
\name{opts_dm}
\alias{opts_dm}
\title{desc options}
\format{
An object of class \code{list} of length 3.
}
\usage{
opts_dm
}
\description{
This list tries to mimic the behavior of opts_chunk from knitr.
Currently these values are maintained with the functions in (the list)
\code{opts_desc}:
\itemize{
\item get - get the current values
\item set - set new values
\item restore - restore default values
}
}
\keyword{datasets}
