% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parity.R
\name{parity}
\alias{parity}
\alias{is.even}
\alias{is.odd}
\alias{get_even}
\alias{get_odd}
\title{Parity functions for determining even and odd values.}
\usage{
is.even(x)

is.odd(x)

get_even(x)

get_odd(x)

parity(x)
}
\arguments{
\item{x}{Numeric (integer).}
}
\value{
Vector (Boolean, numeric, character).
}
\description{
Parity functions for determining even and odd values.
}
\examples{
is.even(1)
is.even(mtcars$carb)
is.odd(2)
is.odd(mtcars$carb)
get_odd(mtcars$carb)
get_even(mtcars$carb)
parity(mtcars$carb)

}
\seealso{
\url{https://github.com/robertschnitman/afp}; `iseven()` and `isodd` in \url{https://docs.julialang.org/en/release-0.4/stdlib/numbers/}
}
