% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_search.R
\name{pattern_conc_search}
\alias{pattern_conc_search}
\title{pattern helper (specific)}
\usage{
pattern_conc_search(s, code.sep = " ")
}
\arguments{
\item{s}{string to match initial part of code}

\item{code.sep}{the string that servers as separator in the concatenation}
}
\description{
get the pattern for correct (simple) searches, characterized by providing the
initial sequence of the string wanted, where the code searched is
concatenated. Note: this function mainly serves as an example for the author
to remember how to do this! Be aware if your search strings are more
complicated as the author is no expert in regular expression.
}
\examples{
x <- c("ABC", "", "FFF XXX", "GGG  ABC FOO",
       "BAR  ABC", "FOO BAR ABB", " ", ";")
s.term = c("ABC", "X")
grepr(pattern = pattern_conc_search(s.term), x = x)
## sample example, different separator:
y <- gsub(pattern = " ", replacement = ";", x)
grepr(pattern = pattern_conc_search(s.term, code.sep = ";"), x = y)
}
