% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAgeComp.R
\name{plotAgeComp}
\alias{plotAgeComp}
\title{Create plots of age composition}
\usage{
plotAgeComp(dmObj, proportions = TRUE)
}
\arguments{
\item{dmObj}{a saved DM object (list) from runModel() with the objects "input","dat","result", "bdat" and "tDat"}

\item{proportions}{If proportions=TRUE, the age composition proportions are plotted.  If proportions=FALSE, the number of samples in each age are plotted.}
}
\value{
The plot is made.
}
\description{
This function is called by plotResults().  Age composition data versus calendar year. The years with age composition data are represented with a stacked barplot. With age 2 on the bottom (black) and age 5 on the top (white). The numbers above each bar represent the age composition sample size.
}
