% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCapacity.R
\name{plotCapacity}
\alias{plotCapacity}
\title{Create plots of capacity posteriors}
\usage{
plotCapacity(dmObj, xLab = "Capacity parameter", plotType = "histogram",
  binSize = NULL, xLims = NULL)
}
\arguments{
\item{dmObj}{a saved DM object (list) from runModel() with the objects "input","dat","result", "bdat" and "tDat"}

\item{xLab}{x-axis label}

\item{plotType}{("histogram")}

\item{binSize}{The width of the bin if the plot type is histogram.}

\item{xLims}{x-axis limits}
}
\value{
The plot is made.
}
\description{
Capacity posterior distibution.  This is the posterior and prior distributions for the spawner-recruit capacity parameter.
}
