% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoefs.R
\name{plotCoefs}
\alias{plotCoefs}
\title{Create plots of covariate coefficients}
\usage{
plotCoefs(dmObj, xLim = c(-3, 3), xLabs = c("Marine Survival coefficient",
  "Flow coefficient"))
}
\arguments{
\item{dmObj}{a saved DM object (list) from runModel() with the objects "input","dat","result", "bdat" and "tDat"}

\item{xLim}{x-axis limits}

\item{xLabs}{x-axis labels}
}
\value{
The plot is made.
}
