% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPriors.R
\name{plotPriors}
\alias{plotPriors}
\title{Create plots of priors for productivity and log capacity.}
\usage{
plotPriors(priors)
}
\arguments{
\item{priors}{List of priors parameters.  Created with createPriors().}
}
\value{
The plot is made.
}
\description{
This function is used by plotResults()
}
