% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProductivity.R
\name{plotProductivity}
\alias{plotProductivity}
\title{Create plots of productivity posteriors}
\usage{
plotProductivity(dmObj, xLab = "Productivity parameter",
  plotType = "histogram", binSize = NULL, xLims = NULL)
}
\arguments{
\item{dmObj}{a saved DM object (list) from runModel() with the objects "input","dat","result", "bdat" and "tDat"}

\item{xLab}{x-axis label}

\item{plotType}{("histogram")}

\item{binSize}{The width of the bin if the plot type is histogram.}

\item{xLims}{x-axis limits}
}
\value{
The plot is made.
}
