% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{plotResults}
\alias{plotResults}
\title{Create plots from DM runs}
\usage{
plotResults(dmObj, plotType = c("a.and.p.data", "priors", "SR", "recruits",
  "covariates", "harvest", "EscData", "EscAndAge", "escapement", "barPlot",
  "ageComp"), plotDest = "default", plotName = "tmp")
}
\arguments{
\item{dmObj}{a saved DM object (list) from runModel() with the objects "input","dat","result", and "tDat"}

\item{plotType}{The plots and order in which to make them
\describe{
\item{covariates}{}
\item{harvest}{}
\item{EscData}{}
\item{EscAndAge}{}
\item{escapement}{}
\item{SR}{}
\item{recruits}{recruit residuals, recruits/spawners #and effective harvest on recruits}
\item{barPlot}{a bar plot showing spawners and recruits broken down by pHOS and harvested}
\item{ageComp}{}
\item{priors}{Priors used in the DM model}
\item{a.and.p.data}{Raw data from the A and P input file.}
}}

\item{plotDest}{Plot destination. Can be "screen", "png", "default", "none". If "screen", separate graphics windows are opened. If "none", nothing is plotted but the figure captions are returned.}

\item{plotName}{prefix to apply to filename for generated plots if plotDest="png"}
}
\value{
The plot is saved to a file or plotted to screen or default device depending on plotDest. A vector of plot captions is returned invisibly.
}
\description{
Create several plots based on the DM model runs. This calls the
functions for various plots: plotPriors, plotPosteriors, plotAgeComp, plotCoefs
}
