% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimple.R
\name{plotSimple}
\alias{plotSimple}
\title{Create posterior credible interval plots from multiple DM runs}
\usage{
plotSimple(runList, paramList = c("obsSD", "procSD", "prod", "logCap"),
  cols = 4)
}
\arguments{
\item{runList}{a list of multiple run objects (dmObj).  Each run object in the list is from one call of runModel}

\item{paramList}{The parameters you want to compare from the model output.}

\item{cols}{The number of columns of figures in the plot.}
}
\value{
The plot is made.
}
