% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{print.dm}
\alias{print.dm}
\title{print 'dm' object}
\usage{
\method{print}{dm}(x, ..., print = TRUE)
}
\arguments{
\item{x}{a 'dm' object}

\item{...}{arguments passed to print.data.frame}

\item{print}{if \code{FALSE} then an data frame is returned}
}
\value{
possibly a data frame
}
\description{
prints a data frame version of selected info in a 'dm' object or returns
    that data frame
}
