% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmd.R
\name{print.dm_derive}
\alias{print.dm_derive}
\title{print 'dm_derive' object}
\usage{
\method{print}{dm_derive}(x, ..., print = TRUE)
}
\arguments{
\item{x}{a 'dm_derive' object}

\item{...}{arguments passed to print.data.frame}

\item{print}{if \code{FALSE} then a data.frame is returned}
}
\description{
prints a data frame version of selected info in a 'dm_derive'
    object or returns that data frame
}
