% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{print.dm_filter}
\alias{print.dm_filter}
\title{print 'dm_filter' object}
\usage{
\method{print}{dm_filter}(x, ..., print = TRUE, seq = NULL, id = NULL, verbose = TRUE)
}
\arguments{
\item{x}{a 'dm_filter' object}

\item{...}{arguments passed to print.data.frame}

\item{print}{if \code{FALSE} then an data frame is returned}

\item{seq}{order in which to look at the filters}

\item{id}{id-vector for units, necessarily in the same order as filter}

\item{verbose}{logical, want funtion to provide messages?}
}
\value{
possibly a data frame
}
\description{
prints a data frame version of selected info in a 'dm_filter'
    object or returns that data frame
}
