% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_dm.R
\name{quick_dm}
\alias{quick_dm}
\title{Build a dm with compact syntax}
\usage{
quick_dm(...)
}
\arguments{
\item{...}{tables or constraints
\itemize{
\item Tables can be provided just like in \code{dm()}
\item \code{~ table} is used to define an empty table in the dm, this is usually not necessary
because \code{table} names are infered from constraints
\item fk to pk relationships are created with \verb{->>},  fk to non pk parent keys are created with \verb{->},
below we illustrate with \verb{->>}
\item \code{child[child_fk_cols] ->> parent[parent_key_cols]} will create a fk relationship,
it will also create the tables, columns and pk if relevant
\item \code{child[child_fk_cols] ->> parent} is the same as \code{child[child_fk_cols] ->> parent[child_fk_cols]},
and \code{child ->> parent[parent_key_cols]} is the same as
\code{child[parent_key_cols] ->> parent[parent_key_cols]} (i.e. no need to repeat
keys on both sides)
\item \code{NULL ->> table[pk_cols]} is used to define a primary key without mentioning
a child table
\item if \code{cols} is of length 1 the syntaxes \code{table[cols]},
\code{table[[cols]]} and \code{table$cols_value}
}}
}
\value{
a dm
}
\description{
\code{quick_dm()} helps you build a syntax with minimal typing. \verb{->>} and \verb{->}
are used here to define foreign key constraints and means something completely
different from what they do in base R. Columns that are not provided explicitly
are built as logical.
}
\examples{
# minimal dms
dm1 <- quick_dm(child$fk ->> parent$pk)
dm_draw(dm1)

dm2 <- quick_dm(~ table_without_key, NULL ->> table_with_pk$pk)
dm_draw(dm2)

# reproduce structure of some dms from the package
dm_nycflights13_quick <-
  quick_dm(
    flights$carrier ->> airlines,
    flights$tailnum ->> planes,
    flights$origin ->> airports,
    flights[c("origin", "time_hour")] ->> weather,
  )
dm_draw(dm_nycflights13_quick)

dm_for_filter_quick <-
  quick_dm(
    tf_2$d ->> tf_1$a,
    NULL ->> tf_2$c,
    tf_2[c("e","e1")] ->> tf_3[c("f", "f1")],
    tf_4[c("j","j1")] ->> tf_3[c("f", "f1")],
    tf_5$l ->> tf_4$h,
    tf_5$m -> tf_6$n,
    NULL ->> tf_5$k,
    NULL ->> tf_6$o
  )
dm_draw(dm_for_filter_quick)

# building prototypes for `dm_unnest_tbl`
airlines_wrapped <-
  dm_nycflights13() \%>\%
    dm_wrap_tbl(airlines)

airlines_wrapped \%>\%
  dm_unnest_tbl(airlines, flights, quick_dm(flights$carrier ->> airlines))
}
