% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{This reads in the csv file saved from an A & P file and creates dat and input}
\usage{
readData(a.and.p.file, input, folder = "./", silent = FALSE)
}
\arguments{
\item{a.and.p.file}{a csv file saved from the DM tab of an A & P excel file}

\item{input}{a list with the other values needed for a DM run. The following are the defaults used, but can be passed in to specify something different: naturalMort = c(0.5,0.4,0.3,0.2,0.1,0), analysisType = "DM", SRfunction = "ricker", includeMarineSurvival = "no", includeFlow = "no"}

\item{silent}{Whether to print progress messages.}
}
\value{
The result is a list with dat and input for the DM functions
}
\description{
The DM tab of the A & P excel file has all the 
necessary inputs to run DM.  This reads that .csv file and makes
a list needed for the DM functions.  creates input which has the 
other user specified values for the rav file.
}
