% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Compact and literal vectorized switch().}
\usage{
recode(x, initial, new)
recode_all <- function(x, ...)
NAvl(x, v)
NAvl0(x)
switchv(x, ...)
swap(x, ...)
switchv_all(x, ...)
}
\arguments{
\item{x}{A vector.}

\item{initial}{Initial set of values--i.e., values to be replaced.}

\item{new}{New set of values--i.e., values to replace the initial ones.}

\item{v}{Value to replace a missing value.}

\item{...}{Parameters passed to \code{recode} in the case of \code{recode_all} or parameters passed to \code{switch} in the case of \code{switchv}, \code{swap}, and \code{switchv_all}}
}
\value{
A vector
}
\description{
Recode variables in a simple manner OR recode variables a la SPSS style.
}
\details{
\code{recode} replaces an initial set of values with a new set of values for a vector.

\code{recode_all} applies \code{recode} to all columns in a dataset.

\code{NAvl} replaces a vector's missing values with a specified value set in \code{v}.

\code{NAvl0} replaces a vector's missing values with 0.

\code{switchv} behaves the same as \code{switch} except that it applies to a vector (i.e., it is a vectorized version of the latter). \code{swap} is its synonym.

\code{switchv_all} applies \code{switchv} to all columns in a data frame.
}
\examples{
recode(mtcars$am, 0:1, 2:3)

with(airquality, NAvl(Ozone, mean(Ozone, na.rm = TRUE)))

switchv(iris$Species, setosa = 1, versicolor = 2, virginica = 3)

}
\seealso{
\url{https://github.com/robertschnitman/dm}
}
