% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_file.R
\name{rename_file}
\alias{rename_file}
\title{Rename files}
\usage{
rename_file(filepath, pattern_now, pattern_new)
}
\arguments{
\item{filepath}{Desired directory of where the files to be renamed are.}

\item{pattern_now}{Regular expression pattern of the file names as they are CURRENTLY.}

\item{pattern_new}{Desired string to replace the file names detected in \code{pattern_now}.}
}
\description{
Rename files in your directory.
}
\details{
Renames files in a specified directory based on a given pattern.
}
\seealso{
\url{https://github.com/robertschnitman/dm}
}
