% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report1.R
\name{Report1}
\alias{Report1}
\title{Creates a PDF report with plots.}
\usage{
Report1(dmObj = NULL, dmObj.RData.file = NULL, output.file = "report1",
  rav.options = list(), output_format = c("html_document", "pdf_document"))
}
\arguments{
\item{dmObj}{a saved DM object (list) from runModel() with the objects "input","dat","result", and "tDat"}

\item{dmObj.RData.file}{a saved DM object as a RData file with the objects "input","dat","result", and "tDat"}

\item{output.file}{Filename to give saved report.}

\item{rav.options}{list of the rav file options to use that do not come from the posteriors.}

\item{output_format}{list of output formats.  Any that rmarkdown:::render allow are fine.}
}
\value{
Nothing.  The report is written to a PDF and tex file.
}
