% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows-truncate.R
\name{rows_truncate}
\alias{rows_truncate}
\title{Truncate all rows}
\usage{
rows_truncate(x, ..., copy = FALSE, in_place = FALSE)
}
\arguments{
\item{x}{A data frame or data frame extension (e.g. a tibble).}

\item{...}{other parameters passed onto methods, for instance, \code{na_matches}
to control how \code{NA} values are matched.  See \link[dplyr]{join.tbl_df} for more.}

\item{copy}{If \code{x} and \code{y} are not from the same data source,
and \code{copy} is \code{TRUE}, then \code{y} will be copied into the
same src as \code{x}.  This allows you to join tables across srcs, but
it is a potentially expensive operation so you must opt into it.}

\item{in_place}{Should \code{x} be modified in place? This argument is only
relevant for mutable backends (e.g. databases, data.tables).

When \code{TRUE}, a modified version of \code{x} is returned invisibly;
when \code{FALSE}, a new object representing the resulting changes is returned.}
}
\description{
\code{rows_truncate()} removes all rows.
This operation corresponds to \code{TRUNCATE} in SQL.
\code{...} is ignored.
}
