% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runJAGS.R
\name{runJAGS}
\alias{runJAGS}
\title{Call JAGS to get the posterior draws.}
\usage{
runJAGS(bdat, calcInits, sims = 10000, numChains = 3, details = FALSE)
}
\arguments{
\item{bdat}{a list of containing the data necessary for the JAGS model}

\item{calcInits}{a list or function that generates a list of intial values for the JAGS model}

\item{sims}{the number of MCMC simulations used for each chain after a burnin of 10 percent of sims}

\item{numChains}{number of MCMC chains to use}

\item{details}{if details=TRUE, more detailed results are returned.}
}
\value{
If details=FALSE, a list of MCMC chains for the monitored model parameters is returned. If details=TRUE, a model object generated by the jags function is returned (this contains the simulations along with additional information)
}
\description{
This function calls jags to get the posterior draws.
}
