% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{Run the DM model}
\usage{
runModel(filename, input = NULL, priors = NULL, run = TRUE,
  sims = 10000, numChains = 3, oldSims = NULL, silent = FALSE)
}
\arguments{
\item{filename}{csv file saved from the DM tab of A & P file.  Must have age data.}

\item{input}{A list that includes parameters for running the model (see readData for details) of particular interest are the parameters SRfunction, analysisType, and escapementObsSD. SRfunction sets the spawner-recruit function to be used. Current options are "ricker", "bevertonHolt" or "hockeyStick".   analysisType can be either "DM" or "SS".  DM corresponds to the traditional Dynamic Model which uses the maturation rates from the DynamicsInput tab in the A&P table. SS corresponds to the State Space model which estimates the maturation rates and includes observation error in the spawner counts.  If escapementObsSD is a positive real number, observation error is fixed at that value. If it is NULL, observation error is a free parameter that is estimated.}

\item{priors}{parameters for prior distributions in the JAGS model}

\item{run}{if run=TRUE, the model is run and the results returned, otherwise all of the information necessary to run the model is returned without running the model}

\item{sims}{length of each MCMC chain; default is 10000}

\item{numChains}{number of MCMC chains. default is 3}

\item{oldSims}{?}

\item{silent}{Whether to print progress messages.}
}
\value{
a dmObj: a list of output for a single model fit.
}
\description{
This function takes the list created by readDMData() and runs
the Bayesian model to estimate the SR parameters.
The output is written to a .csv file and an .Rdata file.
It writes a file, mod1.txt to the user's working directory
}
