% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-helpers.R
\name{schema_supported_dbs}
\alias{schema_supported_dbs}
\title{Databases for which primary and foreign keys guessing from schema is supported}
\usage{
schema_supported_dbs()
}
\description{
Databases for which primary and foreign keys guessing from schema is supported
}
\examples{
schema_supported_dbs()
}
