% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRFunctions.R
\name{selectSR}
\alias{selectSR}
\title{Helper function that takes the name of SR function and returns the recruits}
\usage{
selectSR(x)
}
\arguments{
\item{x}{name of SR function as a text string: "ricker", "bevertonHolt", "hockeyStick"}
}
\value{
a function
}
