% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{sql_schema_drop}
\alias{sql_schema_drop}
\title{Remove a schema from a database}
\usage{
sql_schema_drop(dest, schema, ...)
}
\arguments{
\item{dest}{An object of class \code{"src"} or \code{"DBIConnection"}.}

\item{schema}{Class \code{character} or \code{SQL} (cf. Details), name of the schema}

\item{...}{Passed on to the individual methods.}
}
\value{
\code{NULL} invisibly.
}
\description{
\code{sql_schema_drop()} deletes an empty schema from the database.
}
\details{
Methods are not available for all DBMS.

An error is thrown if no schema of that name exists.

The argument \code{schema} (and \code{dbname} for MSSQL) can be provided as \code{SQL} objects.
Keep in mind, that in this case it is assumed that they are already correctly quoted as identifiers.

Additional arguments are:
\itemize{
\item \code{dbname}: supported for MSSQL. Remove a schema from a different
database on the connected MSSQL-server; default: database addressed by \code{dest}.
}
}
\seealso{
Other schema handling functions: 
\code{\link{sql_schema_create}()},
\code{\link{sql_schema_exists}()},
\code{\link{sql_schema_list}()},
\code{\link{sql_schema_table_list}()}
}
\concept{schema handling functions}
