% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{sql_schema_table_list}
\alias{sql_schema_table_list}
\title{List the tables in a schema on a database}
\usage{
sql_schema_table_list(dest, schema = NULL, ...)
}
\arguments{
\item{dest}{An object of class \code{"src"} or \code{"DBIConnection"}.}

\item{schema}{Name of the schema}

\item{...}{Passed on to the individual methods.}
}
\value{
A tibble with the following columns:
\describe{
\item{\code{table_name}}{name of the table,}
\item{\code{remote_name}}{identifier of the table on the DBMS.
Can be used to access the listed tables with the syntax
\verb{tbl(dest, remote_name).}}
}
}
\description{
\code{sql_schema_table_list()} list the tables in a schema on the database.
}
\details{
Methods are not available for all DBMS.

An error is thrown if no schema of that name exists.

Additional arguments are:
\itemize{
\item \code{dbname}: supported for MSSQL. Look for tables on a different
database on the connected MSSQL-server; default: database addressed by \code{dest}.
}
}
\seealso{
Other schema handling functions: 
\code{\link{sql_schema_create}()},
\code{\link{sql_schema_drop}()},
\code{\link{sql_schema_exists}()},
\code{\link{sql_schema_list}()}
}
\concept{schema handling functions}
