% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRFunctions.R
\name{srfunctions}
\alias{srfunctions}
\alias{ricker}
\alias{ricker}
\alias{bevertonHolt}
\alias{hockeyStick}
\title{SR funcitons}
\usage{
ricker(S, covariates, p)

bevertonHolt(S, covariates, p)

hockeyStick(S, covariates, p)
}
\arguments{
\item{S}{spawner count}

\item{covariates}{A list with the covariates.  The ms covariate is normally the first and the flow covariate is the second.  Though order does not matter as the p's are matched based on order the covariate appears in the list.}

\item{p}{the SR function a, b, and the coefficients for each covariate in the covariate list. All covariates appear in the form exp(exp(p)*cov). p[1]=a; p[2]=b.  The other p match the covariates.}
}
\value{
recruits (scalar)

recruits from ricker: S*exp(p[1])*exp(-S/exp(p[2]))*exp(p[4]*flow)*exp(p[3]*marineInd)

recruits from beverton Holt: S/(S*exp(-p[2])+exp(-p[1]))*exp(p[4]*flow)*exp(p[3]*marineInd)

recruits from hockey stick: ifelse(exp(p[1])*S<exp(p[2]),exp(p[1])*S,exp(p[2]))*exp(p[4]*flow)*exp(p[3]*marineInd)
}
\description{
Takes the S, LOG flow, LOG marine survival covariates and p parameters and returns the recruits based on the specified SR function: ricker(), bevertonHolt(), hockeyStick().  Note that it assumes that the log of the covariates are input. All covariates appear in the form exp(p*cov). p[1]=log a; p[2]=log b.  The other p are the covariate coefficients.
}
