% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSTfuncDM.R
\name{SSTfuncDM}
\alias{SSTfuncDM}
\title{Original dynamic model}
\usage{
SSTfuncDM(p, SRfunc, dat, input, returnAll = FALSE)
}
\arguments{
\item{p}{The values of the SR function parameters logged.  p[1]=log(prod), p[2]=log(cap).  The other p depend on what covariates are in the model. The length of p is 2 + number of covariates.  If there are 2 covariates, then p[3]= log(msCoef) and p[4]=log(flowCoef).}

\item{SRfunc}{a SR function: ricker, bevertonHolt, hockeyStick, as returned by selectSR()}

\item{dat}{data from the A & P file}

\item{input}{a list with the other values needed for a DM run. 
The following are examples naturalMort, analysisType = "DM", SRfunction, covariates, includeMarineSurvival, includeFlow}

\item{returnAll}{(TRUE/FALSE)}
}
\value{
Returns the sum of squares residuals between the observed and estimated escapement. If returnAll is true, the estimated and observed escapement are also returned.
}
\description{
Implements the DM model from the original RER process. Estimates escapement by year using escapement data from previous years, maturation rates, harvest rates, and rates of natural mortality. The estimated escapement is log transformed and subtracted from the log transformed values of observed escapement. These differences are then squared and summed to get the sum of squares residual which is returned by the function. This function can be used to estimated the spawner-recruit parameters through the use of an optimization function.
}
