% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopif.R
\name{stopif}
\alias{stopif}
\title{Conditional halt of an execution}
\usage{
stopif(cond, msg = 'Stop condition met.', ...)
}
\arguments{
\item{cond}{Condition}

\item{msg}{The printed error message.}

\item{...}{Parameters passed to \code{stop}}
}
\description{
Stop an execution based on a specified condition.
}
\seealso{
\url{https://github.com/robertschnitman/dm}
}
