% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_subcondition.R
\name{subset_subcondition}
\alias{subset_subcondition}
\title{subset by subcondition}
\usage{
subset_subcondition(x, meta.var = NULL, ..., select = NULL)
}
\arguments{
\item{meta.var}{the meta variable}

\item{...}{condition for subsetting passed to \code{subset}}

\item{data}{the data, a data frame or some such thing}
}
\value{
a data frame
}
\description{
this function subsets by some condition, but returns all information
    associated with those values of the meta variable 'meta.var' that would
    be in the standard subset returned by the same condition.
}
\examples{
d <- data.frame(id = rep(LETTERS[1:2], each = 2), z = 1:4)
## 'subset' returns the rows satisfying the condition
subset(d, z == 4)
## 'subset_subcondition' returns all rows such that the 'meta.var' value is
## in the standard subset returned, i.e. a (possibly) bigger set
subset_subcondition(d, meta.var = "id", z == 4)
}
