% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeRuns.R
\name{summarizeRuns}
\alias{summarizeRuns}
\title{Summarize runs}
\usage{
summarizeRuns(runList)
}
\arguments{
\item{runList}{a list of multiple run objects (dmObj).  Each run object in the list is from one call of runModel}
}
\value{
a data frame with summary statistics for each run
}
\description{
Summarize DM runs
}
