% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_combine.R
\name{surv_combiner}
\alias{surv_combiner}
\title{Surv combine}
\usage{
surv_combiner(..., comb.type = "EELC", surv.output = TRUE)
}
\arguments{
\item{...}{Surv objects}

\item{comb.type}{how the Surv  objects are combined. 'EELC'  (Early Events,
Late Censored) thinks of the Surv  objects as informative of each other,
e.g. if the first observation in the  first row is censored at 1 for the
first variable  and at 2 for  the second variable, then  the combination
will pick the  second one. More specifically, it will  pick the earliest
event (as an event) if there is at least one event, and it will pick the
latest censored (as censored) if  there are no events. 'earliest' thinks
of the Surv objects  as being very distinct, so it  will simply pick the
earliest time (as event or censored).}

\item{surv.output}{return a Surv object (default \code{TRUE}),  else a data
frame with variables \code{time} and \code{event}.}
}
\description{
combine Surv variables
}
\details{
combine two or more Surv objects into a combined Surv object
}
