% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censored-formats.R
\name{surv_converter}
\alias{surv_converter}
\alias{surv_pair2Surv}
\alias{Surv2surv_pair}
\title{convert censored times}
\usage{
surv_converter(
  data,
  id = NULL,
  toSurv = NULL,
  .fix = c(time = "t.", event = "ev."),
  suffix = TRUE
)

surv_pair2Surv(
  data,
  id = NULL,
  .fix = c(time = "t.", event = "ev."),
  suffix = TRUE
)

Surv2surv_pair(
  data,
  id = NULL,
  .fix = c(time = "t.", event = "ev."),
  suffix = TRUE
)
}
\arguments{
\item{data}{the data, a data frame or some such object}

\item{id}{is there an identifiction variable that should remain in the data set}

\item{toSurv}{logical; resulting variables to be of class 'Surv'?}

\item{.fix}{named vector of pre- or suffix that identifies the 'time' and
'event'-variables (names must contain "time" and "event")}

\item{suffix}{is .fix a suffix?}
}
\value{
a data frame
}
\description{
censored times might already be of the 'Surv' class, or where the 'time' and
    'event' components of such a class is stored separately, either way this
    is a function to transition between these two formats
}
\section{Functions}{
\itemize{
\item \code{surv_pair2Surv}: transition from separate variables to class 'Surv'

\item \code{Surv2surv_pair}: transition from class 'Surv' to separate variables
}}

\examples{
df <- data.frame(
       persnr = sprintf("id\%s", 1:10),
       t.arg = rexp(10, 1/50),
       ev.arg = rbinom(10, 1, 0.7)
)
surv_converter(df, id = "persnr")
}
