% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_truncate.R
\name{surv_truncate}
\alias{surv_truncate}
\title{Surv truncate}
\usage{
surv_truncate(x, ttime = NULL, surv.output = TRUE)
}
\arguments{
\item{x}{Surv object}

\item{ttime}{truncating/censoring time}

\item{surv.output}{return a Surv object (default \code{TRUE}),  else a data
frame with variables \code{time} and \code{event}.}
}
\description{
truncate Surv object
}
\details{
censor at 'ttime' for all later times
}
