% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit_it.R
\name{survfit_it}
\alias{survfit_it}
\title{multiple survival curves}
\usage{
survfit_it(
  surv,
  data,
  strata = NULL,
  glist = NULL,
  w = NULL,
  type = "kaplan-meier",
  prefix = c(event = "ev.", time = "t."),
  vs_age = FALSE,
  age_var = NULL,
  age_bound = NULL,
  progress = FALSE,
  stringsAsFactors = TRUE
)
}
\arguments{
\item{surv}{the \code{Surv} objects (as character, refering to variables in
'data') OR the common names of time + event, if these have similar names
(differing only in perfix)}

\item{data}{the data}

\item{strata}{stratifying variable}

\item{glist}{a grouping list (list of logical vectors of the same length as
rows in data, i.e. can be overlapping groups), or the name of a grouping
variable in 'data'}

\item{w}{weight or name of variable that provides weights}

\item{type}{passed to \code{survival::survfit}}

\item{prefix}{prefix values for times and events}

\item{vs_age}{logical; should KM curves be with age on the x-axis?}

\item{age_var}{character; name of age variable. N.B. important that age is
measured in the same units as the time variable}

\item{age_bound}{numeric; what age to start from}

\item{progress}{primitive displayer of progress}

\item{stringsAsFactors}{logical, if TRUE will keep ordering of ingoing stuff}
}
\value{
a data frame (much like a \code{broom::tidy}-version of the output)
}
\description{
wrapper for survfit to create multiple survival curves for
    multiple endpoints, stratifications and groupings
}
\examples{
n = 1200
df <- data.frame(
        t.foo = abs(rnorm(n, 10, 2)),
        ev.foo = rbinom(n, 1, 0.2),
        t.bar = abs(rnorm(n, 10, 3)),
        ev.bar = rbinom(n, 1, 0.3),
        strutt = sample(letters[1:2], n, TRUE),
        gsak = sample(LETTERS[4:7], n, TRUE),
        wait = .5 + abs(rnorm(n, 0, 1))
)
str(survfit_it(data = df, surv = c("foo", "bar"), w = "wait",
          strata = "strutt", glist = "gsak"))
}
