% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdc_functions.R
\name{tdc_in_study}
\alias{tdc_in_study}
\title{participants over time}
\usage{
tdc_in_study(x, tstart = "tstart", tstop = "tstop")
}
\arguments{
\item{x}{data frame with time periods starting at tstart, ending at tstop.}

\item{tstart}{name of variable acting as tstart}

\item{tstop}{name of variable acting as tstop}
}
\value{
data frame with variables 'time' and 'in.study' (count)
}
\description{
\code{survival::tmerge} creates time periods (per patient) starting at
'tstart' and ending at 'tstop'. This function counts the number of
individuals in the study at each time point of interest (set of all distinct
time points starting or ending an interval). For each such point, a patient
is in the study at T if they have a time period starting at, or before, T,
AND the corresponding end strictly after T.
}
