% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdc_functions.R
\name{tdc_stat}
\alias{tdc_stat}
\title{stats for tdc}
\usage{
tdc_stat(x, var, stat, stat.names = NULL, tstart = "tstart", tstop = "tstop")
}
\arguments{
\item{x}{data frame with time periods starting at tstart, ending at tstop.}

\item{var}{name of variable of interest}

\item{stat}{function to calculate stat. Note: can return a vector}

\item{stat.names}{names of columns created, if set to NULL, the function
will provide them}

\item{tstart}{name of variable acting as tstart}

\item{tstop}{name of variable acting as tstop}
}
\value{
a data frame
}
\description{
Attempt to describe a time dependent covariate (tdc).
\code{survival::tmerge} creates time periods (per patient) starting at
'tstart' and ending at 'tstop'. This function calculates a statistic at each
time point of interest (set of all distinct time points starting or ending
an interval). For each such point, the values of a tdc is gathered to
calculate a statistic (determined by argument 'stat').
}
