% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_to_tibble.R
\name{tibble_to_dm}
\alias{tibble_to_dm}
\title{Convert a wrapped tibble to a dm}
\usage{
tibble_to_dm(x, prototype, root = NULL)
}
\arguments{
\item{x}{A wrapped table, as created by \code{dm_to_tibble()}}

\item{prototype}{A dm, might be an empty prototype.}

\item{root}{The root table (unquoted), optional because we can often
infer it from \code{x} and \code{prototype}}
}
\value{
A dm
}
\description{
Convert a wrapped tibble to a dm
}
\examples{
# often we can infer the root table from the prototype
flights_wrapped <- dm_to_tibble(dm_nycflights13(), flights, silent = TRUE)
tibble_to_dm(flights_wrapped, dm_nycflights13())

# other times, it is ambiguous and should be given
airlines_wrapped <- dm_to_tibble(dm_nycflights13(), airlines, silent = TRUE)
tibble_to_dm(airlines_wrapped, dm_nycflights13(), airlines)
}
\seealso{
\link{tibble_to_dm}, \link{dm_wrap},
\link{dm_wrap}, \link{dm_unwrap},
}
