% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_class.R
\name{type}
\alias{type}
\title{Shortcut functions for knowing types and classes of objects}
\usage{
type(x)
type_class(x)
tc(x)
}
\arguments{
\item{x}{An object.}
}
\description{
Shortcut functions for knowing types and classes of objects
}
\details{
\code{type} provides the type of an object. \code{type_class} provides the type and class of an object--\code{tc} is a synonym.
}
\examples{
type(mtcars)
type_class(mtcars)
tc(mtcars)

}
\seealso{
\url{https://github.com/robertschnitman/dm}
}
