% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncenterProdCap.R
\name{uncenterProdCap}
\alias{uncenterProdCap}
\title{Uncenter productivty and log capacity parameters}
\usage{
uncenterProdCap(input, dat, params)
}
\arguments{
\item{dat}{data from the A & P file}

\item{params}{a vector of parameters with names flowCoef, msCoef, prod, cap.  Assumed to be from the runModel() Bayesian analysis or findOptimum() call, so centered (or not) based on input$centerFlow and input$centerMS}
}
\value{
a vector of uncentered parameters with names flowCoef, msCoef, prod, cap.
}
\description{
Transform capacity and productivity estimates to what they would be if the covariates were centered during estimation (Bayesian or ML).  VRAP uses the uncentered parameters. Note, if the coefficients were centered during estimation, input$centerMS=TRUE and/or input$centerFlow=TRUE.  These values are used to determine whether to apply the correction.  Thus input is assumed to match the parameter values in params.
}
\seealso{
centerProdCap
}
