% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconstruct.R
\name{unclass_keyed_tbl}
\alias{unclass_keyed_tbl}
\title{Remove \code{"dm_keyed_tbl"} class}
\usage{
unclass_keyed_tbl(tbl)
}
\value{
If entered table has \code{"dm_keyed_tbl"} class, it will be removed. All
other classes will be preserved.
}
\description{
Remove \code{"dm_keyed_tbl"} class
}
\examples{
dm <- dm_nycflights13()
class(dm$airlines)
class(dm:::unclass_keyed_tbl(dm$airlines))
}
\keyword{internal}
