% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_spin.R
\name{validate_spin}
\alias{validate_spin}
\title{validate SPIN}
\usage{
validate_spin(spin)
}
\arguments{
\item{spin}{a swedish person identity number of the form "XXXXXX-XXXX" or
"XXXXXXXX-XXXX"}
}
\value{
logical
}
\description{
test if a Swedish Personal Identity Number is plausible by checking the
    control number
}
