% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeReport.R
\name{writeReport}
\alias{writeReport}
\title{Creates a report with plots.}
\usage{
writeReport(dmObj = NULL, dmObj.RData.file = NULL,
  output.file = "report1", rav.options = list(),
  output_format = c("pdf_document"), output.dir = getwd(),
  input.type = "xRnw")
}
\arguments{
\item{dmObj}{a saved DM object (list) from runModel() with the objects "input","dat","result", and "tDat"}

\item{dmObj.RData.file}{a saved DM object as a RData file with the objects "input","dat","result", and "tDat"}

\item{output.file}{Filename to give saved report.}

\item{rav.options}{list of the rav file options to use that do not come from the posteriors.}

\item{output_format}{list of output formats.  If input.type=="xRmd", any format that rmarkdown:::render allows is fine.  If input.type=="xRnw", only "pdf_document" is allowed.}

\item{output.dir}{Directory where to save the output files.  Defaults to working directory.}

\item{input.type}{xRmd for rmarkdown.  xRnw for knitr.}
}
\value{
Nothing.  The report is written to a PDF and tex file.
}
