% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeResultsToFile.R
\name{writeResultsToFile}
\alias{writeResultsToFile}
\title{Save DM output.}
\usage{
writeResultsToFile(input, dat, tDat, mlEst, bdat, filename)
}
\arguments{
\item{input}{a list with the other values needed for a DM run. The following are the defaults used, but can be passed in to specify something different: naturalMort = c(0.5,0.4,0.3,0.2,0.1,0), analysisType = "DM", SRfunction = "ricker", covariates = "no", includeMarineSurvival = "no", includeFlow = "no"}

\item{dat}{data from the A & P file}

\item{tDat}{a data frame of the posteriors and calculated process error and Smsy for each draw.  From calculatePEandSmsy().}

\item{mlEst}{The maximum likelihood estimate for the four SR parameters}

\item{bdat}{data for the bayesian specification of a DM run}

\item{filename}{name to give the outputfiles file}

\item{population}{Name of the population.}
}
\value{
nothing is returned but the csv and RData files are written.
}
\description{
Write DM output (posteriors) to csv file and RData file.
}
