% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_flag.R
\name{zero_flag}
\alias{zero_flag}
\title{Flag a zero at the start of a value}
\usage{
zero_flag(x, w, format = 'd')
}
\arguments{
\item{x}{A vector.}

\item{w}{Width.}

\item{format}{Based on \code{format} in \code{formatC}.}
}
\description{
Flag a zero at the start of a value. Useful for formatting numbers with a specified number of digits such as Social Security Numbers.
}
\seealso{
\url{https://github.com/robertschnitman/dm}
}
