% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdates.R
\name{cdate}
\alias{cdate}
\title{fix censored dates}
\usage{
cdate(
  x,
  sep = NULL,
  low.bound = NULL,
  verbose = TRUE,
  bound4all = FALSE,
  ignore.na = TRUE
)
}
\arguments{
\item{x}{dates, possibly censored, as character or numeric (e.g. 20070101)}

\item{sep}{separator in x, if any}

\item{low.bound}{the lower bound}

\item{verbose}{logical; print additional info, maybe}

\item{bound4all}{logical (default \code{FALSE}); apply bound also to
uncensored dates?}

\item{ignore.na}{logical; ignore missing values in x?}
}
\description{
fix censored dates of the type '20110000' or '20110200' by
    selecting the midpoint of the censored time interval. If a lower bound
    is given, the midpoint between this point and the end of the censored
    interval is chosen.
}
