% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{cdm_add_fk}
\alias{cdm_add_fk}
\title{Add a reference from one table of a \code{\link{dm}} to another}
\usage{
cdm_add_fk(dm, table, column, ref_table, check = TRUE)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}

\item{column}{The column of \code{table} which is to become the foreign key column
referencing the primary key of \code{ref_table}.}

\item{ref_table}{The table which \code{table} is referencing. This table needs to have
a primary key set.}

\item{check}{Boolean, if \code{TRUE} (default), a check is performed, if the values of
\code{column} are a subset of the values of the primary key column of \code{ref_table}.}
}
\description{
Add a reference from one table of a \code{\link{dm}} to another
}
\seealso{
Other Foreign key functions: \code{\link{cdm_enum_fk_candidates}},
  \code{\link{cdm_get_all_fks}}, \code{\link{cdm_get_fk}},
  \code{\link{cdm_has_fk}}, \code{\link{cdm_rm_fk}}
}
\concept{Foreign key functions}
