% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-keys.R
\name{cdm_check_for_pk_candidates}
\alias{cdm_check_for_pk_candidates}
\title{Which columns are candidates for a primary key column of a \code{dm}-object's table?}
\usage{
cdm_check_for_pk_candidates(dm, table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}
}
\description{
\code{cdm_check_for_pk_candidates()} checks for each column of a
table of a \code{dm}-object if this column contains only unique values and is therefore
a unique key of this table.
}
\examples{
library(dplyr)
nycflights_dm <- cdm_nycflights13()

nycflights_dm \%>\% cdm_check_for_pk_candidates(flights)
nycflights_dm \%>\% cdm_check_for_pk_candidates(airports)

}
