% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-keys.R
\name{cdm_check_for_pk_candidates}
\alias{cdm_check_for_pk_candidates}
\title{Which columns are candidates for a primary key column of a \code{\link{dm}} object's table?}
\usage{
cdm_check_for_pk_candidates(dm, table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}
}
\description{
\code{cdm_check_for_pk_candidates()} checks for each column of a
table of a \code{\link{dm}} object if this column contains only unique values and is therefore
a unique key of this table.
}
\examples{
library(dplyr)
nycflights_dm <- cdm_nycflights13()

nycflights_dm \%>\% cdm_check_for_pk_candidates(flights)
nycflights_dm \%>\% cdm_check_for_pk_candidates(airports)
}
\seealso{
Other Primary key functions: \code{\link{cdm_add_pk}},
  \code{\link{cdm_get_all_pks}}, \code{\link{cdm_get_pk}},
  \code{\link{cdm_has_pk}}, \code{\link{cdm_rm_pk}}
}
\concept{Primary key functions}
