% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disentangle.R
\name{cdm_disentangle}
\alias{cdm_disentangle}
\title{Disentangle relationships}
\usage{
cdm_disentangle(dm, child_table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{child_table}{Table from which multiple foreign keys are pointing to another table}
}
\description{
A normalized data model does not contain multiple key relations pointing from one table to another.
If this situation should occur anyway, this function can help to untie this knot.
}
\details{
Sometimes a table points with more than one column to another table.
Usually these two columns have a different meaning, but refer to the same primary key values
of the parent table, as e.g. in \code{cdm_nycflights13(cycle = TRUE)}.
\code{cdm_disentangle()} will insert as many copies of the old parent table(s), as there are duplicate key relations between
the two tables, replace the old parent table(s) with them and install new foreign key relations, thus
removing the cycles in the relationship graph that were introduced in this way.
}
\examples{
cdm_disentangle(cdm_nycflights13(cycle = TRUE), flights)
}
