% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-dm.R
\name{cdm_draw}
\alias{cdm_draw}
\alias{cdm_set_colors}
\alias{cdm_get_colors}
\alias{cdm_get_available_colors}
\title{Draw schema of a \code{\link{dm}}-object's data model}
\usage{
cdm_draw(dm, table_names = NULL, rankdir = "LR", col_attr = "column",
  view_type = "keys_only", columnArrows = TRUE, graph_attrs = "",
  node_attrs = "", edge_attrs = "", focus = NULL,
  graph_name = "Data Model")

cdm_set_colors(dm, ...)

cdm_get_colors(dm)

cdm_get_available_colors()
}
\arguments{
\item{dm}{A \code{\link{dm}} object}

\item{table_names}{Tables to consider in the model visualisation.
If \code{NULL} (default), there is no constraint.}

\item{rankdir}{Graph attribute for direction (eg. 'BT' = bottom --> top)}

\item{col_attr}{Column atributes to display.
Only column name (\code{column}) is included by default.}

\item{view_type}{Can be "all" (by default), "keys_only" or "title_only".
It defines the level of details for the table rendering
(all columns, only primary and foreign keys or no columns)}

\item{columnArrows}{Edges from column to column (default: FALSE)}

\item{graph_attrs}{Additional graph attributes}

\item{node_attrs}{Additional node attributes}

\item{edge_attrs}{Additional edge attributes}

\item{focus}{A list of parameters for rendering (table filter)}

\item{graph_name}{A graph name}

\item{...}{Colors to set in the form \code{table = "<color>"} . Fall-through syntax similarly to
\code{\link[=switch]{switch()}} is supported: \code{table1 = , table2 = "<color>"} sets the color for both \code{table1}
and \code{table2} . This argument supports splicing.}
}
\value{
For \code{cdm_set_colors()}: the updated data model.

For \code{cdm_get_colors()}, a two-column tibble with one row per table.

For \code{cdm_get_available_colors()}, a tibble with the color in the first
column and auxiliary information in other columns.
}
\description{
\code{cdm_draw()} draws a schema of the data model using \code{datamodelr} (which in turn uses \code{DiagrammeR})

\code{cdm_set_colors()} allows to define the colors in which to display the tables of the data model.

\code{cdm_get_colors()} returns the colors defined for a data model.

\code{cdm_get_available_colors()} returns an overview of the available colors and their names
as a tibble.
}
\examples{
library(dplyr)
cdm_draw(cdm_nycflights13())
cdm_draw(cdm_nycflights13(cycle = TRUE))

cdm_nycflights13(color = FALSE) \%>\%
  cdm_set_colors(
    airports = ,
    airlines = ,
    planes = "yellow",
    weather = "dark_blue") \%>\%
    cdm_draw()

# Splicing is supported:
new_colors <- c(
  airports = "yellow", airlines = "yellow", planes = "yellow",
  weather = "dark_blue"
)
cdm_nycflights13(color = FALSE) \%>\%
  cdm_set_colors(!!!new_colors) \%>\%
  cdm_draw()
}
