% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{cdm_enum_fk_candidates}
\alias{cdm_enum_fk_candidates}
\title{Find foreign key candidates in a table}
\usage{
cdm_enum_fk_candidates(dm, table, ref_table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{The table whose columns should be tested for foreign key candidate potential}

\item{ref_table}{A table with a primary key.}
}
\value{
A table with an overview which columns of \code{table} would be suitable candidates as
foreign key columns referencing \code{ref_table}
}
\description{
Which columns are candidates for a foreign key of a table, referencing
the primary key column of another \code{\link{dm}} object's table?
}
\details{
\code{cdm_enum_fk_candidates()} checks first, if \code{ref_table} has a primary key set. Then
it determines for each column of \code{table}, if it has the same class as the primary key of \code{ref_table}.
If this is the case a check is performed, if the column contains only a subset of values of the
primary key column of \code{ref_table}. If this is \code{TRUE}, this column is a candidate for a foreign key
from \code{table} to \code{ref_table}.
}
\examples{
cdm_enum_fk_candidates(cdm_nycflights13(), flights, airports)

}
\seealso{
Other foreign key functions: \code{\link{cdm_add_fk}},
  \code{\link{cdm_get_all_fks}}, \code{\link{cdm_get_fk}},
  \code{\link{cdm_has_fk}}, \code{\link{cdm_rm_fk}}
}
\concept{foreign key functions}
