% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{cdm_enum_fk_candidates}
\alias{cdm_enum_fk_candidates}
\alias{enum_fk_candidates}
\title{Find foreign key candidates in a table}
\usage{
cdm_enum_fk_candidates(dm, table, ref_table)

enum_fk_candidates(zoomed_dm, ref_table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{The table whose columns should be tested for suitability as foreign keys.}

\item{ref_table}{A table with a primary key.}

\item{zoomed_dm}{A \code{dm} with a zoomed table.}
}
\value{
A table that lists which columns of \code{table} would be suitable candidates for
foreign key columns to reference \code{ref_table}, which columns would not be suitable,
and the reason \code{why}.
}
\description{
Determine which columns would be good candidates to be used as foreign keys of a table,
to reference the primary key column of another table of the \code{\link{dm}} object.
}
\details{
\code{cdm_enum_fk_candidates()} first checks if \code{ref_table} has a primary key set,
if not, an error is thrown.

If \code{ref_table} does have a primary key, then a join operation will be tried using
that key as the \code{by} argument of join() to match it to each column of \code{table}.
Attempting to join incompatible columns triggers an error.

The outcome of the join operation determines the value of the \code{why} column in the result:
\itemize{
\item an empty value for a column of \code{table} that is a suitable foreign key candidate
\item the count and percentage of missing matches for a column that is not suitable
\item the error message triggered for unsuitable candidates that may include the types of mismatched columns
}

\code{enum_fk_candidates()} works like \code{cdm_enum_fk_candidates()} with the zoomed table as \code{table}.
}
\examples{
cdm_enum_fk_candidates(cdm_nycflights13(), flights, airports)

cdm_zoom_to_tbl(cdm_nycflights13(), flights) \%>\%
  enum_fk_candidates(airports)

}
\seealso{
Other foreign key functions: 
\code{\link{cdm_add_fk}()},
\code{\link{cdm_get_all_fks}()},
\code{\link{cdm_get_fk}()},
\code{\link{cdm_has_fk}()},
\code{\link{cdm_rm_fk}()}
}
\concept{foreign key functions}
