% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{cdm_enum_fk_candidates}
\alias{cdm_enum_fk_candidates}
\title{Find foreign key candidates in a table}
\usage{
cdm_enum_fk_candidates(dm, table, ref_table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{The table whose columns should be tested for foreign key candidate potential}

\item{ref_table}{A table with a primary key.}
}
\value{
A table with an overview which columns of \code{table} would be suitable candidates as
foreign key columns referencing \code{ref_table}
}
\description{
Which columns are good candidates as a foreign key of a table, referencing
the primary key column of another \code{\link{dm}} object's table?
}
\details{
\code{cdm_enum_fk_candidates()} checks first, if \code{ref_table} has a primary key set.
For each column of \code{table} an \code{\link[=anti_join]{anti_join()}} is then tried, with parameter \code{by} matching
the respective column with the primary key of \code{ref_table}. This tests implicitly for
type compatibility. If the result of the \code{anti_join()} is:
\itemize{
\item a tibble with zero rows: the column is a candidate, there will be an empty entry in column \code{why}
of the result tibble.
\item a tibble with rows: the percentage of missing values in the parent table is calculated and included in
the entry for this column in \code{why}
\item an error: the error message becomes the entry in column \code{why} for this column (often this is because of
mismatched column types)
}
}
\examples{
cdm_enum_fk_candidates(cdm_nycflights13(), flights, airports)

}
\seealso{
Other foreign key functions: \code{\link{cdm_add_fk}},
  \code{\link{cdm_get_all_fks}}, \code{\link{cdm_get_fk}},
  \code{\link{cdm_has_fk}}, \code{\link{cdm_rm_fk}}
}
\concept{foreign key functions}
