% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{cdm_enum_fk_candidates}
\alias{cdm_enum_fk_candidates}
\title{Find foreign key candidates in a table}
\usage{
cdm_enum_fk_candidates(dm, table, ref_table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{The table whose columns should be tested for foreign key candidate potential}

\item{ref_table}{A table with a primary key.}
}
\description{
Which columns are foreign candidates of a table, referencing the primary key column of another \code{\link{dm}} object's table?
\code{cdm_enum_fk_candidates()} checks first, if \code{ref_table} has a primary key set. Then it determines
for each column of \code{table}, if this column contains only a subset of values of the primary key column of
\code{ref_table} and is therefore a candidate for a foreign key from \code{table} to \code{ref_table}.
}
\examples{
library(dplyr)

nycflights_dm <- cdm_nycflights13(cycle = TRUE)

nycflights_dm \%>\%
  cdm_enum_fk_candidates(flights, airports)
}
\seealso{
Other Foreign key functions: \code{\link{cdm_add_fk}},
  \code{\link{cdm_get_all_fks}}, \code{\link{cdm_get_fk}},
  \code{\link{cdm_has_fk}}, \code{\link{cdm_rm_fk}}
}
\concept{Foreign key functions}
