% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-dm.R
\name{cdm_filter}
\alias{cdm_filter}
\alias{cdm_semi_join}
\title{Cascading filter of a \code{\link{dm}} object}
\usage{
cdm_filter(dm, table, ...)

cdm_semi_join(dm, table, reduced_table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}

\item{...}{Logical predicates defined in terms of the variables in .data.
Multiple conditions are combined with &. Only rows where the condition evaluates
to TRUE are kept.

The arguments in ... are automatically quoted and evaluated in the context of
the data frame. They support unquoting and splicing. See vignette("programming")
for an introduction to these concepts.}

\item{reduced_table}{The table indicated in argument \code{table}, but in a filtered state (cf, \code{dplyr::filter()}).}
}
\description{
'cdm_filter()' allows you to set one or more filter conditions for one table
of a \code{\link{dm}} object. If the \code{\link{dm}}'s tables are connected via key constrains, the
filtering will affect all other connected tables, leaving only the rows with
the corresponding key values.

'cdm_semi_join()' performs a cascading "row reduction" of a \code{\link{dm}} object
by an inital semi-join of one of its tables with the same, but filtered table. Subsequently, the
key constraints are used to compute the remainders of the other tables of the \code{\link{dm}} object and
a new \code{\link{dm}} object is returned.
}
\examples{
library(magrittr)
cdm_nycflights13(cycle = FALSE) \%>\%
  cdm_filter(airports, name == "John F Kennedy Intl")
}
