% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_tables_from_dm.R
\name{cdm_find_conn_tbls}
\alias{cdm_find_conn_tbls}
\title{Find the tables connecting two or more tables in a \code{\link{dm}}}
\usage{
cdm_find_conn_tbls(dm, ...)
}
\arguments{
\item{dm}{A \code{\link{dm}} object}

\item{...}{Two or more table names of the \code{\link{dm}} object's tables.}
}
\value{
Character vector with the names of the connecting tables.
}
\description{
Find all tables that need to be passed when traversing the \code{\link{dm}} object
between the indicated tables along the foreign
key relations. Result includes the given tables.
}
\seealso{
Other Functions utilizing foreign key relations: \code{\link{cdm_get_referencing_tables}},
  \code{\link{cdm_is_referenced}},
  \code{\link{cdm_select_tbl}}
}
\concept{Functions utilizing foreign key relations}
