% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{cdm_get_fk}
\alias{cdm_get_fk}
\title{Retrieve the name of the column marked as foreign key, pointing from one table of a \code{dm} to another}
\usage{
cdm_get_fk(dm, table, ref_table)
}
\description{
Retrieve the name of the column marked as foreign key, pointing from one table of a \code{dm} to another
}
