% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-keys.R
\name{cdm_get_pk}
\alias{cdm_get_pk}
\title{Retrieve the name of the column marked as primary key of a table of a \code{\link{dm}} object}
\usage{
cdm_get_pk(dm, table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}
}
\description{
\code{cdm_get_pk()} returns the name of the
column marked as primary key of a table of a \code{\link{dm}} object. If no primary key is
set for the table, an empty character variable is returned.
}
\examples{
library(dplyr)
nycflights_dm <- cdm_nycflights13()

nycflights_dm \%>\%
  cdm_get_pk(planes)
}
\seealso{
Other Primary key functions: \code{\link{cdm_add_pk}},
  \code{\link{cdm_get_all_pks}}, \code{\link{cdm_has_pk}},
  \code{\link{cdm_rm_pk}}, \code{\link{enum_pk_candidates}}
}
\concept{Primary key functions}
