% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{cdm_has_fk}
\alias{cdm_has_fk}
\title{Does a reference from one table of a \code{dm} to another exist?}
\usage{
cdm_has_fk(dm, table, ref_table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}

\item{ref_table}{The table which \code{table} is potentially referencing.}
}
\value{
A boolean value: \code{TRUE}, if a reference from \code{table} to \code{ref_table} exists, \code{FALSE} otherwise.
}
\description{
Does a reference from one table of a \code{dm} to another exist?
}
