% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{cdm_is_referenced}
\alias{cdm_is_referenced}
\title{Is a \code{\link{dm}}'s table referenced by another table?}
\usage{
cdm_is_referenced(dm, table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}
}
\value{
\code{TRUE}, if at least one foreign key exists, pointing to the primary
key of parameter \code{table}, \code{FALSE} otherwise.
}
\description{
Is a \code{\link{dm}}'s table referenced by another table?
}
\seealso{
Other Functions utilizing foreign key relations: \code{\link{cdm_find_conn_tbls}},
  \code{\link{cdm_get_referencing_tables}},
  \code{\link{cdm_select_tbl}}
}
\concept{Functions utilizing foreign key relations}
