% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-helpers.R
\name{cdm_join_tbl}
\alias{cdm_join_tbl}
\title{Perform a join between two tables of a \code{\link{dm}}}
\usage{
cdm_join_tbl(dm, lhs, rhs, join = semi_join)
}
\arguments{
\item{dm}{A \code{\link{dm}} object}

\item{lhs}{The table on the left hand side of the join}

\item{rhs}{The table on the right hand side of the join}

\item{join}{The type of join to be performed, see \code{\link[dplyr]{join}}}
}
\value{
The resulting table of the join
}
\description{
A join of desired type is performed between table \code{lhs} and
table \code{rhs}.
}
\concept{Flattening functions}
