% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-helpers.R
\name{cdm_nrow}
\alias{cdm_nrow}
\title{Number of rows of a table or of the whole \code{dm}-object}
\usage{
cdm_nrow(dm, table = NULL)
}
\description{
Number of rows of a table or of the whole \code{dm}-object
}
