% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-helpers.R
\name{cdm_nrow}
\alias{cdm_nrow}
\title{Number of rows}
\usage{
cdm_nrow(dm, table = NULL)
}
\arguments{
\item{dm}{A \code{\link{dm}} object}
}
\description{
Returns a named vector with the number of rows for each table.
}
