% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{cdm_rm_fk}
\alias{cdm_rm_fk}
\title{Remove the reference(s) from one \code{\link{dm}} table to another}
\usage{
cdm_rm_fk(dm, table, column, ref_table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}.}

\item{column}{The column of \code{table} that should no longer be referencing the primary key of \code{ref_table}.
If \code{NULL}, all columns will be evaluated.}

\item{ref_table}{The table that \code{table} was referencing.}
}
\description{
This function can remove either one reference between two tables, or all references at once, if argument \code{column = NULL}.
All arguments may be provided quoted or unquoted.
}
\seealso{
Other foreign key functions: 
\code{\link{cdm_enum_fk_candidates}()},
\code{\link{cdm_get_all_fks}()},
\code{\link{cdm_get_fk}()},
\code{\link{cdm_has_fk}()},
\code{\link{dm_add_fk}()}
}
\concept{foreign key functions}
