% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-keys.R
\name{cdm_rm_pk}
\alias{cdm_rm_pk}
\title{Remove primary key from a table in a \code{\link{dm}} object}
\usage{
cdm_rm_pk(dm, table, rm_referencing_fks = FALSE)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}

\item{rm_referencing_fks}{Boolean: if \code{FALSE} (default), will throw an error, if
there are foreign keys addressing the primary key to be removed. If \code{TRUE}, will
in addition to the primary key of parameter \code{table}, also remove all foreign key constraints
that are pointing to it.}
}
\description{
\code{cdm_rm_pk()} removes a potentially set primary key from a table in the
underlying \code{data_model}-object and otherwise leaves the \code{\link{dm}} object untouched.

Foreign keys pointing to the table from other tables can optionally be removed as well.
}
\examples{
library(dplyr)
nycflights_dm <- cdm_nycflights13()

nycflights_dm \%>\%
  cdm_rm_pk(airports, rm_referencing_fks = TRUE) \%>\%
  cdm_has_pk(planes)

nycflights_dm \%>\%
  cdm_rm_pk(planes, rm_referencing_fks = TRUE) \%>\%
  cdm_has_pk(planes)
}
\seealso{
Other Primary key functions: \code{\link{cdm_add_pk}},
  \code{\link{cdm_get_all_pks}}, \code{\link{cdm_get_pk}},
  \code{\link{cdm_has_pk}},
  \code{\link{enum_pk_candidates}}
}
\concept{Primary key functions}
