% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-tbl.R
\name{cdm_rm_tbl}
\alias{cdm_rm_tbl}
\title{Remove tables from a \code{\link{dm}}}
\usage{
cdm_rm_tbl(dm, ...)
}
\arguments{
\item{dm}{A \code{\link{dm}} object.}

\item{...}{One or more unquoted table names to remove from the \code{dm}.
\code{tidyselect} is supported, see \code{\link[tidyselect:vars_select]{tidyselect::vars_select()}} and \code{\link[tidyselect:vars_rename]{tidyselect::vars_rename()}}
for details on the semantics.}
}
\value{
The dm without the removed table(s) that were present in the initial \code{dm}.
}
\description{
Removes one or more tables from a \code{\link{dm}}.
}
\seealso{
\code{\link[=cdm_add_tbl]{cdm_add_tbl()}}, \code{\link[=cdm_select_tbl]{cdm_select_tbl()}}
}
