% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-management.R
\name{cdm_rm_tbl}
\alias{cdm_rm_tbl}
\title{Remove tables from a \code{\link{dm}}}
\usage{
cdm_rm_tbl(dm, ...)
}
\arguments{
\item{dm}{A \code{\link{dm}} object}

\item{...}{One or more unquoted tibble names to remove from the \code{dm}.}
}
\value{
The inital \code{dm} without the removed table(s).
}
\description{
Removes one or more tibbles from a \code{\link{dm}}.
}
\seealso{
\code{\link[=cdm_add_tbl]{cdm_add_tbl()}}, \code{\link[=cdm_select_tbl]{cdm_select_tbl()}}
}
