% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-tbl.R
\name{dm_rename_tbl}
\alias{dm_rename_tbl}
\title{Change the names of the tables in a \code{dm}}
\usage{
dm_rename_tbl(dm, ...)
}
\description{
\code{dm_rename_tbl()} renames tables.
}
\examples{
cdm_nycflights13() \%>\%
  cdm_rename_tbl(ap = airports, fl = flights)
}
