% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-tbl.R
\name{cdm_select_tbl}
\alias{cdm_select_tbl}
\alias{cdm_rename_tbl}
\title{Select and rename tables}
\usage{
cdm_select_tbl(dm, ...)

cdm_rename_tbl(dm, ...)
}
\arguments{
\item{dm}{A \code{\link{dm}} object}

\item{...}{One or more table names of the \code{\link{dm}} object's tables.
See \code{\link[tidyselect:vars_select]{tidyselect::vars_select()}} and \code{\link[tidyselect:vars_rename]{tidyselect::vars_rename()}}
for details on the semantics.}
}
\value{
The input \code{dm} with tables renamed or removed.
}
\description{
\code{cdm_select_tbl()} keeps the selected tables and their relationships,
optionally renaming them.

\code{cdm_rename_tbl()} renames tables.
}
