% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_tables_from_dm.R
\name{cdm_select_tbl}
\alias{cdm_select_tbl}
\title{Get a \code{\link{dm}} only containing the indicated tables}
\usage{
cdm_select_tbl(dm, ...)
}
\arguments{
\item{dm}{A \code{\link{dm}} object}

\item{...}{Two or more table names of the \code{\link{dm}} object's tables.
See \code{\link[tidyselect:vars_select]{tidyselect::vars_select()}} for details on the semantics.}
}
\description{
Returns a reduced \code{\link{dm}} object, taking care of the key relations
}
