% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key-helpers.R
\name{check_subset}
\alias{check_subset}
\title{Check column values for subset}
\usage{
check_subset(x, y, ..., x_select = NULL, y_select = NULL)
}
\arguments{
\item{x, y}{A data frame or lazy table.}

\item{...}{These dots are for future extensions and must be empty.}

\item{x_select, y_select}{Names of key columns, processed with
\code{\link[tidyselect:eval_select]{tidyselect::eval_select()}}, to restrict the check.}
}
\value{
Returns \code{x}, invisibly, if the check is passed.
Otherwise an error is thrown and the reason for it is explained.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{check_subset()} tests if \code{x} is a subset of \code{y}.
For convenience, the \code{x_select} and \code{y_select} arguments allow restricting the check
to a set of key columns without affecting the return value.
}
\examples{
data_1 <- tibble::tibble(a = c(1, 2, 1), b = c(1, 4, 1), c = c(5, 6, 7))
data_2 <- tibble::tibble(a = c(1, 2, 3), b = c(4, 5, 6), c = c(7, 8, 9))
# this is passing:
check_subset(data_1, a, data_2, a)

# this is failing:
try(check_subset(data_2, a, data_1, a))
}
