% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_surgery.R
\name{decompose_table}
\alias{decompose_table}
\title{Perform table surgery by extracting a 'parent table' from a table, returning both and linking them by a key}
\usage{
decompose_table(.data, new_id_column, ...)
}
\description{
Perform table surgery by extracting a 'parent table' from a table, returning both and linking them by a key
}
