% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/zzx-deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{cdm_get_src}
\alias{cdm_get_con}
\alias{cdm_get_tables}
\alias{cdm_get_filter}
\alias{cdm_add_tbl}
\alias{cdm_rm_tbl}
\alias{cdm_copy_to}
\alias{cdm_disambiguate_cols}
\alias{cdm_draw}
\alias{cdm_set_colors}
\alias{cdm_get_colors}
\alias{cdm_get_available_colors}
\alias{cdm_filter}
\alias{cdm_nrow}
\alias{cdm_flatten_to_tbl}
\alias{cdm_squash_to_tbl}
\alias{cdm_join_to_tbl}
\alias{cdm_apply_filters}
\alias{cdm_apply_filters_to_tbl}
\alias{cdm_add_pk}
\alias{cdm_add_fk}
\title{Deprecated functions}
\usage{
cdm_get_src(x)

cdm_get_con(x)

cdm_get_tables(x)

cdm_get_filter(x)

cdm_add_tbl(dm, ..., repair = "unique", quiet = FALSE)

cdm_rm_tbl(dm, ...)

cdm_copy_to(
  dest,
  dm,
  ...,
  types = NULL,
  overwrite = NULL,
  indexes = NULL,
  unique_indexes = NULL,
  set_key_constraints = TRUE,
  unique_table_names = FALSE,
  table_names = NULL,
  temporary = TRUE
)

cdm_disambiguate_cols(dm, sep = ".", quiet = FALSE)

cdm_draw(
  dm,
  rankdir = "LR",
  col_attr = "column",
  view_type = "keys_only",
  columnArrows = TRUE,
  graph_attrs = "",
  node_attrs = "",
  edge_attrs = "",
  focus = NULL,
  graph_name = "Data Model"
)

cdm_set_colors(dm, ...)

cdm_get_colors(dm)

cdm_get_available_colors()

cdm_filter(dm, table, ...)

cdm_nrow(dm)

cdm_flatten_to_tbl(dm, start, ..., join = left_join)

cdm_squash_to_tbl(dm, start, ..., join = left_join)

cdm_join_to_tbl(dm, table_1, table_2, join = left_join)

cdm_apply_filters(dm)

cdm_apply_filters_to_tbl(dm = dm, table = table)

cdm_add_pk(dm, table, column, check = FALSE, force = FALSE)

cdm_add_fk(dm, table, column, ref_table, check = FALSE)
}
\description{
These functions are deprecated in favor of better alternatives.
Most functions with the \code{cdm_} prefix have an identical alternative
with a \code{dm_} prefix.
}
\keyword{internal}
