% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/zzx-deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{cdm_get_src}
\title{Deprecated functions}
\usage{
cdm_get_src(x)
}
\description{
These functions are deprecated in favor of better alternatives.
Most functions with the \code{cdm_} prefix have an identical alternative
with a \code{dm_} prefix.
}
\keyword{internal}
