% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/zzx-deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{cdm_get_src}
\alias{cdm_get_con}
\alias{cdm_get_tables}
\alias{cdm_get_filter}
\alias{cdm_add_tbl}
\alias{cdm_rm_tbl}
\title{Deprecated functions}
\usage{
cdm_get_src(x)

cdm_get_con(x)

cdm_get_tables(x)

cdm_get_filter(x)

cdm_add_tbl(dm, ..., repair = "unique", quiet = FALSE)

cdm_rm_tbl(dm, ...)
}
\description{
These functions are deprecated in favor of better alternatives.
Most functions with the \code{cdm_} prefix have an identical alternative
with a \code{dm_} prefix.
}
\keyword{internal}
