% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differ.R
\name{differ2}
\alias{differ2}
\title{Code differences between two vectors}
\usage{
differ2(x, y)
}
\arguments{
\item{x}{A vector.}
}
\description{
Code differences between two vectors as 1; else, 0.
}
\seealso{
\url{https://github.com/robertschnitman/dm}
}
