% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dm}
\alias{dm}
\alias{new_dm}
\alias{validate_dm}
\alias{dm_get_src}
\alias{dm_get_data_model}
\alias{is_dm}
\alias{as_dm}
\title{Data model class}
\usage{
dm(src, data_model = NULL)

new_dm(src, data_model)

validate_dm(x)

dm_get_src(x)

dm_get_data_model(x)

is_dm(x)

as_dm(x)
}
\arguments{
\item{src}{A \pkg{dplyr} table source object.}

\item{data_model}{A \pkg{datamodelr} data model object, or \code{NULL}.}

\item{x}{An object.}
}
\description{
The \code{dm} class wraps \link[dplyr:src]{dplyr::src} and adds a description of table relationships
based on \link[datamodelr:datamodelr-package]{datamodelr::datamodelr-package}.

\code{dm()} coerces its inputs.

\code{new_dm()} only checks if the inputs are of the correct class.

\code{validate_dm()} checks consistency between the \pkg{dplyr} source
and the \pkg{datamodelr} based specification of table relationships.

\code{dm_get_src()} returns the \pkg{dplyr} source component of a \code{dm}.
object.

\code{dm_get_data_model()} returns the \pkg{datamodelr} data model component of a \code{dm}.
object.

\code{is_dm()} returns \code{TRUE} if the input is of class \code{dm}.

\code{as_dm()} coerces objects to the \code{dm} class
}
