% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dm}
\alias{dm}
\alias{new_dm}
\alias{validate_dm}
\alias{dm_get_src}
\alias{dm_get_data_model}
\alias{is_dm}
\alias{as_dm}
\title{Data model class}
\usage{
dm(src, data_model = NULL)

new_dm(src, data_model)

validate_dm(x)

dm_get_src(x)

dm_get_data_model(x)

is_dm(x)

as_dm(x)
}
\arguments{
\item{src}{A \pkg{dplyr} table source object.}

\item{data_model}{A \pkg{datamodelr} data model object, or `NULL`.}

\item{x}{An object.}
}
\description{
The `dm` class wraps [dplyr::src] and adds a description of table relationships
based on [datamodelr::datamodelr-package].

`dm()` coerces its inputs.

`new_dm()` only checks if the inputs are of the correct class.

`validate_dm()` checks consistency between the \pkg{dplyr} source
and the \pkg{datamodelr} based specification of table relationships.

`dm_get_src()` returns the \pkg{dplyr} source component of a `dm`.
object.

`dm_get_data_model()` returns the \pkg{datamodelr} data model component of a `dm`.
object.

`is_dm()` returns `TRUE` if the input is of class `dm`.

`as_dm()` coerces objects to the `dm` class
}
