% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DM.R
\name{DM}
\alias{DM}
\alias{DMVRAP}
\title{Shiny app to run DM and VRAP}
\usage{
DM()
}
\description{
This function takes the input from an A & P files, runs DM to estimate SR parameters and saves a .rav file and .RData file. The .rav file can be downloaded or directly used for running the VRAP simulations by clicking the VRAP tab.
}
