% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dm}
\alias{dm}
\alias{dmv}
\title{Data Management Tool (interactive-ish)}
\usage{
dm(
  var,
  name = var,
  db = NULL,
  recode = NULL,
  transf = NULL,
  group = NULL,
  comment = NULL,
  label = NULL,
  keep.label = TRUE
)

dmv(
  var,
  name = var,
  db = NULL,
  recode = NULL,
  transf = NULL,
  group = NULL,
  comment = NULL,
  label = NULL,
  keep.label = TRUE
)
}
\arguments{
\item{var}{character, database/data frame entry}

\item{name}{character, (new) name of variable, will be 'var' if unspecified}

\item{db}{character, name of data frame, if \code{NULL}, will look for
default setting in \code{opts_dm$get('default_db')}.}

\item{recode}{recode \code{L} argument for recoding}

\item{transf}{function, for transformation}

\item{group}{character, a grouping of the variables}

\item{comment}{character, a comment or some such extra information}

\item{label}{character, something to be stored as label attribute for this
variable}

\item{keep.label}{logical, if \code{var} has a label in \code{db}, should it
be kept?  (only if \code{label} is \code{NULL})}
}
\description{
This creates a list of data management information (defaults to
    a 'dm_doc' object called 'documentation' in the environment 'dm_envir'
    which can be retrieved with \code{dm_doc()}). This information can then
    be used to extract and recode (factors and dates) for (or towards) an
    analytical data base.
}
\section{Functions}{
\itemize{
\item \code{dmv}: An alias for dm ('v' for 'variable')
}}

