% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{dm_add_fk}
\alias{dm_add_fk}
\title{Add a reference from one table of a \code{\link{dm}} to another}
\usage{
dm_add_fk(dm, table, column, ref_table, check = FALSE)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}.}

\item{column}{The column of \code{table} which is to become the foreign key column and
reference the primary key of \code{ref_table}.}

\item{ref_table}{The table which \code{table} is referencing.
This table needs to have a primary key set.}

\item{check}{Boolean, if \code{TRUE}, a check will be performed to determine if the values of
\code{column} are a subset of the values of the primary key column of \code{ref_table}.}
}
\value{
An updated \code{dm} with an additional foreign key relation.
}
\description{
Add a reference from one table of a \code{\link{dm}} to another
}
\examples{
library(dplyr)
iris_key <- mutate(iris, key = row_number()) \%>\%
  select(key, everything())
dm(iris_1 = iris_key, iris_2 = iris_key) \%>\%
  dm_add_pk(iris_2, key) \%>\%
  dm_add_fk(iris_1, key, iris_2)
}
\seealso{
Other foreign key functions: 
\code{\link{dm_enum_fk_candidates}()},
\code{\link{dm_get_all_fks}()},
\code{\link{dm_get_fk}()},
\code{\link{dm_has_fk}()},
\code{\link{dm_rm_fk}()}
}
\concept{foreign key functions}
