% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{dm_add_fk}
\alias{dm_add_fk}
\alias{dm_rm_fk}
\title{Add/remove foreign keys}
\usage{
dm_add_fk(
  dm,
  table,
  columns,
  ref_table,
  ref_columns = NULL,
  ...,
  check = FALSE
)

dm_rm_fk(dm, table, columns, ref_table, ...)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}.}

\item{columns}{For \code{dm_add_fk()}: The columns of \code{table} which are to become the foreign key columns that
reference the primary key of \code{ref_table}.

For \code{dm_rm_fk()}: The columns of \code{table} that should no longer be referencing the primary key of \code{ref_table}.
If \code{NULL}, all columns will be evaluated.

To define a compound key, use \code{c(col1, col2)}.}

\item{ref_table}{For \code{dm_add_fk()}: The table which \code{table} will be referencing.
This table needs to have a primary key set.

For \code{dm_rm_fk()}: The table that \code{table} is referencing.}

\item{ref_columns}{For \code{dm_add_fk()}: The columns of \code{table} which are to become the referenced columns in \code{ref_table}.
By default, the primary key is used.

For \code{dm_rm_fk()}: The columns of \code{table} that should no longer be referencing the primary key of \code{ref_table}.

To define a compound key, use \code{c(col1, col2)}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{check}{Boolean, if \code{TRUE}, a check will be performed to determine if the values of
\code{column} are a subset of the values of the primary key column of \code{ref_table}.}
}
\value{
For \code{dm_add_fk()}: An updated \code{dm} with an additional foreign key relation.

For \code{dm_rm_fk()}: An updated \code{dm} without the given foreign key relation.
}
\description{
\code{dm_add_fk()} marks the specified columns as the foreign key of table \code{table} with
respect to the primary key of table \code{ref_table}.
If \code{check == TRUE}, then it will first check if the values in columns \code{columns} are a subset
of the values of the primary key in table \code{ref_table}.

\code{dm_rm_fk()} can remove either one reference between two tables, or all references at once, if argument \code{columns = NULL}.
All arguments may be provided quoted or unquoted.
}
\examples{
\dontshow{if (rlang::is_installed("nycflights13") && rlang::is_installed("DiagrammeR")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
nycflights_dm <- dm(
  planes = nycflights13::planes,
  flights = nycflights13::flights,
  weather = nycflights13::weather
)

nycflights_dm \%>\%
  dm_draw()

# Create foreign keys:
nycflights_dm \%>\%
  dm_add_pk(planes, tailnum) \%>\%
  dm_add_fk(flights, tailnum, planes) \%>\%
  dm_add_pk(weather, c(origin, time_hour)) \%>\%
  dm_add_fk(flights, c(origin, time_hour), weather) \%>\%
  dm_draw()

# Keys can be checked during creation:
try(
  nycflights_dm \%>\%
    dm_add_pk(planes, tailnum) \%>\%
    dm_add_fk(flights, tailnum, planes, check = TRUE)
)
\dontshow{\}) # examplesIf}
\dontshow{if (rlang::is_installed("nycflights13") && rlang::is_installed("DiagrammeR")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dm_nycflights13(cycle = TRUE) \%>\%
  dm_rm_fk(flights, dest, airports) \%>\%
  dm_draw()
\dontshow{\}) # examplesIf}
}
\seealso{
Other foreign key functions: 
\code{\link{dm_enum_fk_candidates}()},
\code{\link{dm_get_all_fks}()},
\code{\link{dm_get_fk}()},
\code{\link{dm_has_fk}()}

Other foreign key functions: 
\code{\link{dm_enum_fk_candidates}()},
\code{\link{dm_get_all_fks}()},
\code{\link{dm_get_fk}()},
\code{\link{dm_has_fk}()}
}
\concept{foreign key functions}
