% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-keys.R
\name{dm_add_pk}
\alias{dm_add_pk}
\title{Mark columns of a table in a \code{\link{dm}} object as its primary key}
\usage{
dm_add_pk(dm, table, columns, check = FALSE, force = FALSE)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}.}

\item{columns}{Columns of that table.}

\item{check}{Boolean, if \code{TRUE}, a check is made if the combination of columns is a unique key of the table.}

\item{force}{Boolean, if \code{FALSE} (default), an error will be thrown if there is already a primary key
set for this table.
If \code{TRUE}, a potential old \code{pk} is deleted before setting a new one.}
}
\value{
An updated \code{dm} with an additional primary key.
}
\description{
\code{dm_add_pk()} marks the specified columns as the primary key of the specified table.
If \code{check == TRUE}, then it will first check if
the given combination of columns is a unique key of the table.
If \code{force == TRUE}, the function will replace an already
set key.

FYI: Mind, that in one of the upcoming updates, compound keys will be supported, and
therefore the syntax for this function will slightly change.
}
\examples{
library(dplyr)

nycflights_dm <- dm_from_src(src_df(pkg = "nycflights13"))

# the following works
dm_add_pk(nycflights_dm, planes, tailnum)
dm_add_pk(nycflights_dm, airports, faa, check = TRUE)
dm_add_pk(nycflights_dm, planes, manufacturer)

# the following does not work (throws an error)
try(dm_add_pk(nycflights_dm, planes, manufacturer, check = TRUE))
}
\seealso{
Other primary key functions: 
\code{\link{dm_get_all_pks}()},
\code{\link{dm_get_pk}()},
\code{\link{dm_has_pk}()},
\code{\link{dm_rm_pk}()},
\code{\link{enum_pk_candidates}()}
}
\concept{primary key functions}
