% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-keys.R
\name{dm_add_pk}
\alias{dm_add_pk}
\title{Mark a column of a table in a \code{\link{dm}} object as its primary key}
\usage{
dm_add_pk(dm, table, column, check = FALSE, force = FALSE)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}

\item{column}{A column of that table}

\item{check}{Boolean, if \code{TRUE} (default), a check is made if the column is a unique key of the table.}

\item{force}{Boolean, if \code{FALSE} (default), an error will be thrown, if there is
already a primary key set for this table. If \code{TRUE} a potential old \code{pk} is deleted before setting the new one.}
}
\description{
\code{dm_add_pk()} marks the given column as the given table's primary key
in the \code{data_model}-part of the \code{\link{dm}} object. If \code{check == TRUE}, it also first checks if
the given column is a unique key of the table. If \code{force == TRUE}, it replaces an already
set key.
}
\examples{
library(dplyr)


nycflights_dm <- dm(src_df(pkg = "nycflights13"))

# the following works
dm_add_pk(nycflights_dm, planes, tailnum)
dm_add_pk(nycflights_dm, airports, faa)
dm_add_pk(nycflights_dm, planes, manufacturer, check = FALSE)

# the following does not work
try(dm_add_pk(nycflights_dm, planes, manufacturer))
}
\seealso{
Other primary key functions: \code{\link{dm_get_all_pks}},
  \code{\link{dm_get_pk}}, \code{\link{dm_has_pk}},
  \code{\link{dm_rm_pk}}, \code{\link{enum_pk_candidates}}
}
\concept{primary key functions}
