% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_keys.R
\name{dm_add_primary_key}
\alias{dm_add_primary_key}
\title{Mark a column of a table in a \code{dm}-object as its primary key}
\usage{
dm_add_primary_key(dm, table, column, check_if_unique_key = TRUE)
}
\description{
\code{dm_add_primary_key()} first checks , if the given column
is a unique key of the given table in the \code{dm}-object. If yes, it then marks
that column as the table's primary key in the \code{data_model}-part of the \code{dm}-object.
}
\examples{
\dontrun{
library(nycflights13)
library(dplyr)

nycflights_dm <- dm(src_df(pkg = "nycflights13"))

# the following works
dm_add_primary_key(nycflights_dm, "planes", "tailnum")
dm_add_primary_key(nycflights_dm, "airports", faa)
dm_add_primary_key(nycflights_dm, "planes", "manufacturer", check_if_unique_key = FALSE)

# the following does not work
dm_add_primary_key(nycflights_dm, "planes", "manufacturer")
}
}
