% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambiguate.R
\name{dm_disambiguate_cols}
\alias{dm_disambiguate_cols}
\title{Avoid column name conflicts}
\usage{
dm_disambiguate_cols(dm, sep = ".", quiet = FALSE)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{sep}{The character variable separating the table names and the ambiguous column names}

\item{quiet}{Boolean. By default this function lists the renamed columns in a message, pass \code{FALSE} to suppress this message.}
}
\description{
This function checks all tables for column names that are not unique (across the entire \code{dm} object) and renames
those columns by prefixing the respective table name and a separator.
Key columns will not be renamed, since only one column should remain, when two tables
linked by a key relation are joined.
}
\examples{
dm_disambiguate_cols(dm_nycflights13())
}
