% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disentangle.R
\name{dm_disentangle}
\alias{dm_disentangle}
\title{Remove cycles from a \code{\link{dm}}}
\usage{
dm_disentangle(dm, naming_template = NULL, quiet = FALSE)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{naming_template}{Naming template for the tables to be created as a length 1 character variable.
Default \code{NULL} leads to names of the form: \verb{<parent_table>_<row_number>}
Keywords are:
\itemize{
\item \code{.pt}  : original name of parent table
\item \code{.pkc} : primary key column name(s) of original parent table. Multiple columns are separated by an underscore.
\item \code{.ct}  : name of child table
\item \code{.fkc} : foreign key column name(s). Multiple columns are separated by an underscore
\item \code{.num} : integer counting the tables based on each original parent table
}}

\item{quiet}{Boolean.
By default, this function lists the renamed columns in a message, pass \code{TRUE} to suppress this message.}
}
\value{
A cycle-free \code{dm} object.
}
\description{
Produce a cycle-free \code{\link{dm}} by multiplying parent tables
}
\details{
Starting from the graph representation of the \code{dm}, it is tested which
of the graph's components (connected undirectional subgraphs) has a cycle.
Now taking the direction of the relations into account, for components with a cycle
it is now it is determined which vertices (here: parent tables) have both:
\itemize{
\item 2 or more incoming foreign keys
\item multiple possible (undirected) paths between the participating tables in at last one of those foreign keys
}

Those tables are copied and reinserted into the \code{dm} with a new name while distributing the
foreign keys among the newly created tables.

This function is a no-op if:
\itemize{
\item no cycles are detected.
\item at least one "endless" cycles is detected (possible to walk endlessly in direction of arrows).
}
}
\examples{
dm_disentangle(dm_nycflights13())
dm_disentangle(dm_nycflights13(cycle = TRUE))
dm_disentangle(dm_nycflights13(cycle = TRUE), naming_template = ".pt..fkc")
}
