% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dm_doc2latex}
\alias{dm_doc2latex}
\title{documentation as LaTeX}
\usage{
dm_doc2latex(
  doc = NULL,
  file = "",
  where = "htb",
  rowname = NULL,
  ...,
  group = TRUE,
  grouping = NULL,
  which = c("group", "name", "var", "db", "label", "comment"),
  code.key = c(group = "\\\\textbf", name = "\\\\texttt", var = "\\\\texttt", db =
    "\\\\textbf", comment = "\\\\emph")
)
}
\arguments{
\item{doc}{documentation}

\item{file}{argument for \code{Hmisc::latex}, default "" (empty string)}

\item{where}{argument for \code{Hmisc::latex}, default "htb"}

\item{rowname}{argument for \code{Hmisc::latex}, default \code{NULL}}

\item{...}{passed to \code{Hmisc::latex}}

\item{group}{group on 'group' (if it exists)}

\item{grouping}{specify the order to group by (character vector), else as-is}

\item{which}{columns to print}

\item{code.key}{formatting code key}
}
\description{
print information from documentation in LaTeX code
}
