% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{dm_enum_fk_candidates}
\alias{dm_enum_fk_candidates}
\title{Find foreign key candidates in a table}
\usage{
dm_enum_fk_candidates(dm, table, ref_table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{The table whose columns should be tested for foreign key candidate potential}

\item{ref_table}{A table with a primary key.}
}
\value{
A table with an overview which columns of \code{table} would be suitable candidates as
foreign key columns referencing \code{ref_table} and which columns would not.
}
\description{
Which columns are good candidates as a foreign key of a table, referencing
the primary key column of another \code{\link{dm}} object's table?
}
\details{
\code{dm_enum_fk_candidates()} checks first, if \code{ref_table} has a primary key set.
For each column of \code{table} a join operation is then tried, with parameter \code{by} matching
the respective column with the primary key of \code{ref_table}. This tests implicitly for
type compatibility (on most sources). Based on the result of the join, the
entry in the result column \code{why} is:
\itemize{
\item an empty entry, if the column is a candidate
\item the total percentage and individual numbers of missing matches between the entries of the
respective column in table \code{table} and the primary column entries in table \code{ref_table}.
\item the error message triggered by the error (often stating the mismatched column types)
}
}
\examples{
dm_enum_fk_candidates(dm_nycflights13(), flights, airports)

}
\seealso{
Other foreign key functions: \code{\link{dm_add_fk}},
  \code{\link{dm_get_all_fks}}, \code{\link{dm_get_fk}},
  \code{\link{dm_has_fk}}, \code{\link{dm_rm_fk}}
}
\concept{foreign key functions}
