% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key-helpers.R
\name{dm_examine_constraints}
\alias{dm_examine_constraints}
\title{Validate your \code{\link{dm}}: are all key constraints met?}
\usage{
dm_examine_constraints(dm)
}
\arguments{
\item{dm}{A \code{dm} object.}
}
\value{
A tibble with 5 columns:
\enumerate{
\item \code{table}: the table in the \code{dm},
\item \code{kind}: "PK" or "FK",
\item \code{column}: a column of the table,
\item \code{ref_table}: for foreign keys, the referenced table,
\item \code{is_key}: logical,
\item \code{problem}: if \code{is_key = FALSE}, the reason for that.
}
}
\description{
This function returns a tibble with information about which key constraints are met (\code{is_key = TRUE}) or violated (\code{FALSE}).
}
\details{
For the primary key constraints, it is tested if the values in the respective columns are all unique.
For the foreign key constraints, the tests check if for each foreign key constraint, the values of the foreign key column
form a subset of the values of the referenced column.
}
\examples{
dm_examine_constraints(dm_nycflights13())
}
