% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{dm_filter}
\alias{dm_filter}
\title{get or reset documentation for filtering}
\usage{
dm_filter(kill = FALSE, prompt = TRUE)
}
\arguments{
\item{kill}{if \code{TRUE} deletes all documentation}

\item{prompt}{if \code{TRUE} prompts before deleting}
}
\description{
get or reset the list that has documentet a filtering process done by dmf
}
