% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial.R
\name{dm_financial}
\alias{dm_financial}
\alias{dm_financial_sqlite}
\title{Creates a dm object for the Financial data}
\usage{
dm_financial()

dm_financial_sqlite()
}
\value{
A \code{dm} object.
}
\description{
\lifecycle{experimental}

\code{dm_financial()} creates an example \code{\link{dm}} object from the tables at
\url{https://relational.fit.cvut.cz/dataset/Financial}.
The connection is established once per session,
subsequent calls return the same connection.

\code{dm_financial_sqlite()} copies the data to a temporary SQLite database.
The data is downloaded once per session, subsequent calls return the same database.
The \code{trans} table is excluded due to its size.
}
\examples{
\dontrun{
if (rlang::is_installed("RMariaDB")) {
  dm_financial() \%>\%
    dm_draw()
}
}
}
