% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{dm_get_all_fks}
\alias{dm_get_all_fks}
\title{Retrieve all foreign key constraints in a \code{\link{dm}}}
\usage{
dm_get_all_fks(dm)
}
\arguments{
\item{dm}{A \code{dm} object.}
}
\value{
A tibble with the following columns:
\describe{
\item{\code{child_table}}{child table,}
\item{\code{child_fk_col}}{foreign key column in child table,}
\item{\code{parent_table}}{parent table.}
}
}
\description{
Get a summary of all foreign key relations in a \code{\link{dm}}
}
\examples{
dm_get_all_fks(dm_nycflights13())
}
\seealso{
Other foreign key functions: 
\code{\link{dm_add_fk}()},
\code{\link{dm_enum_fk_candidates}()},
\code{\link{dm_get_fk}()},
\code{\link{dm_has_fk}()},
\code{\link{dm_rm_fk}()}
}
\concept{foreign key functions}
