% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{dm_get_filters}
\alias{dm_get_filters}
\title{Get filter expressions}
\usage{
dm_get_filters(x)
}
\value{
A tibble with the following columns:
\describe{
\item{\code{table}}{table that was filtered,}
\item{\code{filter}}{the filter expression,}
\item{\code{zoomed}}{logical, does the filter condition relate to the zoomed table.}
}
}
\description{
\code{dm_get_filters()} returns the filter expressions that have been applied to a \code{dm} object.
These filter expressions are not intended for evaluation, only for
information.
}
