% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{dm_get_fk}
\alias{dm_get_fk}
\title{Retrieve the name of the column marked as foreign key, pointing from one table of a \code{\link{dm}} to another}
\usage{
dm_get_fk(dm, table, ref_table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}

\item{ref_table}{The table which is referenced from \code{table}.}
}
\description{
Retrieve the name of the column marked as foreign key, pointing from one table of a \code{\link{dm}} to another
}
\seealso{
Other foreign key functions: \code{\link{dm_add_fk}},
  \code{\link{dm_enum_fk_candidates}},
  \code{\link{dm_get_all_fks}}, \code{\link{dm_has_fk}},
  \code{\link{dm_rm_fk}}
}
\concept{foreign key functions}
