% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-keys.R
\name{dm_get_pk}
\alias{dm_get_pk}
\title{Retrieve the name of the primary key columns of a \code{dm} table}
\usage{
dm_get_pk(dm, table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}.}
}
\value{
A character vector with the column name(s) of the primary key of \code{table}.
}
\description{
\code{dm_get_pk()} returns the names of the
columns marked as primary key of a table of a \code{\link{dm}} object.
If no primary key is
set for the table, an empty character vector is returned.

FYI: Mind, that in one of the upcoming updates, compound keys will be supported, and
therefore the result for this function can include several columns.
}
\examples{
library(dplyr)
nycflights_dm <- dm_nycflights13()

nycflights_dm \%>\%
  dm_get_pk(planes)
}
\seealso{
Other primary key functions: 
\code{\link{dm_add_pk}()},
\code{\link{dm_get_all_pks}()},
\code{\link{dm_has_pk}()},
\code{\link{dm_rm_pk}()},
\code{\link{enum_pk_candidates}()}
}
\concept{primary key functions}
