% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_keys.R
\name{dm_get_pk}
\alias{dm_get_pk}
\title{Retrieve the name of the column marked as primary key of a table of a \code{dm}-object}
\usage{
dm_get_pk(dm, table)
}
\description{
\code{dm_get_pk()} returns the name of the
column marked as primary key of a table of a \code{dm}-object. If no primary key is
set for the table, an empty character variable is returned.
}
\examples{
\dontrun{
library(nycflights13)
library(dplyr)

nycflights_dm <- dm(src_df(pkg = "nycflights13"))
dm_obj_with_keys <- dm_add_pk(nycflights_dm, "planes", "tailnum")

dm_obj_with_keys \%>\%
  dm_get_pk("planes")
}

}
