% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary-keys.R
\name{dm_get_pk}
\alias{dm_get_pk}
\title{Primary key column names}
\usage{
dm_get_pk(dm, table, ...)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A list with character vectors with the column name(s) of the
primary keys of \code{table}.
The contained character vectors have length greater than one for compound keys.
}
\description{
\code{dm_get_pk()} returns the names of the
columns marked as primary key of a table of a \code{\link{dm}} object.
If no primary key is
set for the table, an empty character vector is returned.
}
\section{Multiple primary keys}{


Each table can have only one primary key.
For symmetry with \code{dm_get_fk()},
this function still returns a list of character vectors.
}

\examples{
\dontshow{if (rlang::is_installed("nycflights13")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dm_nycflights13() \%>\%
  dm_get_pk(flights)
dm_nycflights13() \%>\%
  dm_get_pk(planes)
\dontshow{\}) # examplesIf}
}
\seealso{
Other primary key functions: 
\code{\link{dm_add_pk}()},
\code{\link{dm_get_all_pks}()},
\code{\link{dm_has_pk}()},
\code{\link{enum_pk_candidates}()}
}
\concept{primary key functions}
