% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{dm_has_fk}
\alias{dm_has_fk}
\title{Does there exist a reference from one table of a \code{dm} to another?}
\usage{
dm_has_fk(dm, table, ref_table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}.}

\item{ref_table}{The table that \code{table} is potentially referencing.}
}
\value{
A boolean value: \code{TRUE} if a reference from \code{table} to \code{ref_table} exists, \code{FALSE} otherwise.
}
\description{
Does there exist a reference from one table of a \code{dm} to another?
}
\seealso{
Other foreign key functions: 
\code{\link{dm_add_fk}()},
\code{\link{dm_enum_fk_candidates}()},
\code{\link{dm_get_all_fks}()},
\code{\link{dm_get_fk}()},
\code{\link{dm_rm_fk}()}
}
\concept{foreign key functions}
