% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_nest_tbl.R
\name{dm_nest_tbl}
\alias{dm_nest_tbl}
\alias{dm_pack_tbl}
\title{Nest or pack a table inside its dm}
\usage{
dm_nest_tbl(dm, table, into = NULL, silent = FALSE)

dm_pack_tbl(dm, table, into = NULL, silent = FALSE)
}
\arguments{
\item{dm}{A dm}

\item{table}{A table}

\item{into}{The table to wrap \code{table} into, optional as it can be guessed
from the foreign keys unambiguously but useful to be explicit.}

\item{silent}{if not silent (the default), the code to unwrap will be printed
\itemize{
\item \code{dm_nest_tbl()} will nest a given \code{table} into its parent, \code{table} itself
should not have children tables (i.e. it needs to be a \emph{terminal child table})
\item \code{dm_pack_tbl()} will pack a given \code{table} into its child, \code{table} itself
should not have parent tables (i.e. it needs to be a \emph{terminal parent table})
}}
}
\description{
These functions nest or pack a table inside its parent or child.
}
\seealso{
\link{dm_wrap}, \link{dm_unwrap},
\link{dm_to_tibble}, \link{tibble_to_dm}
}
