% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_nest_tbl.R
\name{dm_nest_tbl}
\alias{dm_nest_tbl}
\alias{dm_pack_tbl}
\title{Nest or pack a table inside its dm}
\usage{
dm_nest_tbl(dm, table, into = NULL)

dm_pack_tbl(dm, table, into = NULL)
}
\arguments{
\item{dm}{A dm.}

\item{table}{A table.}

\item{into}{The table to wrap \code{table} into, optional as it can be guessed
from the foreign keys unambiguously but useful to be explicit.}
}
\description{
\code{dm_nest_tbl()} converts a child table to a nested column in its parent
table.
The child table should not have children itself (i.e. it needs to be a
\emph{terminal child table}).

\code{dm_pack_tbl()} converts a parent table to a packed column in its child
table.
The parent table should not have parent tables itself (i.e. it needs to be a
\emph{terminal parent table}).
}
\seealso{
\code{\link[=dm_wrap_tbl]{dm_wrap_tbl()}}, \code{\link[=dm_unwrap_tbl]{dm_unwrap_tbl()}}
}
