% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-helpers.R
\name{dm_nrow}
\alias{dm_nrow}
\title{Number of rows}
\usage{
dm_nrow(dm)
}
\arguments{
\item{dm}{A \code{\link{dm}} object.}
}
\value{
A named vector with the number of rows for each table.
}
\description{
Returns a named vector with the number of rows for each table.
}
\examples{
\dontshow{if (rlang::is_installed("nycflights13")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dm_nycflights13() \%>\%
  dm_filter(airports, faa \%in\% c("EWR", "LGA")) \%>\%
  dm_apply_filters() \%>\%
  dm_nrow()
\dontshow{\}) # examplesIf}
}
