% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptype.R
\name{dm_ptype}
\alias{dm_ptype}
\title{Prototype for a dm object}
\usage{
dm_ptype(dm)
}
\arguments{
\item{dm}{A \code{dm} object.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The prototype contains all tables, all primary and foreign keys,
but no data.
All tables are truncated and converted to zero-row tibbles.
Column names retain their type.
This is useful for performing creation and population of a database
in separate steps.
}
\examples{
\dontshow{if (dm:::dm_has_financial()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dm_financial() \%>\%
  dm_ptype()

dm_financial() \%>\%
  dm_ptype() \%>\%
  dm_nrow()
\dontshow{\}) # examplesIf}
}
