% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{dm_rename}
\alias{dm_rename}
\title{Rename one or more columns of a \code{\link{dm}} table}
\usage{
dm_rename(dm, table, ...)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}

\item{...}{One or more unquoted expressions separated by commas. You can treat
variable names like they are positions, so you can use expressions like x:y
to select ranges of variables.

Use named arguments, e.g. new_name = old_name, to rename selected variables.

The arguments in ... are automatically quoted and evaluated in a context where
column names represent column positions. They also support unquoting and splicing.
See \code{vignette("programming", package = "dplyr")} for an introduction to these concepts.

See select helpers for more details and examples about tidyselect helpers such as starts_with(), everything(), ...}
}
\description{
Rename columns of your \code{\link{dm}} with a similar syntax to \code{dplyr::rename()}.
}
\details{
If key columns are renamed the meta-information of the \code{dm} is updated accordingly
}
\examples{
dm_nycflights13() \%>\%
  dm_rename(airports, code = faa, altitude = alt)
}
