% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_constraints.R
\name{dm_repair_constraints}
\alias{dm_repair_constraints}
\title{Repair a data model's constraints}
\usage{
dm_repair_constraints(
  dm,
  fk_repair = c("add", "delete", "none"),
  progress = NA,
  in_place = NULL
)
}
\arguments{
\item{dm}{A dm object}

\item{fk_repair}{Either \code{"add"} or \code{"delete"}, \code{"add"} will update
the parent table by adding the missing primary keys while \code{"delete"}
will remove from the child table the rows featuring the orphan foreign keys.}

\item{progress}{Whether to display a progress bar, if \code{NA} (the default)
hide in non-interactive mode, show in interactive mode. Requires the
'progress' package.}

\item{in_place}{Should \code{x} be modified in place? This argument is only
relevant for mutable backends (e.g. databases, data.tables).

When \code{TRUE}, a modified version of \code{x} is returned invisibly;
when \code{FALSE}, a new object representing the resulting changes is returned.}
}
\value{
a dm
}
\description{
Repair a data model's constraints
}
\examples{
nycflights <- dm_nycflights13()

repaired_add <-
  nycflights \%>\%
  dm_repair_constraints("add")

repaired_delete <-
  nycflights \%>\%
  dm_repair_constraints("delete")

nrow(nycflights$planes)
nrow(nycflights$flights)

# we have more planes, and the same amount of flights
nrow(repaired_add$planes)
nrow(repaired_add$flights)


# we (should) have the same amount of planes, and fewer flights
nrow(repaired_delete$planes)
nrow(repaired_delete$flights) # this is not right


}
