% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign-keys.R
\name{dm_rm_fk}
\alias{dm_rm_fk}
\title{Remove reference(s) from one table of a \code{\link{dm}} to another}
\usage{
dm_rm_fk(dm, table, column, ref_table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{A table in the \code{dm}}

\item{column}{The column of \code{table} which should no longer be referencing the primary
key of \code{ref_table}.
If \code{NULL}, all columns will be considered.}

\item{ref_table}{The table which \code{table} was referencing.}
}
\description{
Can either remove one reference between the two tables or all at once if parameter \code{column = NULL}.
All parameters can be provided unquoted or quoted.
}
\seealso{
Other foreign key functions: \code{\link{dm_add_fk}},
  \code{\link{dm_enum_fk_candidates}},
  \code{\link{dm_get_all_fks}}, \code{\link{dm_get_fk}},
  \code{\link{dm_has_fk}}
}
\concept{foreign key functions}
