% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{dm_select}
\alias{dm_select}
\title{Select and/or rename one or more columns of a \code{\link{dm}} table}
\usage{
dm_select(dm, table, ...)
}
\value{
An updated \code{dm} with the columns of \code{table} potentially reduced and/or renamed
}
\description{
Select columns of your \code{\link{dm}} using syntax that is similar to \code{dplyr::select()}.
}
\details{
If key columns are renamed, then the meta-information of the \code{dm} is updated accordingly.
If key columns are removed, then all related relations are dropped as well.
}
\examples{
cdm_nycflights13() \%>\%
  cdm_select(airports, code = faa, altitude = alt)

}
