% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-dm.R
\name{dm_set_table_description}
\alias{dm_set_table_description}
\alias{dm_get_table_description}
\title{Add info about a dm's tables}
\usage{
dm_set_table_description(dm, ...)

dm_get_table_description(dm, table = NULL, ...)
}
\arguments{
\item{dm}{A \code{\link{dm}} object.}

\item{...}{Descriptions for tables to set in the form \code{description = table}.
These descriptions will be displayed when creating a diagram with \code{\link[=dm_draw]{dm_draw()}}.
Multi-line descriptions can be achieved using the newline symbol \verb{\\n}.
The currently set descriptions can be checked using \code{\link[=dm_get_table_description]{dm_get_table_description()}}.
Descriptions can be removed using \code{NULL = table}.

\code{tidyselect} is supported, see \code{\link[dplyr:select]{dplyr::select()}} for details on the semantics.}

\item{table}{A table in the \code{dm}.}
}
\value{
A \code{dm} object containing descriptions for specified tables.

A named vector of tables, with the descriptions in the names.
}
\description{
Add info about a dm's tables
}
