% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_to_tibble.R
\name{dm_unnest_tbl}
\alias{dm_unnest_tbl}
\alias{dm_unpack_tbl}
\title{Unnest or unpack columns from a wrapped table}
\usage{
dm_unnest_tbl(
  dm,
  table,
  col,
  parent_fk = NULL,
  child_pk_names = NULL,
  child_fk_names = NULL,
  prototype = NULL
)

dm_unpack_tbl(
  dm,
  table,
  col,
  child_fk = NULL,
  parent_pk_names = NULL,
  parent_fk_names = NULL,
  prototype = NULL
)
}
\arguments{
\item{dm}{A dm}

\item{table}{A table}

\item{col}{The column to unpack or unnest (unquoted)}

\item{parent_fk}{Columns in the table to unnest that the unnested child's foreign keys point to}

\item{child_pk_names}{Names of the unnested child's primary keys}

\item{child_fk_names}{Names of the unnested child's foreign keys}

\item{prototype}{A dm}

\item{child_fk}{Foreign key columns of the table to unpack}

\item{parent_pk_names}{Names of the unpacked parent's primary keys}

\item{parent_fk_names}{Names of the unpacked parent's foreign keys}
}
\value{
A dm
}
\description{
\code{dm_unpack_tbl()} and \code{dm_unnest_tbl()} target a specific column to unpack/unnest
from the given table in a given dm. A prototype or a set of keys should be given,
not both.
}
\examples{
airlines_wrapped <- dm_wrap(dm_nycflights13(), "airlines")

airlines_wrapped \%>\%
dm_unnest_tbl(airlines, flights, parent_fk = carrier, child_fk_names = "carrier") \%>\%
  dm_unpack_tbl(
    flights, weather, child_fk = c(origin, time_hour),
    parent_fk_names = c("origin", "time_hour"),
    parent_pk_names = c("origin", "time_hour")) \%>\%
  dm_unpack_tbl(
    flights, planes, child_fk = tailnum, parent_fk_names = "tailnum",
    parent_pk_names = "tailnum") \%>\%
  dm_unpack_tbl(
    flights, airports, child_fk = origin, parent_fk_names = "faa",
    parent_pk_names = "faa")

airlines_wrapped \%>\%
  dm_unnest_tbl(airlines, flights, prototype = dm_nycflights13()) \%>\%
  dm_unpack_tbl(flights, weather, prototype = dm_nycflights13()) \%>\%
  dm_unpack_tbl(flights, planes, prototype = dm_nycflights13()) \%>\%
  dm_unpack_tbl(flights, airports, prototype = dm_nycflights13())
}
\seealso{
\link{dm_wrap}, \link{dm_wrap}, \link{dm_unwrap},
\link{dm_to_tibble}, \link{tibble_to_dm}
}
