% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_wrap.R
\name{dm_unwrap}
\alias{dm_unwrap}
\title{Unwrap a single table dm}
\usage{
dm_unwrap(dm, ptype)
}
\arguments{
\item{dm}{A dm.}

\item{ptype}{A dm, might be an empty ptype.}
}
\value{
A dm.
}
\description{
\code{dm_unwrap()} unwraps all tables in a dm object so that the resulting dm
matches a given ptype dm.
It runs a sequence of \code{\link[=dm_unnest_tbl]{dm_unnest_tbl()}} and \code{\link[=dm_unpack_tbl]{dm_unpack_tbl()}} operations
on the dm.
}
\examples{

roundtrip <-
  dm_nycflights13() \%>\%
  dm_wrap(root = flights, silent = TRUE) \%>\%
  dm_unwrap(ptype = dm_nycflights13())
roundtrip

# The roundtrip has the same structure but fewer rows:
dm_nrow(dm_nycflights13())
dm_nrow(roundtrip)
}
