% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_to_tibble.R
\name{dm_unwrap}
\alias{dm_unwrap}
\alias{dm_unnest_tbl}
\alias{dm_unpack_tbl}
\title{Unwrap a table inside its dm}
\usage{
dm_unwrap(dm, table, keys)

dm_unnest_tbl(dm, table, col, keys)

dm_unpack_tbl(dm, table, col, keys)
}
\arguments{
\item{dm}{A dm}

\item{table}{A table}

\item{keys}{A dm, or a list of character vectors containing the necessary
items among \code{child_pk}, \code{child_fk}, \code{parent_pk}, \code{parent_fk}}

\item{col}{The column to unpack or unnest (unquoted)}
}
\value{
A dm
}
\description{
These functions unpack or unnest columns from a wrapped table. \code{dm_unpack_tbl()}
and \code{dm_unnest_tbl()} target a specific column to unpack/unnest, while \code{dm_unwrap()}
unpacks or unnests all relevant columns.
}
\examples{
airlines_wrapped <- dm_wrap_all(dm_nycflights13(), airlines, silent = TRUE)
dm_unwrap(airlines_wrapped, airlines, keys = dm_nycflights13())
airlines_wrapped \%>\%
  dm_unnest_tbl(airlines, flights, keys = list(
    child_fk = "carrier", parent_fk = "carrier")) \%>\%
  dm_unpack_tbl(flights, weather, keys = list(
    child_fk = c("origin", "time_hour"),
    parent_pk = c("origin",  "time_hour"),
    parent_fk = c("origin", "time_hour")
  ))

}
\seealso{
\link{dm_wrap}, \link{dm_wrap_all}, \link{dm_unwrap_all},
\link{dm_to_tibble}, \link{tibble_to_dm}
}
