% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{dm_validate}
\alias{dm_validate}
\title{Validator}
\usage{
dm_validate(x)
}
\arguments{
\item{x}{An object.}
}
\value{
For \code{dm_validate()}: Returns the \code{dm}, invisibly, after finishing all checks.
}
\description{
\code{dm_validate()} checks the internal consistency of a \code{dm} object.
}
\details{
In theory, all \code{dm} objects created or modified by functions in this package
should be valid, and this function should never be useful.
Please fila an issue if any dm operation creates an invalid object.
}
\examples{
dm_validate(dm())

bad_dm <- structure(list(bad = "dm"), class = "dm")
try(dm_validate(bad_dm))
}
