% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial-db-con.R
\name{financial_db_con}
\alias{financial_db_con}
\title{Connection to SQL Financial Database}
\usage{
financial_db_con()
}
\value{
A \code{MariaDBConnection} object
}
\description{
Connects to relational.fit.cvut.cz unless the service is
unavailable, in which case db-edu.pacha.dev is used as a fallback
}
