% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{glimpse.dm}
\alias{glimpse.dm}
\title{Get a glimpse of \code{dm} object}
\usage{
\method{glimpse}{dm}(x, width = NULL, ...)
}
\arguments{
\item{x}{A \code{dm} object.}

\item{width}{Controls the maximum number of columns on a line used in
printing. If \code{NULL}, \code{getOption("width")} will be consulted.}

\item{...}{Passed to \code{\link[pillar:glimpse]{pillar::glimpse()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{glimpse()} provides an overview (dimensions, column data types, primary
keys, etc.) of all tables included in the \code{dm} object. It will additionally
print details about outgoing foreign keys for the child table.

\code{glimpse()} is provided by the pillar package, and re-exported by {dm}.
See \code{\link[pillar:glimpse]{pillar::glimpse()}} for more details.
}
\examples{

dm_nycflights13() \%>\% glimpse()

}
