% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm.R
\name{pull_tbl}
\alias{pull_tbl}
\title{Retrieve a table from a \code{dm} or a \code{zoomed_dm}}
\usage{
pull_tbl(dm, table)
}
\arguments{
\item{dm}{A \code{dm} object.}

\item{table}{One unquoted table name for \code{pull_tbl.dm()}, ignored for \code{pull_tbl.zoomed_dm()}.}
}
\value{
The requested table
}
\description{
This function has methods for both \code{dm} classes:
\enumerate{
\item With \code{pull_tbl.dm()} you can chose which table of the \code{dm} you want to retrieve.
\item With \code{pull_tbl.zoomed_dm()} you will retrieve the zoomed table in the current state.
}
}
\examples{
# both examples lead to the same tibble:
dm_nycflights13() \%>\%
  pull_tbl(airlines)
dm_nycflights13() \%>\%
  dm_zoom_to_tbl(airlines) \%>\%
  pull_tbl()

}
