% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\alias{dcode}
\title{manipulate factor levels}
\usage{
recode(x, L, asFactor = TRUE, newFirst = TRUE)

dcode(...)
}
\arguments{
\item{x}{factor or character vector to be relabelled or reordered}

\item{L}{a list specifying the relabeling/reordering, list names specify a
new (or old) name and the list elements are vectors of (old) values that
are to be assigned that name. The order of L is the order of the levels
(with old values 'outside' the list either after of before, depending on
the \code{newFirst} value). List elements \code{=NULL} just means that
this name is to be unchanged but be placed in that place (order-wise).}

\item{asFactor}{if \code{TRUE} then a factor is returned (default \code{TRUE})}

\item{newFirst}{if \code{TRUE} then added levels will be placed first in the list of levels (default \code{TRUE})}

\item{...}{dcode arguments passed to recode}
}
\description{
This functions enables relabeling, reordering and merging of factor levels.
}
\section{Functions}{
\itemize{
\item \code{dcode}: 'dcode' is an alias for 'recode' since the latter is in
use in some proper R package
}}

\examples{
x <- LETTERS[2:1]
factor(x) ## levels in alphabetical order
recode(x, L = list('B' = NULL)) ## makes 'B' be the first level
x <- c('', LETTERS[1:5], NA)
factor(x)
L <- list('new_BC_or_NA' = c('B', 'C', NA),
          'E' = NULL,
          'new_D_(level_after_E)' = 'D',
          'blank' = '')
y <- recode(x, L, newFirst = FALSE) ## all unspecified values comes first
levels(y)
table(y, x, useNA = 'ifany')
}
\seealso{
\code{\link{factor}}, \code{\link{relevel}}
}
