% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{filter}
\alias{mutate}
\alias{transmute}
\alias{summarise}
\alias{summarize}
\alias{select}
\alias{rename}
\alias{group_by}
\alias{ungroup}
\alias{left_join}
\alias{inner_join}
\alias{full_join}
\alias{semi_join}
\alias{anti_join}
\alias{right_join}
\alias{collect}
\alias{compute}
\alias{copy_to}
\alias{src_tbls}
\alias{tbl}
\alias{arrange}
\alias{rows_insert}
\alias{rows_append}
\alias{rows_update}
\alias{rows_patch}
\alias{rows_upsert}
\alias{rows_delete}
\alias{glimpse}
\alias{tibble}
\alias{unite}
\alias{separate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyrdplyr:\%>\%anti_joinarrangecollectcomputecopy_tofilterfull_joingroup_byinner_joinleft_joinmutaterenameright_joinrows_appendrows_deleterows_insertrows_patchrows_updaterows_upsertselectsemi_joinsrc_tblssummarisesummarizetbltransmuteungroup]{\%>\%anti_joinarrangecollectcomputecopy_tofilterfull_joingroup_byinner_joinleft_joinmutaterenameright_joinrows_appendrows_deleterows_insertrows_patchrows_updaterows_upsertselectsemi_joinsrc_tblssummarisesummarizetbltransmuteungroup}()()()()()()()()()()()()()()()()()()()()()()()()()()()}}

  \item{tibble}{\code{\link[tibbletibble:glimpsetibble]{glimpsetibble}()()}}

  \item{tidyr}{\code{\link[tidyrtidyr:separateunite]{separateunite}()()}}
}}

