% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-surgery.R
\name{reunite_parent_child}
\alias{reunite_parent_child}
\alias{reunite_parent_child_from_list}
\title{Merge two tables linked by a foreign key relation}
\usage{
reunite_parent_child(child_table, parent_table, id_column)

reunite_parent_child_from_list(list_of_parent_child_tables, id_column)
}
\description{
Perform table fusion by combining two tables by a common (key) column and then removing this column.

\code{reunite_parent_child()}: After joining the two tables by the column \code{id_column}, this column is removed. The transformation is roughly the
inverse of what \code{decompose_table()} does.

\code{reunite_parent_child_from_list()}: After joining the two tables by the column \code{id_column}, this column is removed.
The function is almost exactly the inverse of \code{decompose_table()} (the order of the columns is not retained and original rownames are lost).
}
