% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows-db.R
\name{rows-db}
\alias{rows-db}
\alias{rows_insert.tbl_SQLiteConnection}
\alias{rows_update.tbl_SQLiteConnection}
\title{Persisting data for database tables}
\usage{
\method{rows_insert}{tbl_SQLiteConnection}(
  .data,
  .key = NULL,
  ...,
  .persist = NULL,
  .copy = NULL,
  .check = NULL
)

\method{rows_update}{tbl_SQLiteConnection}(
  .data,
  .key = NULL,
  ...,
  .persist = NULL,
  .copy = NULL,
  .check = NULL
)
}
\arguments{
\item{.data}{Target table object.}

\item{.key}{Key columns, unquoted.}

\item{...}{New data.
If unnamed, must be a single object of the same class as \code{.data}.
If named, will be passed on to \code{\link[=tibble]{tibble()}}.}

\item{.persist}{Set to \code{FALSE} for running the operation without persisting.
In this mode, a modified version of \code{.data} is returned.
This allows verifying the results of an operation before actually
applying it.
Set to \code{TRUE} to perform the update on the database table.
The default is \code{FALSE} with an informative message.}

\item{.copy}{Set to \code{TRUE} to automatically copy the new data to the database if
necessary.}

\item{.check}{Set to \code{TRUE} to always check keys, or \code{FALSE} to never check.
The default is to check only if \code{.persist} is \code{TRUE} or \code{NULL}.}
}
\value{
A tbl object of the same structure as \code{.data}.
If \code{.persist = TRUE}, \link{invisible} and identical to \code{.data}.
}
\description{
\lifecycle{experimental}

These methods provide a framework for manipulating individual rows
in existing tables.
All operations expect that both existing and new data are presented
in two compatible \link{tbl} objects on the same data source.

On mutable backends like databases, these operations manipulate the
underlying storage.
In contrast to all other operations,
these operations may lead to irreversible changes to the underlying database.
Therefore, persistence must be requested explicitly with \code{.persist = TRUE}.
By default, an informative message is given.
Unlike \code{\link[=compute]{compute()}} or \code{\link[=copy_to]{copy_to()}}, no new tables are created.
}
\examples{
data <- memdb_frame(a = 1:3, b = letters[c(1:2, NA)], c = 0.5 + 0:2)

try(rows_insert(data, a, a = 4, b = "z"))
rows_insert(data, a, a = 4, b = "z", .copy = TRUE)
rows_update(data, a, a = 2:3, b = "w", .copy = TRUE, .persist = FALSE)

rows_insert(data, a, memdb_frame(a = 4, b = "z"), .persist = TRUE)
data
rows_update(data, a, memdb_frame(a = 2:3, b = "w"), .persist = TRUE)
data
}
