% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows-truncate.R
\name{rows_truncate}
\alias{rows_truncate}
\alias{sql_rows_truncate}
\title{Truncate all rows}
\usage{
rows_truncate(x, ..., in_place = FALSE)

sql_rows_truncate(x, ...)
}
\arguments{
\item{x}{A data frame or data frame extension (e.g. a tibble).}

\item{...}{Other parameters passed onto methods.}

\item{in_place}{Should \code{x} be modified in place? This argument is only
relevant for mutable backends (e.g. databases, data.tables).

When \code{TRUE}, a modified version of \code{x} is returned invisibly;
when \code{FALSE}, a new object representing the resulting changes is returned.}
}
\description{
\code{rows_truncate()} removes all rows.
This operation corresponds to \code{TRUNCATE} in SQL.
\code{...} is ignored.
}
