% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncenterProdCap.R
\name{uncenterProdCap}
\alias{uncenterProdCap}
\title{Uncenter productivty and log capacity parameters}
\usage{
uncenterProdCap(input, dat, params)
}
\arguments{
\item{dat}{data from the A & P file}

\item{params}{a vector of parameters with names flowCoef, msCoef, prod, cap.  Assumed to be from the Bayesian analysis, so centered (or not) based on input$centerFlow and input$centerMS}
}
\value{
a vector of uncentered parameters with names flowCoef, msCoef, prod, cap.
}
\description{
Uncenter capacity and productivity if the covariates were centered in the Bayesian Analysis.  VRAP uses the uncentered parameters. The input$centerFlow and input$centerMS variables specify whether and which covariates were centered.
}
\seealso{
centerProdCap
}
